Build Ready. Removed duplicate files.
