--- docs/assets/index-BA9C0b2H.js
@@ -1,67 +1,0 @@
-function Hc(e,t){for(var n=0;n<t.length;n++){const r=t[n];if(typeof r!="string"&&!Array.isArray(r)){for(const l in r)if(l!=="default"&&!(l in e)){const i=Object.getOwnPropertyDescriptor(r,l);i&&Object.defineProperty(e,l,i.get?i:{enumerable:!0,get:()=>r[l]})}}}return Object.freeze(Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}))}(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const l of document.querySelectorAll('link[rel="modulepreload"]'))r(l);new MutationObserver(l=>{for(const i of l)if(i.type==="childList")for(const o of i.addedNodes)o.tagName==="LINK"&&o.rel==="modulepreload"&&r(o)}).observe(document,{childList:!0,subtree:!0});function n(l){const i={};return l.integrity&&(i.integrity=l.integrity),l.referrerPolicy&&(i.referrerPolicy=l.referrerPolicy),l.crossOrigin==="use-credentials"?i.credentials="include":l.crossOrigin==="anonymous"?i.credentials="omit":i.credentials="same-origin",i}function r(l){if(l.ep)return;l.ep=!0;const i=n(l);fetch(l.href,i)}})();function Qc(e){return e&&e.__esModule&&Object.prototype.hasOwnProperty.call(e,"default")?e.default:e}var ws={exports:{}},ml={},xs={exports:{}},T={};/**
- * @license React
- * react.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var ir=Symbol.for("react.element"),Kc=Symbol.for("react.portal"),Gc=Symbol.for("react.fragment"),Yc=Symbol.for("react.strict_mode"),Xc=Symbol.for("react.profiler"),Jc=Symbol.for("react.provider"),Zc=Symbol.for("react.context"),qc=Symbol.for("react.forward_ref"),bc=Symbol.for("react.suspense"),ef=Symbol.for("react.memo"),tf=Symbol.for("react.lazy"),tu=Symbol.iterator;function nf(e){return e===null||typeof e!="object"?null:(e=tu&&e[tu]||e["@@iterator"],typeof e=="function"?e:null)}var Ss={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},ks=Object.assign,Cs={};function fn(e,t,n){this.props=e,this.context=t,this.refs=Cs,this.updater=n||Ss}fn.prototype.isReactComponent={};fn.prototype.setState=function(e,t){if(typeof e!="object"&&typeof e!="function"&&e!=null)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,t,"setState")};fn.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")};function Es(){}Es.prototype=fn.prototype;function ro(e,t,n){this.props=e,this.context=t,this.refs=Cs,this.updater=n||Ss}var lo=ro.prototype=new Es;lo.constructor=ro;ks(lo,fn.prototype);lo.isPureReactComponent=!0;var nu=Array.isArray,js=Object.prototype.hasOwnProperty,io={current:null},_s={key:!0,ref:!0,__self:!0,__source:!0};function Ps(e,t,n){var r,l={},i=null,o=null;if(t!=null)for(r in t.ref!==void 0&&(o=t.ref),t.key!==void 0&&(i=""+t.key),t)js.call(t,r)&&!_s.hasOwnProperty(r)&&(l[r]=t[r]);var u=arguments.length-2;if(u===1)l.children=n;else if(1<u){for(var s=Array(u),a=0;a<u;a++)s[a]=arguments[a+2];l.children=s}if(e&&e.defaultProps)for(r in u=e.defaultProps,u)l[r]===void 0&&(l[r]=u[r]);return{$$typeof:ir,type:e,key:i,ref:o,props:l,_owner:io.current}}function rf(e,t){return{$$typeof:ir,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}function oo(e){return typeof e=="object"&&e!==null&&e.$$typeof===ir}function lf(e){var t={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,function(n){return t[n]})}var ru=/\/+/g;function Fl(e,t){return typeof e=="object"&&e!==null&&e.key!=null?lf(""+e.key):t.toString(36)}function zr(e,t,n,r,l){var i=typeof e;(i==="undefined"||i==="boolean")&&(e=null);var o=!1;if(e===null)o=!0;else switch(i){case"string":case"number":o=!0;break;case"object":switch(e.$$typeof){case ir:case Kc:o=!0}}if(o)return o=e,l=l(o),e=r===""?"."+Fl(o,0):r,nu(l)?(n="",e!=null&&(n=e.replace(ru,"$&/")+"/"),zr(l,t,n,"",function(a){return a})):l!=null&&(oo(l)&&(l=rf(l,n+(!l.key||o&&o.key===l.key?"":(""+l.key).replace(ru,"$&/")+"/")+e)),t.push(l)),1;if(o=0,r=r===""?".":r+":",nu(e))for(var u=0;u<e.length;u++){i=e[u];var s=r+Fl(i,u);o+=zr(i,t,n,s,l)}else if(s=nf(e),typeof s=="function")for(e=s.call(e),u=0;!(i=e.next()).done;)i=i.value,s=r+Fl(i,u++),o+=zr(i,t,n,s,l);else if(i==="object")throw t=String(e),Error("Objects are not valid as a React child (found: "+(t==="[object Object]"?"object with keys {"+Object.keys(e).join(", ")+"}":t)+"). If you meant to render a collection of children, use an array instead.");return o}function hr(e,t,n){if(e==null)return e;var r=[],l=0;return zr(e,r,"","",function(i){return t.call(n,i,l++)}),r}function of(e){if(e._status===-1){var t=e._result;t=t(),t.then(function(n){(e._status===0||e._status===-1)&&(e._status=1,e._result=n)},function(n){(e._status===0||e._status===-1)&&(e._status=2,e._result=n)}),e._status===-1&&(e._status=0,e._result=t)}if(e._status===1)return e._result.default;throw e._result}var ae={current:null},Tr={transition:null},uf={ReactCurrentDispatcher:ae,ReactCurrentBatchConfig:Tr,ReactCurrentOwner:io};function Ns(){throw Error("act(...) is not supported in production builds of React.")}T.Children={map:hr,forEach:function(e,t,n){hr(e,function(){t.apply(this,arguments)},n)},count:function(e){var t=0;return hr(e,function(){t++}),t},toArray:function(e){return hr(e,function(t){return t})||[]},only:function(e){if(!oo(e))throw Error("React.Children.only expected to receive a single React element child.");return e}};T.Component=fn;T.Fragment=Gc;T.Profiler=Xc;T.PureComponent=ro;T.StrictMode=Yc;T.Suspense=bc;T.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=uf;T.act=Ns;T.cloneElement=function(e,t,n){if(e==null)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var r=ks({},e.props),l=e.key,i=e.ref,o=e._owner;if(t!=null){if(t.ref!==void 0&&(i=t.ref,o=io.current),t.key!==void 0&&(l=""+t.key),e.type&&e.type.defaultProps)var u=e.type.defaultProps;for(s in t)js.call(t,s)&&!_s.hasOwnProperty(s)&&(r[s]=t[s]===void 0&&u!==void 0?u[s]:t[s])}var s=arguments.length-2;if(s===1)r.children=n;else if(1<s){u=Array(s);for(var a=0;a<s;a++)u[a]=arguments[a+2];r.children=u}return{$$typeof:ir,type:e.type,key:l,ref:i,props:r,_owner:o}};T.createContext=function(e){return e={$$typeof:Zc,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null},e.Provider={$$typeof:Jc,_context:e},e.Consumer=e};T.createElement=Ps;T.createFactory=function(e){var t=Ps.bind(null,e);return t.type=e,t};T.createRef=function(){return{current:null}};T.forwardRef=function(e){return{$$typeof:qc,render:e}};T.isValidElement=oo;T.lazy=function(e){return{$$typeof:tf,_payload:{_status:-1,_result:e},_init:of}};T.memo=function(e,t){return{$$typeof:ef,type:e,compare:t===void 0?null:t}};T.startTransition=function(e){var t=Tr.transition;Tr.transition={};try{e()}finally{Tr.transition=t}};T.unstable_act=Ns;T.useCallback=function(e,t){return ae.current.useCallback(e,t)};T.useContext=function(e){return ae.current.useContext(e)};T.useDebugValue=function(){};T.useDeferredValue=function(e){return ae.current.useDeferredValue(e)};T.useEffect=function(e,t){return ae.current.useEffect(e,t)};T.useId=function(){return ae.current.useId()};T.useImperativeHandle=function(e,t,n){return ae.current.useImperativeHandle(e,t,n)};T.useInsertionEffect=function(e,t){return ae.current.useInsertionEffect(e,t)};T.useLayoutEffect=function(e,t){return ae.current.useLayoutEffect(e,t)};T.useMemo=function(e,t){return ae.current.useMemo(e,t)};T.useReducer=function(e,t,n){return ae.current.useReducer(e,t,n)};T.useRef=function(e){return ae.current.useRef(e)};T.useState=function(e){return ae.current.useState(e)};T.useSyncExternalStore=function(e,t,n){return ae.current.useSyncExternalStore(e,t,n)};T.useTransition=function(){return ae.current.useTransition()};T.version="18.3.1";xs.exports=T;var E=xs.exports;const sf=Qc(E),af=Hc({__proto__:null,default:sf},[E]);/**
- * @license React
- * react-jsx-runtime.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var cf=E,ff=Symbol.for("react.element"),df=Symbol.for("react.fragment"),pf=Object.prototype.hasOwnProperty,hf=cf.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,mf={key:!0,ref:!0,__self:!0,__source:!0};function Ls(e,t,n){var r,l={},i=null,o=null;n!==void 0&&(i=""+n),t.key!==void 0&&(i=""+t.key),t.ref!==void 0&&(o=t.ref);for(r in t)pf.call(t,r)&&!mf.hasOwnProperty(r)&&(l[r]=t[r]);if(e&&e.defaultProps)for(r in t=e.defaultProps,t)l[r]===void 0&&(l[r]=t[r]);return{$$typeof:ff,type:e,key:i,ref:o,props:l,_owner:hf.current}}ml.Fragment=df;ml.jsx=Ls;ml.jsxs=Ls;ws.exports=ml;var v=ws.exports,Rs={exports:{}},xe={},zs={exports:{}},Ts={};/**
- * @license React
- * scheduler.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */(function(e){function t(_,R){var z=_.length;_.push(R);e:for(;0<z;){var Q=z-1>>>1,Z=_[Q];if(0<l(Z,R))_[Q]=R,_[z]=Z,z=Q;else break e}}function n(_){return _.length===0?null:_[0]}function r(_){if(_.length===0)return null;var R=_[0],z=_.pop();if(z!==R){_[0]=z;e:for(var Q=0,Z=_.length,dr=Z>>>1;Q<dr;){var wt=2*(Q+1)-1,Ol=_[wt],xt=wt+1,pr=_[xt];if(0>l(Ol,z))xt<Z&&0>l(pr,Ol)?(_[Q]=pr,_[xt]=z,Q=xt):(_[Q]=Ol,_[wt]=z,Q=wt);else if(xt<Z&&0>l(pr,z))_[Q]=pr,_[xt]=z,Q=xt;else break e}}return R}function l(_,R){var z=_.sortIndex-R.sortIndex;return z!==0?z:_.id-R.id}if(typeof performance=="object"&&typeof performance.now=="function"){var i=performance;e.unstable_now=function(){return i.now()}}else{var o=Date,u=o.now();e.unstable_now=function(){return o.now()-u}}var s=[],a=[],h=1,p=null,m=3,w=!1,x=!1,y=!1,C=typeof setTimeout=="function"?setTimeout:null,f=typeof clearTimeout=="function"?clearTimeout:null,c=typeof setImmediate<"u"?setImmediate:null;typeof navigator<"u"&&navigator.scheduling!==void 0&&navigator.scheduling.isInputPending!==void 0&&navigator.scheduling.isInputPending.bind(navigator.scheduling);function d(_){for(var R=n(a);R!==null;){if(R.callback===null)r(a);else if(R.startTime<=_)r(a),R.sortIndex=R.expirationTime,t(s,R);else break;R=n(a)}}function g(_){if(y=!1,d(_),!x)if(n(s)!==null)x=!0,Tl(k);else{var R=n(a);R!==null&&Ml(g,R.startTime-_)}}function k(_,R){x=!1,y&&(y=!1,f(L),L=-1),w=!0;var z=m;try{for(d(R),p=n(s);p!==null&&(!(p.expirationTime>R)||_&&!Ne());){var Q=p.callback;if(typeof Q=="function"){p.callback=null,m=p.priorityLevel;var Z=Q(p.expirationTime<=R);R=e.unstable_now(),typeof Z=="function"?p.callback=Z:p===n(s)&&r(s),d(R)}else r(s);p=n(s)}if(p!==null)var dr=!0;else{var wt=n(a);wt!==null&&Ml(g,wt.startTime-R),dr=!1}return dr}finally{p=null,m=z,w=!1}}var P=!1,N=null,L=-1,H=5,M=-1;function Ne(){return!(e.unstable_now()-M<H)}function vn(){if(N!==null){var _=e.unstable_now();M=_;var R=!0;try{R=N(!0,_)}finally{R?gn():(P=!1,N=null)}}else P=!1}var gn;if(typeof c=="function")gn=function(){c(vn)};else if(typeof MessageChannel<"u"){var eu=new MessageChannel,Wc=eu.port2;eu.port1.onmessage=vn,gn=function(){Wc.postMessage(null)}}else gn=function(){C(vn,0)};function Tl(_){N=_,P||(P=!0,gn())}function Ml(_,R){L=C(function(){_(e.unstable_now())},R)}e.unstable_IdlePriority=5,e.unstable_ImmediatePriority=1,e.unstable_LowPriority=4,e.unstable_NormalPriority=3,e.unstable_Profiling=null,e.unstable_UserBlockingPriority=2,e.unstable_cancelCallback=function(_){_.callback=null},e.unstable_continueExecution=function(){x||w||(x=!0,Tl(k))},e.unstable_forceFrameRate=function(_){0>_||125<_?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):H=0<_?Math.floor(1e3/_):5},e.unstable_getCurrentPriorityLevel=function(){return m},e.unstable_getFirstCallbackNode=function(){return n(s)},e.unstable_next=function(_){switch(m){case 1:case 2:case 3:var R=3;break;default:R=m}var z=m;m=R;try{return _()}finally{m=z}},e.unstable_pauseExecution=function(){},e.unstable_requestPaint=function(){},e.unstable_runWithPriority=function(_,R){switch(_){case 1:case 2:case 3:case 4:case 5:break;default:_=3}var z=m;m=_;try{return R()}finally{m=z}},e.unstable_scheduleCallback=function(_,R,z){var Q=e.unstable_now();switch(typeof z=="object"&&z!==null?(z=z.delay,z=typeof z=="number"&&0<z?Q+z:Q):z=Q,_){case 1:var Z=-1;break;case 2:Z=250;break;case 5:Z=1073741823;break;case 4:Z=1e4;break;default:Z=5e3}return Z=z+Z,_={id:h++,callback:R,priorityLevel:_,startTime:z,expirationTime:Z,sortIndex:-1},z>Q?(_.sortIndex=z,t(a,_),n(s)===null&&_===n(a)&&(y?(f(L),L=-1):y=!0,Ml(g,z-Q))):(_.sortIndex=Z,t(s,_),x||w||(x=!0,Tl(k))),_},e.unstable_shouldYield=Ne,e.unstable_wrapCallback=function(_){var R=m;return function(){var z=m;m=R;try{return _.apply(this,arguments)}finally{m=z}}}})(Ts);zs.exports=Ts;var vf=zs.exports;/**
- * @license React
- * react-dom.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var gf=E,we=vf;function S(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var Ms=new Set,Bn={};function Ot(e,t){rn(e,t),rn(e+"Capture",t)}function rn(e,t){for(Bn[e]=t,e=0;e<t.length;e++)Ms.add(t[e])}var Ke=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),si=Object.prototype.hasOwnProperty,yf=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,lu={},iu={};function wf(e){return si.call(iu,e)?!0:si.call(lu,e)?!1:yf.test(e)?iu[e]=!0:(lu[e]=!0,!1)}function xf(e,t,n,r){if(n!==null&&n.type===0)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return r?!1:n!==null?!n.acceptsBooleans:(e=e.toLowerCase().slice(0,5),e!=="data-"&&e!=="aria-");default:return!1}}function Sf(e,t,n,r){if(t===null||typeof t>"u"||xf(e,t,n,r))return!0;if(r)return!1;if(n!==null)switch(n.type){case 3:return!t;case 4:return t===!1;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}function ce(e,t,n,r,l,i,o){this.acceptsBooleans=t===2||t===3||t===4,this.attributeName=r,this.attributeNamespace=l,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=i,this.removeEmptyString=o}var ne={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach(function(e){ne[e]=new ce(e,0,!1,e,null,!1,!1)});[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach(function(e){var t=e[0];ne[t]=new ce(t,1,!1,e[1],null,!1,!1)});["contentEditable","draggable","spellCheck","value"].forEach(function(e){ne[e]=new ce(e,2,!1,e.toLowerCase(),null,!1,!1)});["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach(function(e){ne[e]=new ce(e,2,!1,e,null,!1,!1)});"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach(function(e){ne[e]=new ce(e,3,!1,e.toLowerCase(),null,!1,!1)});["checked","multiple","muted","selected"].forEach(function(e){ne[e]=new ce(e,3,!0,e,null,!1,!1)});["capture","download"].forEach(function(e){ne[e]=new ce(e,4,!1,e,null,!1,!1)});["cols","rows","size","span"].forEach(function(e){ne[e]=new ce(e,6,!1,e,null,!1,!1)});["rowSpan","start"].forEach(function(e){ne[e]=new ce(e,5,!1,e.toLowerCase(),null,!1,!1)});var uo=/[\-:]([a-z])/g;function so(e){return e[1].toUpperCase()}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach(function(e){var t=e.replace(uo,so);ne[t]=new ce(t,1,!1,e,null,!1,!1)});"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach(function(e){var t=e.replace(uo,so);ne[t]=new ce(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)});["xml:base","xml:lang","xml:space"].forEach(function(e){var t=e.replace(uo,so);ne[t]=new ce(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)});["tabIndex","crossOrigin"].forEach(function(e){ne[e]=new ce(e,1,!1,e.toLowerCase(),null,!1,!1)});ne.xlinkHref=new ce("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1);["src","href","action","formAction"].forEach(function(e){ne[e]=new ce(e,1,!1,e.toLowerCase(),null,!0,!0)});function ao(e,t,n,r){var l=ne.hasOwnProperty(t)?ne[t]:null;(l!==null?l.type!==0:r||!(2<t.length)||t[0]!=="o"&&t[0]!=="O"||t[1]!=="n"&&t[1]!=="N")&&(Sf(t,n,l,r)&&(n=null),r||l===null?wf(t)&&(n===null?e.removeAttribute(t):e.setAttribute(t,""+n)):l.mustUseProperty?e[l.propertyName]=n===null?l.type===3?!1:"":n:(t=l.attributeName,r=l.attributeNamespace,n===null?e.removeAttribute(t):(l=l.type,n=l===3||l===4&&n===!0?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}var Je=gf.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,mr=Symbol.for("react.element"),$t=Symbol.for("react.portal"),Bt=Symbol.for("react.fragment"),co=Symbol.for("react.strict_mode"),ai=Symbol.for("react.profiler"),Os=Symbol.for("react.provider"),Fs=Symbol.for("react.context"),fo=Symbol.for("react.forward_ref"),ci=Symbol.for("react.suspense"),fi=Symbol.for("react.suspense_list"),po=Symbol.for("react.memo"),qe=Symbol.for("react.lazy"),Ds=Symbol.for("react.offscreen"),ou=Symbol.iterator;function yn(e){return e===null||typeof e!="object"?null:(e=ou&&e[ou]||e["@@iterator"],typeof e=="function"?e:null)}var V=Object.assign,Dl;function _n(e){if(Dl===void 0)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);Dl=t&&t[1]||""}return`
-`+Dl+e}var Il=!1;function Ul(e,t){if(!e||Il)return"";Il=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),typeof Reflect=="object"&&Reflect.construct){try{Reflect.construct(t,[])}catch(a){var r=a}Reflect.construct(e,[],t)}else{try{t.call()}catch(a){r=a}e.call(t.prototype)}else{try{throw Error()}catch(a){r=a}e()}}catch(a){if(a&&r&&typeof a.stack=="string"){for(var l=a.stack.split(`
-`),i=r.stack.split(`
-`),o=l.length-1,u=i.length-1;1<=o&&0<=u&&l[o]!==i[u];)u--;for(;1<=o&&0<=u;o--,u--)if(l[o]!==i[u]){if(o!==1||u!==1)do if(o--,u--,0>u||l[o]!==i[u]){var s=`
-`+l[o].replace(" at new "," at ");return e.displayName&&s.includes("<anonymous>")&&(s=s.replace("<anonymous>",e.displayName)),s}while(1<=o&&0<=u);break}}}finally{Il=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?_n(e):""}function kf(e){switch(e.tag){case 5:return _n(e.type);case 16:return _n("Lazy");case 13:return _n("Suspense");case 19:return _n("SuspenseList");case 0:case 2:case 15:return e=Ul(e.type,!1),e;case 11:return e=Ul(e.type.render,!1),e;case 1:return e=Ul(e.type,!0),e;default:return""}}function di(e){if(e==null)return null;if(typeof e=="function")return e.displayName||e.name||null;if(typeof e=="string")return e;switch(e){case Bt:return"Fragment";case $t:return"Portal";case ai:return"Profiler";case co:return"StrictMode";case ci:return"Suspense";case fi:return"SuspenseList"}if(typeof e=="object")switch(e.$$typeof){case Fs:return(e.displayName||"Context")+".Consumer";case Os:return(e._context.displayName||"Context")+".Provider";case fo:var t=e.render;return e=e.displayName,e||(e=t.displayName||t.name||"",e=e!==""?"ForwardRef("+e+")":"ForwardRef"),e;case po:return t=e.displayName||null,t!==null?t:di(e.type)||"Memo";case qe:t=e._payload,e=e._init;try{return di(e(t))}catch{}}return null}function Cf(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=t.render,e=e.displayName||e.name||"",t.displayName||(e!==""?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return di(t);case 8:return t===co?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if(typeof t=="function")return t.displayName||t.name||null;if(typeof t=="string")return t}return null}function ht(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":return e;case"object":return e;default:return""}}function Is(e){var t=e.type;return(e=e.nodeName)&&e.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function Ef(e){var t=Is(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&typeof n<"u"&&typeof n.get=="function"&&typeof n.set=="function"){var l=n.get,i=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return l.call(this)},set:function(o){r=""+o,i.call(this,o)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(o){r=""+o},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}function vr(e){e._valueTracker||(e._valueTracker=Ef(e))}function Us(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=Is(e)?e.checked?"true":"false":e.value),e=r,e!==n?(t.setValue(e),!0):!1}function Wr(e){if(e=e||(typeof document<"u"?document:void 0),typeof e>"u")return null;try{return e.activeElement||e.body}catch{return e.body}}function pi(e,t){var n=t.checked;return V({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:n??e._wrapperState.initialChecked})}function uu(e,t){var n=t.defaultValue==null?"":t.defaultValue,r=t.checked!=null?t.checked:t.defaultChecked;n=ht(t.value!=null?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:t.type==="checkbox"||t.type==="radio"?t.checked!=null:t.value!=null}}function $s(e,t){t=t.checked,t!=null&&ao(e,"checked",t,!1)}function hi(e,t){$s(e,t);var n=ht(t.value),r=t.type;if(n!=null)r==="number"?(n===0&&e.value===""||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if(r==="submit"||r==="reset"){e.removeAttribute("value");return}t.hasOwnProperty("value")?mi(e,t.type,n):t.hasOwnProperty("defaultValue")&&mi(e,t.type,ht(t.defaultValue)),t.checked==null&&t.defaultChecked!=null&&(e.defaultChecked=!!t.defaultChecked)}function su(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!(r!=="submit"&&r!=="reset"||t.value!==void 0&&t.value!==null))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}n=e.name,n!==""&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,n!==""&&(e.name=n)}function mi(e,t,n){(t!=="number"||Wr(e.ownerDocument)!==e)&&(n==null?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var Pn=Array.isArray;function Zt(e,t,n,r){if(e=e.options,t){t={};for(var l=0;l<n.length;l++)t["$"+n[l]]=!0;for(n=0;n<e.length;n++)l=t.hasOwnProperty("$"+e[n].value),e[n].selected!==l&&(e[n].selected=l),l&&r&&(e[n].defaultSelected=!0)}else{for(n=""+ht(n),t=null,l=0;l<e.length;l++){if(e[l].value===n){e[l].selected=!0,r&&(e[l].defaultSelected=!0);return}t!==null||e[l].disabled||(t=e[l])}t!==null&&(t.selected=!0)}}function vi(e,t){if(t.dangerouslySetInnerHTML!=null)throw Error(S(91));return V({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function au(e,t){var n=t.value;if(n==null){if(n=t.children,t=t.defaultValue,n!=null){if(t!=null)throw Error(S(92));if(Pn(n)){if(1<n.length)throw Error(S(93));n=n[0]}t=n}t==null&&(t=""),n=t}e._wrapperState={initialValue:ht(n)}}function Bs(e,t){var n=ht(t.value),r=ht(t.defaultValue);n!=null&&(n=""+n,n!==e.value&&(e.value=n),t.defaultValue==null&&e.defaultValue!==n&&(e.defaultValue=n)),r!=null&&(e.defaultValue=""+r)}function cu(e){var t=e.textContent;t===e._wrapperState.initialValue&&t!==""&&t!==null&&(e.value=t)}function As(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function gi(e,t){return e==null||e==="http://www.w3.org/1999/xhtml"?As(t):e==="http://www.w3.org/2000/svg"&&t==="foreignObject"?"http://www.w3.org/1999/xhtml":e}var gr,Vs=function(e){return typeof MSApp<"u"&&MSApp.execUnsafeLocalFunction?function(t,n,r,l){MSApp.execUnsafeLocalFunction(function(){return e(t,n,r,l)})}:e}(function(e,t){if(e.namespaceURI!=="http://www.w3.org/2000/svg"||"innerHTML"in e)e.innerHTML=t;else{for(gr=gr||document.createElement("div"),gr.innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=gr.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}});function An(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&n.nodeType===3){n.nodeValue=t;return}}e.textContent=t}var zn={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},jf=["Webkit","ms","Moz","O"];Object.keys(zn).forEach(function(e){jf.forEach(function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),zn[t]=zn[e]})});function Ws(e,t,n){return t==null||typeof t=="boolean"||t===""?"":n||typeof t!="number"||t===0||zn.hasOwnProperty(e)&&zn[e]?(""+t).trim():t+"px"}function Hs(e,t){e=e.style;for(var n in t)if(t.hasOwnProperty(n)){var r=n.indexOf("--")===0,l=Ws(n,t[n],r);n==="float"&&(n="cssFloat"),r?e.setProperty(n,l):e[n]=l}}var _f=V({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function yi(e,t){if(t){if(_f[e]&&(t.children!=null||t.dangerouslySetInnerHTML!=null))throw Error(S(137,e));if(t.dangerouslySetInnerHTML!=null){if(t.children!=null)throw Error(S(60));if(typeof t.dangerouslySetInnerHTML!="object"||!("__html"in t.dangerouslySetInnerHTML))throw Error(S(61))}if(t.style!=null&&typeof t.style!="object")throw Error(S(62))}}function wi(e,t){if(e.indexOf("-")===-1)return typeof t.is=="string";switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var xi=null;function ho(e){return e=e.target||e.srcElement||window,e.correspondingUseElement&&(e=e.correspondingUseElement),e.nodeType===3?e.parentNode:e}var Si=null,qt=null,bt=null;function fu(e){if(e=sr(e)){if(typeof Si!="function")throw Error(S(280));var t=e.stateNode;t&&(t=xl(t),Si(e.stateNode,e.type,t))}}function Qs(e){qt?bt?bt.push(e):bt=[e]:qt=e}function Ks(){if(qt){var e=qt,t=bt;if(bt=qt=null,fu(e),t)for(e=0;e<t.length;e++)fu(t[e])}}function Gs(e,t){return e(t)}function Ys(){}var $l=!1;function Xs(e,t,n){if($l)return e(t,n);$l=!0;try{return Gs(e,t,n)}finally{$l=!1,(qt!==null||bt!==null)&&(Ys(),Ks())}}function Vn(e,t){var n=e.stateNode;if(n===null)return null;var r=xl(n);if(r===null)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(e=e.type,r=!(e==="button"||e==="input"||e==="select"||e==="textarea")),e=!r;break e;default:e=!1}if(e)return null;if(n&&typeof n!="function")throw Error(S(231,t,typeof n));return n}var ki=!1;if(Ke)try{var wn={};Object.defineProperty(wn,"passive",{get:function(){ki=!0}}),window.addEventListener("test",wn,wn),window.removeEventListener("test",wn,wn)}catch{ki=!1}function Pf(e,t,n,r,l,i,o,u,s){var a=Array.prototype.slice.call(arguments,3);try{t.apply(n,a)}catch(h){this.onError(h)}}var Tn=!1,Hr=null,Qr=!1,Ci=null,Nf={onError:function(e){Tn=!0,Hr=e}};function Lf(e,t,n,r,l,i,o,u,s){Tn=!1,Hr=null,Pf.apply(Nf,arguments)}function Rf(e,t,n,r,l,i,o,u,s){if(Lf.apply(this,arguments),Tn){if(Tn){var a=Hr;Tn=!1,Hr=null}else throw Error(S(198));Qr||(Qr=!0,Ci=a)}}function Ft(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do t=e,t.flags&4098&&(n=t.return),e=t.return;while(e)}return t.tag===3?n:null}function Js(e){if(e.tag===13){var t=e.memoizedState;if(t===null&&(e=e.alternate,e!==null&&(t=e.memoizedState)),t!==null)return t.dehydrated}return null}function du(e){if(Ft(e)!==e)throw Error(S(188))}function zf(e){var t=e.alternate;if(!t){if(t=Ft(e),t===null)throw Error(S(188));return t!==e?null:e}for(var n=e,r=t;;){var l=n.return;if(l===null)break;var i=l.alternate;if(i===null){if(r=l.return,r!==null){n=r;continue}break}if(l.child===i.child){for(i=l.child;i;){if(i===n)return du(l),e;if(i===r)return du(l),t;i=i.sibling}throw Error(S(188))}if(n.return!==r.return)n=l,r=i;else{for(var o=!1,u=l.child;u;){if(u===n){o=!0,n=l,r=i;break}if(u===r){o=!0,r=l,n=i;break}u=u.sibling}if(!o){for(u=i.child;u;){if(u===n){o=!0,n=i,r=l;break}if(u===r){o=!0,r=i,n=l;break}u=u.sibling}if(!o)throw Error(S(189))}}if(n.alternate!==r)throw Error(S(190))}if(n.tag!==3)throw Error(S(188));return n.stateNode.current===n?e:t}function Zs(e){return e=zf(e),e!==null?qs(e):null}function qs(e){if(e.tag===5||e.tag===6)return e;for(e=e.child;e!==null;){var t=qs(e);if(t!==null)return t;e=e.sibling}return null}var bs=we.unstable_scheduleCallback,pu=we.unstable_cancelCallback,Tf=we.unstable_shouldYield,Mf=we.unstable_requestPaint,K=we.unstable_now,Of=we.unstable_getCurrentPriorityLevel,mo=we.unstable_ImmediatePriority,ea=we.unstable_UserBlockingPriority,Kr=we.unstable_NormalPriority,Ff=we.unstable_LowPriority,ta=we.unstable_IdlePriority,vl=null,$e=null;function Df(e){if($e&&typeof $e.onCommitFiberRoot=="function")try{$e.onCommitFiberRoot(vl,e,void 0,(e.current.flags&128)===128)}catch{}}var Me=Math.clz32?Math.clz32:$f,If=Math.log,Uf=Math.LN2;function $f(e){return e>>>=0,e===0?32:31-(If(e)/Uf|0)|0}var yr=64,wr=4194304;function Nn(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return e&4194240;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return e&130023424;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function Gr(e,t){var n=e.pendingLanes;if(n===0)return 0;var r=0,l=e.suspendedLanes,i=e.pingedLanes,o=n&268435455;if(o!==0){var u=o&~l;u!==0?r=Nn(u):(i&=o,i!==0&&(r=Nn(i)))}else o=n&~l,o!==0?r=Nn(o):i!==0&&(r=Nn(i));if(r===0)return 0;if(t!==0&&t!==r&&!(t&l)&&(l=r&-r,i=t&-t,l>=i||l===16&&(i&4194240)!==0))return t;if(r&4&&(r|=n&16),t=e.entangledLanes,t!==0)for(e=e.entanglements,t&=r;0<t;)n=31-Me(t),l=1<<n,r|=e[n],t&=~l;return r}function Bf(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return-1;case 134217728:case 268435456:case 536870912:case 1073741824:return-1;default:return-1}}function Af(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,l=e.expirationTimes,i=e.pendingLanes;0<i;){var o=31-Me(i),u=1<<o,s=l[o];s===-1?(!(u&n)||u&r)&&(l[o]=Bf(u,t)):s<=t&&(e.expiredLanes|=u),i&=~u}}function Ei(e){return e=e.pendingLanes&-1073741825,e!==0?e:e&1073741824?1073741824:0}function na(){var e=yr;return yr<<=1,!(yr&4194240)&&(yr=64),e}function Bl(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function or(e,t,n){e.pendingLanes|=t,t!==536870912&&(e.suspendedLanes=0,e.pingedLanes=0),e=e.eventTimes,t=31-Me(t),e[t]=n}function Vf(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var l=31-Me(n),i=1<<l;t[l]=0,r[l]=-1,e[l]=-1,n&=~i}}function vo(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-Me(n),l=1<<r;l&t|e[r]&t&&(e[r]|=t),n&=~l}}var F=0;function ra(e){return e&=-e,1<e?4<e?e&268435455?16:536870912:4:1}var la,go,ia,oa,ua,ji=!1,xr=[],it=null,ot=null,ut=null,Wn=new Map,Hn=new Map,et=[],Wf="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function hu(e,t){switch(e){case"focusin":case"focusout":it=null;break;case"dragenter":case"dragleave":ot=null;break;case"mouseover":case"mouseout":ut=null;break;case"pointerover":case"pointerout":Wn.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":Hn.delete(t.pointerId)}}function xn(e,t,n,r,l,i){return e===null||e.nativeEvent!==i?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:i,targetContainers:[l]},t!==null&&(t=sr(t),t!==null&&go(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,l!==null&&t.indexOf(l)===-1&&t.push(l),e)}function Hf(e,t,n,r,l){switch(t){case"focusin":return it=xn(it,e,t,n,r,l),!0;case"dragenter":return ot=xn(ot,e,t,n,r,l),!0;case"mouseover":return ut=xn(ut,e,t,n,r,l),!0;case"pointerover":var i=l.pointerId;return Wn.set(i,xn(Wn.get(i)||null,e,t,n,r,l)),!0;case"gotpointercapture":return i=l.pointerId,Hn.set(i,xn(Hn.get(i)||null,e,t,n,r,l)),!0}return!1}function sa(e){var t=Et(e.target);if(t!==null){var n=Ft(t);if(n!==null){if(t=n.tag,t===13){if(t=Js(n),t!==null){e.blockedOn=t,ua(e.priority,function(){ia(n)});return}}else if(t===3&&n.stateNode.current.memoizedState.isDehydrated){e.blockedOn=n.tag===3?n.stateNode.containerInfo:null;return}}}e.blockedOn=null}function Mr(e){if(e.blockedOn!==null)return!1;for(var t=e.targetContainers;0<t.length;){var n=_i(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(n===null){n=e.nativeEvent;var r=new n.constructor(n.type,n);xi=r,n.target.dispatchEvent(r),xi=null}else return t=sr(n),t!==null&&go(t),e.blockedOn=n,!1;t.shift()}return!0}function mu(e,t,n){Mr(e)&&n.delete(t)}function Qf(){ji=!1,it!==null&&Mr(it)&&(it=null),ot!==null&&Mr(ot)&&(ot=null),ut!==null&&Mr(ut)&&(ut=null),Wn.forEach(mu),Hn.forEach(mu)}function Sn(e,t){e.blockedOn===t&&(e.blockedOn=null,ji||(ji=!0,we.unstable_scheduleCallback(we.unstable_NormalPriority,Qf)))}function Qn(e){function t(l){return Sn(l,e)}if(0<xr.length){Sn(xr[0],e);for(var n=1;n<xr.length;n++){var r=xr[n];r.blockedOn===e&&(r.blockedOn=null)}}for(it!==null&&Sn(it,e),ot!==null&&Sn(ot,e),ut!==null&&Sn(ut,e),Wn.forEach(t),Hn.forEach(t),n=0;n<et.length;n++)r=et[n],r.blockedOn===e&&(r.blockedOn=null);for(;0<et.length&&(n=et[0],n.blockedOn===null);)sa(n),n.blockedOn===null&&et.shift()}var en=Je.ReactCurrentBatchConfig,Yr=!0;function Kf(e,t,n,r){var l=F,i=en.transition;en.transition=null;try{F=1,yo(e,t,n,r)}finally{F=l,en.transition=i}}function Gf(e,t,n,r){var l=F,i=en.transition;en.transition=null;try{F=4,yo(e,t,n,r)}finally{F=l,en.transition=i}}function yo(e,t,n,r){if(Yr){var l=_i(e,t,n,r);if(l===null)Jl(e,t,r,Xr,n),hu(e,r);else if(Hf(l,e,t,n,r))r.stopPropagation();else if(hu(e,r),t&4&&-1<Wf.indexOf(e)){for(;l!==null;){var i=sr(l);if(i!==null&&la(i),i=_i(e,t,n,r),i===null&&Jl(e,t,r,Xr,n),i===l)break;l=i}l!==null&&r.stopPropagation()}else Jl(e,t,r,null,n)}}var Xr=null;function _i(e,t,n,r){if(Xr=null,e=ho(r),e=Et(e),e!==null)if(t=Ft(e),t===null)e=null;else if(n=t.tag,n===13){if(e=Js(t),e!==null)return e;e=null}else if(n===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return Xr=e,null}function aa(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(Of()){case mo:return 1;case ea:return 4;case Kr:case Ff:return 16;case ta:return 536870912;default:return 16}default:return 16}}var nt=null,wo=null,Or=null;function ca(){if(Or)return Or;var e,t=wo,n=t.length,r,l="value"in nt?nt.value:nt.textContent,i=l.length;for(e=0;e<n&&t[e]===l[e];e++);var o=n-e;for(r=1;r<=o&&t[n-r]===l[i-r];r++);return Or=l.slice(e,1<r?1-r:void 0)}function Fr(e){var t=e.keyCode;return"charCode"in e?(e=e.charCode,e===0&&t===13&&(e=13)):e=t,e===10&&(e=13),32<=e||e===13?e:0}function Sr(){return!0}function vu(){return!1}function Se(e){function t(n,r,l,i,o){this._reactName=n,this._targetInst=l,this.type=r,this.nativeEvent=i,this.target=o,this.currentTarget=null;for(var u in e)e.hasOwnProperty(u)&&(n=e[u],this[u]=n?n(i):i[u]);return this.isDefaultPrevented=(i.defaultPrevented!=null?i.defaultPrevented:i.returnValue===!1)?Sr:vu,this.isPropagationStopped=vu,this}return V(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var n=this.nativeEvent;n&&(n.preventDefault?n.preventDefault():typeof n.returnValue!="unknown"&&(n.returnValue=!1),this.isDefaultPrevented=Sr)},stopPropagation:function(){var n=this.nativeEvent;n&&(n.stopPropagation?n.stopPropagation():typeof n.cancelBubble!="unknown"&&(n.cancelBubble=!0),this.isPropagationStopped=Sr)},persist:function(){},isPersistent:Sr}),t}var dn={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},xo=Se(dn),ur=V({},dn,{view:0,detail:0}),Yf=Se(ur),Al,Vl,kn,gl=V({},ur,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:So,button:0,buttons:0,relatedTarget:function(e){return e.relatedTarget===void 0?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==kn&&(kn&&e.type==="mousemove"?(Al=e.screenX-kn.screenX,Vl=e.screenY-kn.screenY):Vl=Al=0,kn=e),Al)},movementY:function(e){return"movementY"in e?e.movementY:Vl}}),gu=Se(gl),Xf=V({},gl,{dataTransfer:0}),Jf=Se(Xf),Zf=V({},ur,{relatedTarget:0}),Wl=Se(Zf),qf=V({},dn,{animationName:0,elapsedTime:0,pseudoElement:0}),bf=Se(qf),ed=V({},dn,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),td=Se(ed),nd=V({},dn,{data:0}),yu=Se(nd),rd={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},ld={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},id={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function od(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):(e=id[e])?!!t[e]:!1}function So(){return od}var ud=V({},ur,{key:function(e){if(e.key){var t=rd[e.key]||e.key;if(t!=="Unidentified")return t}return e.type==="keypress"?(e=Fr(e),e===13?"Enter":String.fromCharCode(e)):e.type==="keydown"||e.type==="keyup"?ld[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:So,charCode:function(e){return e.type==="keypress"?Fr(e):0},keyCode:function(e){return e.type==="keydown"||e.type==="keyup"?e.keyCode:0},which:function(e){return e.type==="keypress"?Fr(e):e.type==="keydown"||e.type==="keyup"?e.keyCode:0}}),sd=Se(ud),ad=V({},gl,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),wu=Se(ad),cd=V({},ur,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:So}),fd=Se(cd),dd=V({},dn,{propertyName:0,elapsedTime:0,pseudoElement:0}),pd=Se(dd),hd=V({},gl,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),md=Se(hd),vd=[9,13,27,32],ko=Ke&&"CompositionEvent"in window,Mn=null;Ke&&"documentMode"in document&&(Mn=document.documentMode);var gd=Ke&&"TextEvent"in window&&!Mn,fa=Ke&&(!ko||Mn&&8<Mn&&11>=Mn),xu=" ",Su=!1;function da(e,t){switch(e){case"keyup":return vd.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function pa(e){return e=e.detail,typeof e=="object"&&"data"in e?e.data:null}var At=!1;function yd(e,t){switch(e){case"compositionend":return pa(t);case"keypress":return t.which!==32?null:(Su=!0,xu);case"textInput":return e=t.data,e===xu&&Su?null:e;default:return null}}function wd(e,t){if(At)return e==="compositionend"||!ko&&da(e,t)?(e=ca(),Or=wo=nt=null,At=!1,e):null;switch(e){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return fa&&t.locale!=="ko"?null:t.data;default:return null}}var xd={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function ku(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t==="input"?!!xd[e.type]:t==="textarea"}function ha(e,t,n,r){Qs(r),t=Jr(t,"onChange"),0<t.length&&(n=new xo("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var On=null,Kn=null;function Sd(e){ja(e,0)}function yl(e){var t=Ht(e);if(Us(t))return e}function kd(e,t){if(e==="change")return t}var ma=!1;if(Ke){var Hl;if(Ke){var Ql="oninput"in document;if(!Ql){var Cu=document.createElement("div");Cu.setAttribute("oninput","return;"),Ql=typeof Cu.oninput=="function"}Hl=Ql}else Hl=!1;ma=Hl&&(!document.documentMode||9<document.documentMode)}function Eu(){On&&(On.detachEvent("onpropertychange",va),Kn=On=null)}function va(e){if(e.propertyName==="value"&&yl(Kn)){var t=[];ha(t,Kn,e,ho(e)),Xs(Sd,t)}}function Cd(e,t,n){e==="focusin"?(Eu(),On=t,Kn=n,On.attachEvent("onpropertychange",va)):e==="focusout"&&Eu()}function Ed(e){if(e==="selectionchange"||e==="keyup"||e==="keydown")return yl(Kn)}function jd(e,t){if(e==="click")return yl(t)}function _d(e,t){if(e==="input"||e==="change")return yl(t)}function Pd(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}var Fe=typeof Object.is=="function"?Object.is:Pd;function Gn(e,t){if(Fe(e,t))return!0;if(typeof e!="object"||e===null||typeof t!="object"||t===null)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var l=n[r];if(!si.call(t,l)||!Fe(e[l],t[l]))return!1}return!0}function ju(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function _u(e,t){var n=ju(e);e=0;for(var r;n;){if(n.nodeType===3){if(r=e+n.textContent.length,e<=t&&r>=t)return{node:n,offset:t-e};e=r}e:{for(;n;){if(n.nextSibling){n=n.nextSibling;break e}n=n.parentNode}n=void 0}n=ju(n)}}function ga(e,t){return e&&t?e===t?!0:e&&e.nodeType===3?!1:t&&t.nodeType===3?ga(e,t.parentNode):"contains"in e?e.contains(t):e.compareDocumentPosition?!!(e.compareDocumentPosition(t)&16):!1:!1}function ya(){for(var e=window,t=Wr();t instanceof e.HTMLIFrameElement;){try{var n=typeof t.contentWindow.location.href=="string"}catch{n=!1}if(n)e=t.contentWindow;else break;t=Wr(e.document)}return t}function Co(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&(t==="input"&&(e.type==="text"||e.type==="search"||e.type==="tel"||e.type==="url"||e.type==="password")||t==="textarea"||e.contentEditable==="true")}function Nd(e){var t=ya(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&ga(n.ownerDocument.documentElement,n)){if(r!==null&&Co(n)){if(t=r.start,e=r.end,e===void 0&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if(e=(t=n.ownerDocument||document)&&t.defaultView||window,e.getSelection){e=e.getSelection();var l=n.textContent.length,i=Math.min(r.start,l);r=r.end===void 0?i:Math.min(r.end,l),!e.extend&&i>r&&(l=r,r=i,i=l),l=_u(n,i);var o=_u(n,r);l&&o&&(e.rangeCount!==1||e.anchorNode!==l.node||e.anchorOffset!==l.offset||e.focusNode!==o.node||e.focusOffset!==o.offset)&&(t=t.createRange(),t.setStart(l.node,l.offset),e.removeAllRanges(),i>r?(e.addRange(t),e.extend(o.node,o.offset)):(t.setEnd(o.node,o.offset),e.addRange(t)))}}for(t=[],e=n;e=e.parentNode;)e.nodeType===1&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for(typeof n.focus=="function"&&n.focus(),n=0;n<t.length;n++)e=t[n],e.element.scrollLeft=e.left,e.element.scrollTop=e.top}}var Ld=Ke&&"documentMode"in document&&11>=document.documentMode,Vt=null,Pi=null,Fn=null,Ni=!1;function Pu(e,t,n){var r=n.window===n?n.document:n.nodeType===9?n:n.ownerDocument;Ni||Vt==null||Vt!==Wr(r)||(r=Vt,"selectionStart"in r&&Co(r)?r={start:r.selectionStart,end:r.selectionEnd}:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection(),r={anchorNode:r.anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset}),Fn&&Gn(Fn,r)||(Fn=r,r=Jr(Pi,"onSelect"),0<r.length&&(t=new xo("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=Vt)))}function kr(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var Wt={animationend:kr("Animation","AnimationEnd"),animationiteration:kr("Animation","AnimationIteration"),animationstart:kr("Animation","AnimationStart"),transitionend:kr("Transition","TransitionEnd")},Kl={},wa={};Ke&&(wa=document.createElement("div").style,"AnimationEvent"in window||(delete Wt.animationend.animation,delete Wt.animationiteration.animation,delete Wt.animationstart.animation),"TransitionEvent"in window||delete Wt.transitionend.transition);function wl(e){if(Kl[e])return Kl[e];if(!Wt[e])return e;var t=Wt[e],n;for(n in t)if(t.hasOwnProperty(n)&&n in wa)return Kl[e]=t[n];return e}var xa=wl("animationend"),Sa=wl("animationiteration"),ka=wl("animationstart"),Ca=wl("transitionend"),Ea=new Map,Nu="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function vt(e,t){Ea.set(e,t),Ot(t,[e])}for(var Gl=0;Gl<Nu.length;Gl++){var Yl=Nu[Gl],Rd=Yl.toLowerCase(),zd=Yl[0].toUpperCase()+Yl.slice(1);vt(Rd,"on"+zd)}vt(xa,"onAnimationEnd");vt(Sa,"onAnimationIteration");vt(ka,"onAnimationStart");vt("dblclick","onDoubleClick");vt("focusin","onFocus");vt("focusout","onBlur");vt(Ca,"onTransitionEnd");rn("onMouseEnter",["mouseout","mouseover"]);rn("onMouseLeave",["mouseout","mouseover"]);rn("onPointerEnter",["pointerout","pointerover"]);rn("onPointerLeave",["pointerout","pointerover"]);Ot("onChange","change click focusin focusout input keydown keyup selectionchange".split(" "));Ot("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" "));Ot("onBeforeInput",["compositionend","keypress","textInput","paste"]);Ot("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" "));Ot("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" "));Ot("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var Ln="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),Td=new Set("cancel close invalid load scroll toggle".split(" ").concat(Ln));function Lu(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,Rf(r,t,void 0,e),e.currentTarget=null}function ja(e,t){t=(t&4)!==0;for(var n=0;n<e.length;n++){var r=e[n],l=r.event;r=r.listeners;e:{var i=void 0;if(t)for(var o=r.length-1;0<=o;o--){var u=r[o],s=u.instance,a=u.currentTarget;if(u=u.listener,s!==i&&l.isPropagationStopped())break e;Lu(l,u,a),i=s}else for(o=0;o<r.length;o++){if(u=r[o],s=u.instance,a=u.currentTarget,u=u.listener,s!==i&&l.isPropagationStopped())break e;Lu(l,u,a),i=s}}}if(Qr)throw e=Ci,Qr=!1,Ci=null,e}function I(e,t){var n=t[Mi];n===void 0&&(n=t[Mi]=new Set);var r=e+"__bubble";n.has(r)||(_a(t,e,2,!1),n.add(r))}function Xl(e,t,n){var r=0;t&&(r|=4),_a(n,e,r,t)}var Cr="_reactListening"+Math.random().toString(36).slice(2);function Yn(e){if(!e[Cr]){e[Cr]=!0,Ms.forEach(function(n){n!=="selectionchange"&&(Td.has(n)||Xl(n,!1,e),Xl(n,!0,e))});var t=e.nodeType===9?e:e.ownerDocument;t===null||t[Cr]||(t[Cr]=!0,Xl("selectionchange",!1,t))}}function _a(e,t,n,r){switch(aa(t)){case 1:var l=Kf;break;case 4:l=Gf;break;default:l=yo}n=l.bind(null,t,n,e),l=void 0,!ki||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(l=!0),r?l!==void 0?e.addEventListener(t,n,{capture:!0,passive:l}):e.addEventListener(t,n,!0):l!==void 0?e.addEventListener(t,n,{passive:l}):e.addEventListener(t,n,!1)}function Jl(e,t,n,r,l){var i=r;if(!(t&1)&&!(t&2)&&r!==null)e:for(;;){if(r===null)return;var o=r.tag;if(o===3||o===4){var u=r.stateNode.containerInfo;if(u===l||u.nodeType===8&&u.parentNode===l)break;if(o===4)for(o=r.return;o!==null;){var s=o.tag;if((s===3||s===4)&&(s=o.stateNode.containerInfo,s===l||s.nodeType===8&&s.parentNode===l))return;o=o.return}for(;u!==null;){if(o=Et(u),o===null)return;if(s=o.tag,s===5||s===6){r=i=o;continue e}u=u.parentNode}}r=r.return}Xs(function(){var a=i,h=ho(n),p=[];e:{var m=Ea.get(e);if(m!==void 0){var w=xo,x=e;switch(e){case"keypress":if(Fr(n)===0)break e;case"keydown":case"keyup":w=sd;break;case"focusin":x="focus",w=Wl;break;case"focusout":x="blur",w=Wl;break;case"beforeblur":case"afterblur":w=Wl;break;case"click":if(n.button===2)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":w=gu;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":w=Jf;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":w=fd;break;case xa:case Sa:case ka:w=bf;break;case Ca:w=pd;break;case"scroll":w=Yf;break;case"wheel":w=md;break;case"copy":case"cut":case"paste":w=td;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":w=wu}var y=(t&4)!==0,C=!y&&e==="scroll",f=y?m!==null?m+"Capture":null:m;y=[];for(var c=a,d;c!==null;){d=c;var g=d.stateNode;if(d.tag===5&&g!==null&&(d=g,f!==null&&(g=Vn(c,f),g!=null&&y.push(Xn(c,g,d)))),C)break;c=c.return}0<y.length&&(m=new w(m,x,null,n,h),p.push({event:m,listeners:y}))}}if(!(t&7)){e:{if(m=e==="mouseover"||e==="pointerover",w=e==="mouseout"||e==="pointerout",m&&n!==xi&&(x=n.relatedTarget||n.fromElement)&&(Et(x)||x[Ge]))break e;if((w||m)&&(m=h.window===h?h:(m=h.ownerDocument)?m.defaultView||m.parentWindow:window,w?(x=n.relatedTarget||n.toElement,w=a,x=x?Et(x):null,x!==null&&(C=Ft(x),x!==C||x.tag!==5&&x.tag!==6)&&(x=null)):(w=null,x=a),w!==x)){if(y=gu,g="onMouseLeave",f="onMouseEnter",c="mouse",(e==="pointerout"||e==="pointerover")&&(y=wu,g="onPointerLeave",f="onPointerEnter",c="pointer"),C=w==null?m:Ht(w),d=x==null?m:Ht(x),m=new y(g,c+"leave",w,n,h),m.target=C,m.relatedTarget=d,g=null,Et(h)===a&&(y=new y(f,c+"enter",x,n,h),y.target=d,y.relatedTarget=C,g=y),C=g,w&&x)t:{for(y=w,f=x,c=0,d=y;d;d=Ut(d))c++;for(d=0,g=f;g;g=Ut(g))d++;for(;0<c-d;)y=Ut(y),c--;for(;0<d-c;)f=Ut(f),d--;for(;c--;){if(y===f||f!==null&&y===f.alternate)break t;y=Ut(y),f=Ut(f)}y=null}else y=null;w!==null&&Ru(p,m,w,y,!1),x!==null&&C!==null&&Ru(p,C,x,y,!0)}}e:{if(m=a?Ht(a):window,w=m.nodeName&&m.nodeName.toLowerCase(),w==="select"||w==="input"&&m.type==="file")var k=kd;else if(ku(m))if(ma)k=_d;else{k=Ed;var P=Cd}else(w=m.nodeName)&&w.toLowerCase()==="input"&&(m.type==="checkbox"||m.type==="radio")&&(k=jd);if(k&&(k=k(e,a))){ha(p,k,n,h);break e}P&&P(e,m,a),e==="focusout"&&(P=m._wrapperState)&&P.controlled&&m.type==="number"&&mi(m,"number",m.value)}switch(P=a?Ht(a):window,e){case"focusin":(ku(P)||P.contentEditable==="true")&&(Vt=P,Pi=a,Fn=null);break;case"focusout":Fn=Pi=Vt=null;break;case"mousedown":Ni=!0;break;case"contextmenu":case"mouseup":case"dragend":Ni=!1,Pu(p,n,h);break;case"selectionchange":if(Ld)break;case"keydown":case"keyup":Pu(p,n,h)}var N;if(ko)e:{switch(e){case"compositionstart":var L="onCompositionStart";break e;case"compositionend":L="onCompositionEnd";break e;case"compositionupdate":L="onCompositionUpdate";break e}L=void 0}else At?da(e,n)&&(L="onCompositionEnd"):e==="keydown"&&n.keyCode===229&&(L="onCompositionStart");L&&(fa&&n.locale!=="ko"&&(At||L!=="onCompositionStart"?L==="onCompositionEnd"&&At&&(N=ca()):(nt=h,wo="value"in nt?nt.value:nt.textContent,At=!0)),P=Jr(a,L),0<P.length&&(L=new yu(L,e,null,n,h),p.push({event:L,listeners:P}),N?L.data=N:(N=pa(n),N!==null&&(L.data=N)))),(N=gd?yd(e,n):wd(e,n))&&(a=Jr(a,"onBeforeInput"),0<a.length&&(h=new yu("onBeforeInput","beforeinput",null,n,h),p.push({event:h,listeners:a}),h.data=N))}ja(p,t)})}function Xn(e,t,n){return{instance:e,listener:t,currentTarget:n}}function Jr(e,t){for(var n=t+"Capture",r=[];e!==null;){var l=e,i=l.stateNode;l.tag===5&&i!==null&&(l=i,i=Vn(e,n),i!=null&&r.unshift(Xn(e,i,l)),i=Vn(e,t),i!=null&&r.push(Xn(e,i,l))),e=e.return}return r}function Ut(e){if(e===null)return null;do e=e.return;while(e&&e.tag!==5);return e||null}function Ru(e,t,n,r,l){for(var i=t._reactName,o=[];n!==null&&n!==r;){var u=n,s=u.alternate,a=u.stateNode;if(s!==null&&s===r)break;u.tag===5&&a!==null&&(u=a,l?(s=Vn(n,i),s!=null&&o.unshift(Xn(n,s,u))):l||(s=Vn(n,i),s!=null&&o.push(Xn(n,s,u)))),n=n.return}o.length!==0&&e.push({event:t,listeners:o})}var Md=/\r\n?/g,Od=/\u0000|\uFFFD/g;function zu(e){return(typeof e=="string"?e:""+e).replace(Md,`
-`).replace(Od,"")}function Er(e,t,n){if(t=zu(t),zu(e)!==t&&n)throw Error(S(425))}function Zr(){}var Li=null,Ri=null;function zi(e,t){return e==="textarea"||e==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var Ti=typeof setTimeout=="function"?setTimeout:void 0,Fd=typeof clearTimeout=="function"?clearTimeout:void 0,Tu=typeof Promise=="function"?Promise:void 0,Dd=typeof queueMicrotask=="function"?queueMicrotask:typeof Tu<"u"?function(e){return Tu.resolve(null).then(e).catch(Id)}:Ti;function Id(e){setTimeout(function(){throw e})}function Zl(e,t){var n=t,r=0;do{var l=n.nextSibling;if(e.removeChild(n),l&&l.nodeType===8)if(n=l.data,n==="/$"){if(r===0){e.removeChild(l),Qn(t);return}r--}else n!=="$"&&n!=="$?"&&n!=="$!"||r++;n=l}while(n);Qn(t)}function st(e){for(;e!=null;e=e.nextSibling){var t=e.nodeType;if(t===1||t===3)break;if(t===8){if(t=e.data,t==="$"||t==="$!"||t==="$?")break;if(t==="/$")return null}}return e}function Mu(e){e=e.previousSibling;for(var t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="$"||n==="$!"||n==="$?"){if(t===0)return e;t--}else n==="/$"&&t++}e=e.previousSibling}return null}var pn=Math.random().toString(36).slice(2),Ue="__reactFiber$"+pn,Jn="__reactProps$"+pn,Ge="__reactContainer$"+pn,Mi="__reactEvents$"+pn,Ud="__reactListeners$"+pn,$d="__reactHandles$"+pn;function Et(e){var t=e[Ue];if(t)return t;for(var n=e.parentNode;n;){if(t=n[Ge]||n[Ue]){if(n=t.alternate,t.child!==null||n!==null&&n.child!==null)for(e=Mu(e);e!==null;){if(n=e[Ue])return n;e=Mu(e)}return t}e=n,n=e.parentNode}return null}function sr(e){return e=e[Ue]||e[Ge],!e||e.tag!==5&&e.tag!==6&&e.tag!==13&&e.tag!==3?null:e}function Ht(e){if(e.tag===5||e.tag===6)return e.stateNode;throw Error(S(33))}function xl(e){return e[Jn]||null}var Oi=[],Qt=-1;function gt(e){return{current:e}}function U(e){0>Qt||(e.current=Oi[Qt],Oi[Qt]=null,Qt--)}function D(e,t){Qt++,Oi[Qt]=e.current,e.current=t}var mt={},oe=gt(mt),pe=gt(!1),Lt=mt;function ln(e,t){var n=e.type.contextTypes;if(!n)return mt;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var l={},i;for(i in n)l[i]=t[i];return r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=l),l}function he(e){return e=e.childContextTypes,e!=null}function qr(){U(pe),U(oe)}function Ou(e,t,n){if(oe.current!==mt)throw Error(S(168));D(oe,t),D(pe,n)}function Pa(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,typeof r.getChildContext!="function")return n;r=r.getChildContext();for(var l in r)if(!(l in t))throw Error(S(108,Cf(e)||"Unknown",l));return V({},n,r)}function br(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||mt,Lt=oe.current,D(oe,e),D(pe,pe.current),!0}function Fu(e,t,n){var r=e.stateNode;if(!r)throw Error(S(169));n?(e=Pa(e,t,Lt),r.__reactInternalMemoizedMergedChildContext=e,U(pe),U(oe),D(oe,e)):U(pe),D(pe,n)}var Ve=null,Sl=!1,ql=!1;function Na(e){Ve===null?Ve=[e]:Ve.push(e)}function Bd(e){Sl=!0,Na(e)}function yt(){if(!ql&&Ve!==null){ql=!0;var e=0,t=F;try{var n=Ve;for(F=1;e<n.length;e++){var r=n[e];do r=r(!0);while(r!==null)}Ve=null,Sl=!1}catch(l){throw Ve!==null&&(Ve=Ve.slice(e+1)),bs(mo,yt),l}finally{F=t,ql=!1}}return null}var Kt=[],Gt=0,el=null,tl=0,ke=[],Ce=0,Rt=null,We=1,He="";function kt(e,t){Kt[Gt++]=tl,Kt[Gt++]=el,el=e,tl=t}function La(e,t,n){ke[Ce++]=We,ke[Ce++]=He,ke[Ce++]=Rt,Rt=e;var r=We;e=He;var l=32-Me(r)-1;r&=~(1<<l),n+=1;var i=32-Me(t)+l;if(30<i){var o=l-l%5;i=(r&(1<<o)-1).toString(32),r>>=o,l-=o,We=1<<32-Me(t)+l|n<<l|r,He=i+e}else We=1<<i|n<<l|r,He=e}function Eo(e){e.return!==null&&(kt(e,1),La(e,1,0))}function jo(e){for(;e===el;)el=Kt[--Gt],Kt[Gt]=null,tl=Kt[--Gt],Kt[Gt]=null;for(;e===Rt;)Rt=ke[--Ce],ke[Ce]=null,He=ke[--Ce],ke[Ce]=null,We=ke[--Ce],ke[Ce]=null}var ye=null,ge=null,$=!1,Te=null;function Ra(e,t){var n=Ee(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,t=e.deletions,t===null?(e.deletions=[n],e.flags|=16):t.push(n)}function Du(e,t){switch(e.tag){case 5:var n=e.type;return t=t.nodeType!==1||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t,t!==null?(e.stateNode=t,ye=e,ge=st(t.firstChild),!0):!1;case 6:return t=e.pendingProps===""||t.nodeType!==3?null:t,t!==null?(e.stateNode=t,ye=e,ge=null,!0):!1;case 13:return t=t.nodeType!==8?null:t,t!==null?(n=Rt!==null?{id:We,overflow:He}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},n=Ee(18,null,null,0),n.stateNode=t,n.return=e,e.child=n,ye=e,ge=null,!0):!1;default:return!1}}function Fi(e){return(e.mode&1)!==0&&(e.flags&128)===0}function Di(e){if($){var t=ge;if(t){var n=t;if(!Du(e,t)){if(Fi(e))throw Error(S(418));t=st(n.nextSibling);var r=ye;t&&Du(e,t)?Ra(r,n):(e.flags=e.flags&-4097|2,$=!1,ye=e)}}else{if(Fi(e))throw Error(S(418));e.flags=e.flags&-4097|2,$=!1,ye=e}}}function Iu(e){for(e=e.return;e!==null&&e.tag!==5&&e.tag!==3&&e.tag!==13;)e=e.return;ye=e}function jr(e){if(e!==ye)return!1;if(!$)return Iu(e),$=!0,!1;var t;if((t=e.tag!==3)&&!(t=e.tag!==5)&&(t=e.type,t=t!=="head"&&t!=="body"&&!zi(e.type,e.memoizedProps)),t&&(t=ge)){if(Fi(e))throw za(),Error(S(418));for(;t;)Ra(e,t),t=st(t.nextSibling)}if(Iu(e),e.tag===13){if(e=e.memoizedState,e=e!==null?e.dehydrated:null,!e)throw Error(S(317));e:{for(e=e.nextSibling,t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="/$"){if(t===0){ge=st(e.nextSibling);break e}t--}else n!=="$"&&n!=="$!"&&n!=="$?"||t++}e=e.nextSibling}ge=null}}else ge=ye?st(e.stateNode.nextSibling):null;return!0}function za(){for(var e=ge;e;)e=st(e.nextSibling)}function on(){ge=ye=null,$=!1}function _o(e){Te===null?Te=[e]:Te.push(e)}var Ad=Je.ReactCurrentBatchConfig;function Cn(e,t,n){if(e=n.ref,e!==null&&typeof e!="function"&&typeof e!="object"){if(n._owner){if(n=n._owner,n){if(n.tag!==1)throw Error(S(309));var r=n.stateNode}if(!r)throw Error(S(147,e));var l=r,i=""+e;return t!==null&&t.ref!==null&&typeof t.ref=="function"&&t.ref._stringRef===i?t.ref:(t=function(o){var u=l.refs;o===null?delete u[i]:u[i]=o},t._stringRef=i,t)}if(typeof e!="string")throw Error(S(284));if(!n._owner)throw Error(S(290,e))}return e}function _r(e,t){throw e=Object.prototype.toString.call(t),Error(S(31,e==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function Uu(e){var t=e._init;return t(e._payload)}function Ta(e){function t(f,c){if(e){var d=f.deletions;d===null?(f.deletions=[c],f.flags|=16):d.push(c)}}function n(f,c){if(!e)return null;for(;c!==null;)t(f,c),c=c.sibling;return null}function r(f,c){for(f=new Map;c!==null;)c.key!==null?f.set(c.key,c):f.set(c.index,c),c=c.sibling;return f}function l(f,c){return f=dt(f,c),f.index=0,f.sibling=null,f}function i(f,c,d){return f.index=d,e?(d=f.alternate,d!==null?(d=d.index,d<c?(f.flags|=2,c):d):(f.flags|=2,c)):(f.flags|=1048576,c)}function o(f){return e&&f.alternate===null&&(f.flags|=2),f}function u(f,c,d,g){return c===null||c.tag!==6?(c=ii(d,f.mode,g),c.return=f,c):(c=l(c,d),c.return=f,c)}function s(f,c,d,g){var k=d.type;return k===Bt?h(f,c,d.props.children,g,d.key):c!==null&&(c.elementType===k||typeof k=="object"&&k!==null&&k.$$typeof===qe&&Uu(k)===c.type)?(g=l(c,d.props),g.ref=Cn(f,c,d),g.return=f,g):(g=Vr(d.type,d.key,d.props,null,f.mode,g),g.ref=Cn(f,c,d),g.return=f,g)}function a(f,c,d,g){return c===null||c.tag!==4||c.stateNode.containerInfo!==d.containerInfo||c.stateNode.implementation!==d.implementation?(c=oi(d,f.mode,g),c.return=f,c):(c=l(c,d.children||[]),c.return=f,c)}function h(f,c,d,g,k){return c===null||c.tag!==7?(c=Nt(d,f.mode,g,k),c.return=f,c):(c=l(c,d),c.return=f,c)}function p(f,c,d){if(typeof c=="string"&&c!==""||typeof c=="number")return c=ii(""+c,f.mode,d),c.return=f,c;if(typeof c=="object"&&c!==null){switch(c.$$typeof){case mr:return d=Vr(c.type,c.key,c.props,null,f.mode,d),d.ref=Cn(f,null,c),d.return=f,d;case $t:return c=oi(c,f.mode,d),c.return=f,c;case qe:var g=c._init;return p(f,g(c._payload),d)}if(Pn(c)||yn(c))return c=Nt(c,f.mode,d,null),c.return=f,c;_r(f,c)}return null}function m(f,c,d,g){var k=c!==null?c.key:null;if(typeof d=="string"&&d!==""||typeof d=="number")return k!==null?null:u(f,c,""+d,g);if(typeof d=="object"&&d!==null){switch(d.$$typeof){case mr:return d.key===k?s(f,c,d,g):null;case $t:return d.key===k?a(f,c,d,g):null;case qe:return k=d._init,m(f,c,k(d._payload),g)}if(Pn(d)||yn(d))return k!==null?null:h(f,c,d,g,null);_r(f,d)}return null}function w(f,c,d,g,k){if(typeof g=="string"&&g!==""||typeof g=="number")return f=f.get(d)||null,u(c,f,""+g,k);if(typeof g=="object"&&g!==null){switch(g.$$typeof){case mr:return f=f.get(g.key===null?d:g.key)||null,s(c,f,g,k);case $t:return f=f.get(g.key===null?d:g.key)||null,a(c,f,g,k);case qe:var P=g._init;return w(f,c,d,P(g._payload),k)}if(Pn(g)||yn(g))return f=f.get(d)||null,h(c,f,g,k,null);_r(c,g)}return null}function x(f,c,d,g){for(var k=null,P=null,N=c,L=c=0,H=null;N!==null&&L<d.length;L++){N.index>L?(H=N,N=null):H=N.sibling;var M=m(f,N,d[L],g);if(M===null){N===null&&(N=H);break}e&&N&&M.alternate===null&&t(f,N),c=i(M,c,L),P===null?k=M:P.sibling=M,P=M,N=H}if(L===d.length)return n(f,N),$&&kt(f,L),k;if(N===null){for(;L<d.length;L++)N=p(f,d[L],g),N!==null&&(c=i(N,c,L),P===null?k=N:P.sibling=N,P=N);return $&&kt(f,L),k}for(N=r(f,N);L<d.length;L++)H=w(N,f,L,d[L],g),H!==null&&(e&&H.alternate!==null&&N.delete(H.key===null?L:H.key),c=i(H,c,L),P===null?k=H:P.sibling=H,P=H);return e&&N.forEach(function(Ne){return t(f,Ne)}),$&&kt(f,L),k}function y(f,c,d,g){var k=yn(d);if(typeof k!="function")throw Error(S(150));if(d=k.call(d),d==null)throw Error(S(151));for(var P=k=null,N=c,L=c=0,H=null,M=d.next();N!==null&&!M.done;L++,M=d.next()){N.index>L?(H=N,N=null):H=N.sibling;var Ne=m(f,N,M.value,g);if(Ne===null){N===null&&(N=H);break}e&&N&&Ne.alternate===null&&t(f,N),c=i(Ne,c,L),P===null?k=Ne:P.sibling=Ne,P=Ne,N=H}if(M.done)return n(f,N),$&&kt(f,L),k;if(N===null){for(;!M.done;L++,M=d.next())M=p(f,M.value,g),M!==null&&(c=i(M,c,L),P===null?k=M:P.sibling=M,P=M);return $&&kt(f,L),k}for(N=r(f,N);!M.done;L++,M=d.next())M=w(N,f,L,M.value,g),M!==null&&(e&&M.alternate!==null&&N.delete(M.key===null?L:M.key),c=i(M,c,L),P===null?k=M:P.sibling=M,P=M);return e&&N.forEach(function(vn){return t(f,vn)}),$&&kt(f,L),k}function C(f,c,d,g){if(typeof d=="object"&&d!==null&&d.type===Bt&&d.key===null&&(d=d.props.children),typeof d=="object"&&d!==null){switch(d.$$typeof){case mr:e:{for(var k=d.key,P=c;P!==null;){if(P.key===k){if(k=d.type,k===Bt){if(P.tag===7){n(f,P.sibling),c=l(P,d.props.children),c.return=f,f=c;break e}}else if(P.elementType===k||typeof k=="object"&&k!==null&&k.$$typeof===qe&&Uu(k)===P.type){n(f,P.sibling),c=l(P,d.props),c.ref=Cn(f,P,d),c.return=f,f=c;break e}n(f,P);break}else t(f,P);P=P.sibling}d.type===Bt?(c=Nt(d.props.children,f.mode,g,d.key),c.return=f,f=c):(g=Vr(d.type,d.key,d.props,null,f.mode,g),g.ref=Cn(f,c,d),g.return=f,f=g)}return o(f);case $t:e:{for(P=d.key;c!==null;){if(c.key===P)if(c.tag===4&&c.stateNode.containerInfo===d.containerInfo&&c.stateNode.implementation===d.implementation){n(f,c.sibling),c=l(c,d.children||[]),c.return=f,f=c;break e}else{n(f,c);break}else t(f,c);c=c.sibling}c=oi(d,f.mode,g),c.return=f,f=c}return o(f);case qe:return P=d._init,C(f,c,P(d._payload),g)}if(Pn(d))return x(f,c,d,g);if(yn(d))return y(f,c,d,g);_r(f,d)}return typeof d=="string"&&d!==""||typeof d=="number"?(d=""+d,c!==null&&c.tag===6?(n(f,c.sibling),c=l(c,d),c.return=f,f=c):(n(f,c),c=ii(d,f.mode,g),c.return=f,f=c),o(f)):n(f,c)}return C}var un=Ta(!0),Ma=Ta(!1),nl=gt(null),rl=null,Yt=null,Po=null;function No(){Po=Yt=rl=null}function Lo(e){var t=nl.current;U(nl),e._currentValue=t}function Ii(e,t,n){for(;e!==null;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,r!==null&&(r.childLanes|=t)):r!==null&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function tn(e,t){rl=e,Po=Yt=null,e=e.dependencies,e!==null&&e.firstContext!==null&&(e.lanes&t&&(de=!0),e.firstContext=null)}function _e(e){var t=e._currentValue;if(Po!==e)if(e={context:e,memoizedValue:t,next:null},Yt===null){if(rl===null)throw Error(S(308));Yt=e,rl.dependencies={lanes:0,firstContext:e}}else Yt=Yt.next=e;return t}var jt=null;function Ro(e){jt===null?jt=[e]:jt.push(e)}function Oa(e,t,n,r){var l=t.interleaved;return l===null?(n.next=n,Ro(t)):(n.next=l.next,l.next=n),t.interleaved=n,Ye(e,r)}function Ye(e,t){e.lanes|=t;var n=e.alternate;for(n!==null&&(n.lanes|=t),n=e,e=e.return;e!==null;)e.childLanes|=t,n=e.alternate,n!==null&&(n.childLanes|=t),n=e,e=e.return;return n.tag===3?n.stateNode:null}var be=!1;function zo(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function Fa(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function Qe(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function at(e,t,n){var r=e.updateQueue;if(r===null)return null;if(r=r.shared,O&2){var l=r.pending;return l===null?t.next=t:(t.next=l.next,l.next=t),r.pending=t,Ye(e,n)}return l=r.interleaved,l===null?(t.next=t,Ro(r)):(t.next=l.next,l.next=t),r.interleaved=t,Ye(e,n)}function Dr(e,t,n){if(t=t.updateQueue,t!==null&&(t=t.shared,(n&4194240)!==0)){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,vo(e,n)}}function $u(e,t){var n=e.updateQueue,r=e.alternate;if(r!==null&&(r=r.updateQueue,n===r)){var l=null,i=null;if(n=n.firstBaseUpdate,n!==null){do{var o={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};i===null?l=i=o:i=i.next=o,n=n.next}while(n!==null);i===null?l=i=t:i=i.next=t}else l=i=t;n={baseState:r.baseState,firstBaseUpdate:l,lastBaseUpdate:i,shared:r.shared,effects:r.effects},e.updateQueue=n;return}e=n.lastBaseUpdate,e===null?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function ll(e,t,n,r){var l=e.updateQueue;be=!1;var i=l.firstBaseUpdate,o=l.lastBaseUpdate,u=l.shared.pending;if(u!==null){l.shared.pending=null;var s=u,a=s.next;s.next=null,o===null?i=a:o.next=a,o=s;var h=e.alternate;h!==null&&(h=h.updateQueue,u=h.lastBaseUpdate,u!==o&&(u===null?h.firstBaseUpdate=a:u.next=a,h.lastBaseUpdate=s))}if(i!==null){var p=l.baseState;o=0,h=a=s=null,u=i;do{var m=u.lane,w=u.eventTime;if((r&m)===m){h!==null&&(h=h.next={eventTime:w,lane:0,tag:u.tag,payload:u.payload,callback:u.callback,next:null});e:{var x=e,y=u;switch(m=t,w=n,y.tag){case 1:if(x=y.payload,typeof x=="function"){p=x.call(w,p,m);break e}p=x;break e;case 3:x.flags=x.flags&-65537|128;case 0:if(x=y.payload,m=typeof x=="function"?x.call(w,p,m):x,m==null)break e;p=V({},p,m);break e;case 2:be=!0}}u.callback!==null&&u.lane!==0&&(e.flags|=64,m=l.effects,m===null?l.effects=[u]:m.push(u))}else w={eventTime:w,lane:m,tag:u.tag,payload:u.payload,callback:u.callback,next:null},h===null?(a=h=w,s=p):h=h.next=w,o|=m;if(u=u.next,u===null){if(u=l.shared.pending,u===null)break;m=u,u=m.next,m.next=null,l.lastBaseUpdate=m,l.shared.pending=null}}while(!0);if(h===null&&(s=p),l.baseState=s,l.firstBaseUpdate=a,l.lastBaseUpdate=h,t=l.shared.interleaved,t!==null){l=t;do o|=l.lane,l=l.next;while(l!==t)}else i===null&&(l.shared.lanes=0);Tt|=o,e.lanes=o,e.memoizedState=p}}function Bu(e,t,n){if(e=t.effects,t.effects=null,e!==null)for(t=0;t<e.length;t++){var r=e[t],l=r.callback;if(l!==null){if(r.callback=null,r=n,typeof l!="function")throw Error(S(191,l));l.call(r)}}}var ar={},Be=gt(ar),Zn=gt(ar),qn=gt(ar);function _t(e){if(e===ar)throw Error(S(174));return e}function To(e,t){switch(D(qn,t),D(Zn,e),D(Be,ar),e=t.nodeType,e){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:gi(null,"");break;default:e=e===8?t.parentNode:t,t=e.namespaceURI||null,e=e.tagName,t=gi(t,e)}U(Be),D(Be,t)}function sn(){U(Be),U(Zn),U(qn)}function Da(e){_t(qn.current);var t=_t(Be.current),n=gi(t,e.type);t!==n&&(D(Zn,e),D(Be,n))}function Mo(e){Zn.current===e&&(U(Be),U(Zn))}var B=gt(0);function il(e){for(var t=e;t!==null;){if(t.tag===13){var n=t.memoizedState;if(n!==null&&(n=n.dehydrated,n===null||n.data==="$?"||n.data==="$!"))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if(t.flags&128)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var bl=[];function Oo(){for(var e=0;e<bl.length;e++)bl[e]._workInProgressVersionPrimary=null;bl.length=0}var Ir=Je.ReactCurrentDispatcher,ei=Je.ReactCurrentBatchConfig,zt=0,A=null,X=null,q=null,ol=!1,Dn=!1,bn=0,Vd=0;function re(){throw Error(S(321))}function Fo(e,t){if(t===null)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!Fe(e[n],t[n]))return!1;return!0}function Do(e,t,n,r,l,i){if(zt=i,A=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,Ir.current=e===null||e.memoizedState===null?Kd:Gd,e=n(r,l),Dn){i=0;do{if(Dn=!1,bn=0,25<=i)throw Error(S(301));i+=1,q=X=null,t.updateQueue=null,Ir.current=Yd,e=n(r,l)}while(Dn)}if(Ir.current=ul,t=X!==null&&X.next!==null,zt=0,q=X=A=null,ol=!1,t)throw Error(S(300));return e}function Io(){var e=bn!==0;return bn=0,e}function Ie(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return q===null?A.memoizedState=q=e:q=q.next=e,q}function Pe(){if(X===null){var e=A.alternate;e=e!==null?e.memoizedState:null}else e=X.next;var t=q===null?A.memoizedState:q.next;if(t!==null)q=t,X=e;else{if(e===null)throw Error(S(310));X=e,e={memoizedState:X.memoizedState,baseState:X.baseState,baseQueue:X.baseQueue,queue:X.queue,next:null},q===null?A.memoizedState=q=e:q=q.next=e}return q}function er(e,t){return typeof t=="function"?t(e):t}function ti(e){var t=Pe(),n=t.queue;if(n===null)throw Error(S(311));n.lastRenderedReducer=e;var r=X,l=r.baseQueue,i=n.pending;if(i!==null){if(l!==null){var o=l.next;l.next=i.next,i.next=o}r.baseQueue=l=i,n.pending=null}if(l!==null){i=l.next,r=r.baseState;var u=o=null,s=null,a=i;do{var h=a.lane;if((zt&h)===h)s!==null&&(s=s.next={lane:0,action:a.action,hasEagerState:a.hasEagerState,eagerState:a.eagerState,next:null}),r=a.hasEagerState?a.eagerState:e(r,a.action);else{var p={lane:h,action:a.action,hasEagerState:a.hasEagerState,eagerState:a.eagerState,next:null};s===null?(u=s=p,o=r):s=s.next=p,A.lanes|=h,Tt|=h}a=a.next}while(a!==null&&a!==i);s===null?o=r:s.next=u,Fe(r,t.memoizedState)||(de=!0),t.memoizedState=r,t.baseState=o,t.baseQueue=s,n.lastRenderedState=r}if(e=n.interleaved,e!==null){l=e;do i=l.lane,A.lanes|=i,Tt|=i,l=l.next;while(l!==e)}else l===null&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function ni(e){var t=Pe(),n=t.queue;if(n===null)throw Error(S(311));n.lastRenderedReducer=e;var r=n.dispatch,l=n.pending,i=t.memoizedState;if(l!==null){n.pending=null;var o=l=l.next;do i=e(i,o.action),o=o.next;while(o!==l);Fe(i,t.memoizedState)||(de=!0),t.memoizedState=i,t.baseQueue===null&&(t.baseState=i),n.lastRenderedState=i}return[i,r]}function Ia(){}function Ua(e,t){var n=A,r=Pe(),l=t(),i=!Fe(r.memoizedState,l);if(i&&(r.memoizedState=l,de=!0),r=r.queue,Uo(Aa.bind(null,n,r,e),[e]),r.getSnapshot!==t||i||q!==null&&q.memoizedState.tag&1){if(n.flags|=2048,tr(9,Ba.bind(null,n,r,l,t),void 0,null),b===null)throw Error(S(349));zt&30||$a(n,t,l)}return l}function $a(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},t=A.updateQueue,t===null?(t={lastEffect:null,stores:null},A.updateQueue=t,t.stores=[e]):(n=t.stores,n===null?t.stores=[e]:n.push(e))}function Ba(e,t,n,r){t.value=n,t.getSnapshot=r,Va(t)&&Wa(e)}function Aa(e,t,n){return n(function(){Va(t)&&Wa(e)})}function Va(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!Fe(e,n)}catch{return!0}}function Wa(e){var t=Ye(e,1);t!==null&&Oe(t,e,1,-1)}function Au(e){var t=Ie();return typeof e=="function"&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:er,lastRenderedState:e},t.queue=e,e=e.dispatch=Qd.bind(null,A,e),[t.memoizedState,e]}function tr(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},t=A.updateQueue,t===null?(t={lastEffect:null,stores:null},A.updateQueue=t,t.lastEffect=e.next=e):(n=t.lastEffect,n===null?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e)),e}function Ha(){return Pe().memoizedState}function Ur(e,t,n,r){var l=Ie();A.flags|=e,l.memoizedState=tr(1|t,n,void 0,r===void 0?null:r)}function kl(e,t,n,r){var l=Pe();r=r===void 0?null:r;var i=void 0;if(X!==null){var o=X.memoizedState;if(i=o.destroy,r!==null&&Fo(r,o.deps)){l.memoizedState=tr(t,n,i,r);return}}A.flags|=e,l.memoizedState=tr(1|t,n,i,r)}function Vu(e,t){return Ur(8390656,8,e,t)}function Uo(e,t){return kl(2048,8,e,t)}function Qa(e,t){return kl(4,2,e,t)}function Ka(e,t){return kl(4,4,e,t)}function Ga(e,t){if(typeof t=="function")return e=e(),t(e),function(){t(null)};if(t!=null)return e=e(),t.current=e,function(){t.current=null}}function Ya(e,t,n){return n=n!=null?n.concat([e]):null,kl(4,4,Ga.bind(null,t,e),n)}function $o(){}function Xa(e,t){var n=Pe();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&Fo(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function Ja(e,t){var n=Pe();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&Fo(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function Za(e,t,n){return zt&21?(Fe(n,t)||(n=na(),A.lanes|=n,Tt|=n,e.baseState=!0),t):(e.baseState&&(e.baseState=!1,de=!0),e.memoizedState=n)}function Wd(e,t){var n=F;F=n!==0&&4>n?n:4,e(!0);var r=ei.transition;ei.transition={};try{e(!1),t()}finally{F=n,ei.transition=r}}function qa(){return Pe().memoizedState}function Hd(e,t,n){var r=ft(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},ba(e))ec(t,n);else if(n=Oa(e,t,n,r),n!==null){var l=se();Oe(n,e,r,l),tc(n,t,r)}}function Qd(e,t,n){var r=ft(e),l={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(ba(e))ec(t,l);else{var i=e.alternate;if(e.lanes===0&&(i===null||i.lanes===0)&&(i=t.lastRenderedReducer,i!==null))try{var o=t.lastRenderedState,u=i(o,n);if(l.hasEagerState=!0,l.eagerState=u,Fe(u,o)){var s=t.interleaved;s===null?(l.next=l,Ro(t)):(l.next=s.next,s.next=l),t.interleaved=l;return}}catch{}finally{}n=Oa(e,t,l,r),n!==null&&(l=se(),Oe(n,e,r,l),tc(n,t,r))}}function ba(e){var t=e.alternate;return e===A||t!==null&&t===A}function ec(e,t){Dn=ol=!0;var n=e.pending;n===null?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function tc(e,t,n){if(n&4194240){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,vo(e,n)}}var ul={readContext:_e,useCallback:re,useContext:re,useEffect:re,useImperativeHandle:re,useInsertionEffect:re,useLayoutEffect:re,useMemo:re,useReducer:re,useRef:re,useState:re,useDebugValue:re,useDeferredValue:re,useTransition:re,useMutableSource:re,useSyncExternalStore:re,useId:re,unstable_isNewReconciler:!1},Kd={readContext:_e,useCallback:function(e,t){return Ie().memoizedState=[e,t===void 0?null:t],e},useContext:_e,useEffect:Vu,useImperativeHandle:function(e,t,n){return n=n!=null?n.concat([e]):null,Ur(4194308,4,Ga.bind(null,t,e),n)},useLayoutEffect:function(e,t){return Ur(4194308,4,e,t)},useInsertionEffect:function(e,t){return Ur(4,2,e,t)},useMemo:function(e,t){var n=Ie();return t=t===void 0?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=Ie();return t=n!==void 0?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=Hd.bind(null,A,e),[r.memoizedState,e]},useRef:function(e){var t=Ie();return e={current:e},t.memoizedState=e},useState:Au,useDebugValue:$o,useDeferredValue:function(e){return Ie().memoizedState=e},useTransition:function(){var e=Au(!1),t=e[0];return e=Wd.bind(null,e[1]),Ie().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=A,l=Ie();if($){if(n===void 0)throw Error(S(407));n=n()}else{if(n=t(),b===null)throw Error(S(349));zt&30||$a(r,t,n)}l.memoizedState=n;var i={value:n,getSnapshot:t};return l.queue=i,Vu(Aa.bind(null,r,i,e),[e]),r.flags|=2048,tr(9,Ba.bind(null,r,i,n,t),void 0,null),n},useId:function(){var e=Ie(),t=b.identifierPrefix;if($){var n=He,r=We;n=(r&~(1<<32-Me(r)-1)).toString(32)+n,t=":"+t+"R"+n,n=bn++,0<n&&(t+="H"+n.toString(32)),t+=":"}else n=Vd++,t=":"+t+"r"+n.toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},Gd={readContext:_e,useCallback:Xa,useContext:_e,useEffect:Uo,useImperativeHandle:Ya,useInsertionEffect:Qa,useLayoutEffect:Ka,useMemo:Ja,useReducer:ti,useRef:Ha,useState:function(){return ti(er)},useDebugValue:$o,useDeferredValue:function(e){var t=Pe();return Za(t,X.memoizedState,e)},useTransition:function(){var e=ti(er)[0],t=Pe().memoizedState;return[e,t]},useMutableSource:Ia,useSyncExternalStore:Ua,useId:qa,unstable_isNewReconciler:!1},Yd={readContext:_e,useCallback:Xa,useContext:_e,useEffect:Uo,useImperativeHandle:Ya,useInsertionEffect:Qa,useLayoutEffect:Ka,useMemo:Ja,useReducer:ni,useRef:Ha,useState:function(){return ni(er)},useDebugValue:$o,useDeferredValue:function(e){var t=Pe();return X===null?t.memoizedState=e:Za(t,X.memoizedState,e)},useTransition:function(){var e=ni(er)[0],t=Pe().memoizedState;return[e,t]},useMutableSource:Ia,useSyncExternalStore:Ua,useId:qa,unstable_isNewReconciler:!1};function Re(e,t){if(e&&e.defaultProps){t=V({},t),e=e.defaultProps;for(var n in e)t[n]===void 0&&(t[n]=e[n]);return t}return t}function Ui(e,t,n,r){t=e.memoizedState,n=n(r,t),n=n==null?t:V({},t,n),e.memoizedState=n,e.lanes===0&&(e.updateQueue.baseState=n)}var Cl={isMounted:function(e){return(e=e._reactInternals)?Ft(e)===e:!1},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=se(),l=ft(e),i=Qe(r,l);i.payload=t,n!=null&&(i.callback=n),t=at(e,i,l),t!==null&&(Oe(t,e,l,r),Dr(t,e,l))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=se(),l=ft(e),i=Qe(r,l);i.tag=1,i.payload=t,n!=null&&(i.callback=n),t=at(e,i,l),t!==null&&(Oe(t,e,l,r),Dr(t,e,l))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=se(),r=ft(e),l=Qe(n,r);l.tag=2,t!=null&&(l.callback=t),t=at(e,l,r),t!==null&&(Oe(t,e,r,n),Dr(t,e,r))}};function Wu(e,t,n,r,l,i,o){return e=e.stateNode,typeof e.shouldComponentUpdate=="function"?e.shouldComponentUpdate(r,i,o):t.prototype&&t.prototype.isPureReactComponent?!Gn(n,r)||!Gn(l,i):!0}function nc(e,t,n){var r=!1,l=mt,i=t.contextType;return typeof i=="object"&&i!==null?i=_e(i):(l=he(t)?Lt:oe.current,r=t.contextTypes,i=(r=r!=null)?ln(e,l):mt),t=new t(n,i),e.memoizedState=t.state!==null&&t.state!==void 0?t.state:null,t.updater=Cl,e.stateNode=t,t._reactInternals=e,r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=l,e.__reactInternalMemoizedMaskedChildContext=i),t}function Hu(e,t,n,r){e=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(n,r),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&Cl.enqueueReplaceState(t,t.state,null)}function $i(e,t,n,r){var l=e.stateNode;l.props=n,l.state=e.memoizedState,l.refs={},zo(e);var i=t.contextType;typeof i=="object"&&i!==null?l.context=_e(i):(i=he(t)?Lt:oe.current,l.context=ln(e,i)),l.state=e.memoizedState,i=t.getDerivedStateFromProps,typeof i=="function"&&(Ui(e,t,i,n),l.state=e.memoizedState),typeof t.getDerivedStateFromProps=="function"||typeof l.getSnapshotBeforeUpdate=="function"||typeof l.UNSAFE_componentWillMount!="function"&&typeof l.componentWillMount!="function"||(t=l.state,typeof l.componentWillMount=="function"&&l.componentWillMount(),typeof l.UNSAFE_componentWillMount=="function"&&l.UNSAFE_componentWillMount(),t!==l.state&&Cl.enqueueReplaceState(l,l.state,null),ll(e,n,l,r),l.state=e.memoizedState),typeof l.componentDidMount=="function"&&(e.flags|=4194308)}function an(e,t){try{var n="",r=t;do n+=kf(r),r=r.return;while(r);var l=n}catch(i){l=`
-Error generating stack: `+i.message+`
-`+i.stack}return{value:e,source:t,stack:l,digest:null}}function ri(e,t,n){return{value:e,source:null,stack:n??null,digest:t??null}}function Bi(e,t){try{console.error(t.value)}catch(n){setTimeout(function(){throw n})}}var Xd=typeof WeakMap=="function"?WeakMap:Map;function rc(e,t,n){n=Qe(-1,n),n.tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){al||(al=!0,Ji=r),Bi(e,t)},n}function lc(e,t,n){n=Qe(-1,n),n.tag=3;var r=e.type.getDerivedStateFromError;if(typeof r=="function"){var l=t.value;n.payload=function(){return r(l)},n.callback=function(){Bi(e,t)}}var i=e.stateNode;return i!==null&&typeof i.componentDidCatch=="function"&&(n.callback=function(){Bi(e,t),typeof r!="function"&&(ct===null?ct=new Set([this]):ct.add(this));var o=t.stack;this.componentDidCatch(t.value,{componentStack:o!==null?o:""})}),n}function Qu(e,t,n){var r=e.pingCache;if(r===null){r=e.pingCache=new Xd;var l=new Set;r.set(t,l)}else l=r.get(t),l===void 0&&(l=new Set,r.set(t,l));l.has(n)||(l.add(n),e=ap.bind(null,e,t,n),t.then(e,e))}function Ku(e){do{var t;if((t=e.tag===13)&&(t=e.memoizedState,t=t!==null?t.dehydrated!==null:!0),t)return e;e=e.return}while(e!==null);return null}function Gu(e,t,n,r,l){return e.mode&1?(e.flags|=65536,e.lanes=l,e):(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,n.tag===1&&(n.alternate===null?n.tag=17:(t=Qe(-1,1),t.tag=2,at(n,t,1))),n.lanes|=1),e)}var Jd=Je.ReactCurrentOwner,de=!1;function ue(e,t,n,r){t.child=e===null?Ma(t,null,n,r):un(t,e.child,n,r)}function Yu(e,t,n,r,l){n=n.render;var i=t.ref;return tn(t,l),r=Do(e,t,n,r,i,l),n=Io(),e!==null&&!de?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,Xe(e,t,l)):($&&n&&Eo(t),t.flags|=1,ue(e,t,r,l),t.child)}function Xu(e,t,n,r,l){if(e===null){var i=n.type;return typeof i=="function"&&!Go(i)&&i.defaultProps===void 0&&n.compare===null&&n.defaultProps===void 0?(t.tag=15,t.type=i,ic(e,t,i,r,l)):(e=Vr(n.type,null,r,t,t.mode,l),e.ref=t.ref,e.return=t,t.child=e)}if(i=e.child,!(e.lanes&l)){var o=i.memoizedProps;if(n=n.compare,n=n!==null?n:Gn,n(o,r)&&e.ref===t.ref)return Xe(e,t,l)}return t.flags|=1,e=dt(i,r),e.ref=t.ref,e.return=t,t.child=e}function ic(e,t,n,r,l){if(e!==null){var i=e.memoizedProps;if(Gn(i,r)&&e.ref===t.ref)if(de=!1,t.pendingProps=r=i,(e.lanes&l)!==0)e.flags&131072&&(de=!0);else return t.lanes=e.lanes,Xe(e,t,l)}return Ai(e,t,n,r,l)}function oc(e,t,n){var r=t.pendingProps,l=r.children,i=e!==null?e.memoizedState:null;if(r.mode==="hidden")if(!(t.mode&1))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},D(Jt,ve),ve|=n;else{if(!(n&1073741824))return e=i!==null?i.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,D(Jt,ve),ve|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=i!==null?i.baseLanes:n,D(Jt,ve),ve|=r}else i!==null?(r=i.baseLanes|n,t.memoizedState=null):r=n,D(Jt,ve),ve|=r;return ue(e,t,l,n),t.child}function uc(e,t){var n=t.ref;(e===null&&n!==null||e!==null&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function Ai(e,t,n,r,l){var i=he(n)?Lt:oe.current;return i=ln(t,i),tn(t,l),n=Do(e,t,n,r,i,l),r=Io(),e!==null&&!de?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,Xe(e,t,l)):($&&r&&Eo(t),t.flags|=1,ue(e,t,n,l),t.child)}function Ju(e,t,n,r,l){if(he(n)){var i=!0;br(t)}else i=!1;if(tn(t,l),t.stateNode===null)$r(e,t),nc(t,n,r),$i(t,n,r,l),r=!0;else if(e===null){var o=t.stateNode,u=t.memoizedProps;o.props=u;var s=o.context,a=n.contextType;typeof a=="object"&&a!==null?a=_e(a):(a=he(n)?Lt:oe.current,a=ln(t,a));var h=n.getDerivedStateFromProps,p=typeof h=="function"||typeof o.getSnapshotBeforeUpdate=="function";p||typeof o.UNSAFE_componentWillReceiveProps!="function"&&typeof o.componentWillReceiveProps!="function"||(u!==r||s!==a)&&Hu(t,o,r,a),be=!1;var m=t.memoizedState;o.state=m,ll(t,r,o,l),s=t.memoizedState,u!==r||m!==s||pe.current||be?(typeof h=="function"&&(Ui(t,n,h,r),s=t.memoizedState),(u=be||Wu(t,n,u,r,m,s,a))?(p||typeof o.UNSAFE_componentWillMount!="function"&&typeof o.componentWillMount!="function"||(typeof o.componentWillMount=="function"&&o.componentWillMount(),typeof o.UNSAFE_componentWillMount=="function"&&o.UNSAFE_componentWillMount()),typeof o.componentDidMount=="function"&&(t.flags|=4194308)):(typeof o.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=s),o.props=r,o.state=s,o.context=a,r=u):(typeof o.componentDidMount=="function"&&(t.flags|=4194308),r=!1)}else{o=t.stateNode,Fa(e,t),u=t.memoizedProps,a=t.type===t.elementType?u:Re(t.type,u),o.props=a,p=t.pendingProps,m=o.context,s=n.contextType,typeof s=="object"&&s!==null?s=_e(s):(s=he(n)?Lt:oe.current,s=ln(t,s));var w=n.getDerivedStateFromProps;(h=typeof w=="function"||typeof o.getSnapshotBeforeUpdate=="function")||typeof o.UNSAFE_componentWillReceiveProps!="function"&&typeof o.componentWillReceiveProps!="function"||(u!==p||m!==s)&&Hu(t,o,r,s),be=!1,m=t.memoizedState,o.state=m,ll(t,r,o,l);var x=t.memoizedState;u!==p||m!==x||pe.current||be?(typeof w=="function"&&(Ui(t,n,w,r),x=t.memoizedState),(a=be||Wu(t,n,a,r,m,x,s)||!1)?(h||typeof o.UNSAFE_componentWillUpdate!="function"&&typeof o.componentWillUpdate!="function"||(typeof o.componentWillUpdate=="function"&&o.componentWillUpdate(r,x,s),typeof o.UNSAFE_componentWillUpdate=="function"&&o.UNSAFE_componentWillUpdate(r,x,s)),typeof o.componentDidUpdate=="function"&&(t.flags|=4),typeof o.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof o.componentDidUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=4),typeof o.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=x),o.props=r,o.state=x,o.context=s,r=a):(typeof o.componentDidUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=4),typeof o.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=1024),r=!1)}return Vi(e,t,n,r,i,l)}function Vi(e,t,n,r,l,i){uc(e,t);var o=(t.flags&128)!==0;if(!r&&!o)return l&&Fu(t,n,!1),Xe(e,t,i);r=t.stateNode,Jd.current=t;var u=o&&typeof n.getDerivedStateFromError!="function"?null:r.render();return t.flags|=1,e!==null&&o?(t.child=un(t,e.child,null,i),t.child=un(t,null,u,i)):ue(e,t,u,i),t.memoizedState=r.state,l&&Fu(t,n,!0),t.child}function sc(e){var t=e.stateNode;t.pendingContext?Ou(e,t.pendingContext,t.pendingContext!==t.context):t.context&&Ou(e,t.context,!1),To(e,t.containerInfo)}function Zu(e,t,n,r,l){return on(),_o(l),t.flags|=256,ue(e,t,n,r),t.child}var Wi={dehydrated:null,treeContext:null,retryLane:0};function Hi(e){return{baseLanes:e,cachePool:null,transitions:null}}function ac(e,t,n){var r=t.pendingProps,l=B.current,i=!1,o=(t.flags&128)!==0,u;if((u=o)||(u=e!==null&&e.memoizedState===null?!1:(l&2)!==0),u?(i=!0,t.flags&=-129):(e===null||e.memoizedState!==null)&&(l|=1),D(B,l&1),e===null)return Di(t),e=t.memoizedState,e!==null&&(e=e.dehydrated,e!==null)?(t.mode&1?e.data==="$!"?t.lanes=8:t.lanes=1073741824:t.lanes=1,null):(o=r.children,e=r.fallback,i?(r=t.mode,i=t.child,o={mode:"hidden",children:o},!(r&1)&&i!==null?(i.childLanes=0,i.pendingProps=o):i=_l(o,r,0,null),e=Nt(e,r,n,null),i.return=t,e.return=t,i.sibling=e,t.child=i,t.child.memoizedState=Hi(n),t.memoizedState=Wi,e):Bo(t,o));if(l=e.memoizedState,l!==null&&(u=l.dehydrated,u!==null))return Zd(e,t,o,r,u,l,n);if(i){i=r.fallback,o=t.mode,l=e.child,u=l.sibling;var s={mode:"hidden",children:r.children};return!(o&1)&&t.child!==l?(r=t.child,r.childLanes=0,r.pendingProps=s,t.deletions=null):(r=dt(l,s),r.subtreeFlags=l.subtreeFlags&14680064),u!==null?i=dt(u,i):(i=Nt(i,o,n,null),i.flags|=2),i.return=t,r.return=t,r.sibling=i,t.child=r,r=i,i=t.child,o=e.child.memoizedState,o=o===null?Hi(n):{baseLanes:o.baseLanes|n,cachePool:null,transitions:o.transitions},i.memoizedState=o,i.childLanes=e.childLanes&~n,t.memoizedState=Wi,r}return i=e.child,e=i.sibling,r=dt(i,{mode:"visible",children:r.children}),!(t.mode&1)&&(r.lanes=n),r.return=t,r.sibling=null,e!==null&&(n=t.deletions,n===null?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=r,t.memoizedState=null,r}function Bo(e,t){return t=_l({mode:"visible",children:t},e.mode,0,null),t.return=e,e.child=t}function Pr(e,t,n,r){return r!==null&&_o(r),un(t,e.child,null,n),e=Bo(t,t.pendingProps.children),e.flags|=2,t.memoizedState=null,e}function Zd(e,t,n,r,l,i,o){if(n)return t.flags&256?(t.flags&=-257,r=ri(Error(S(422))),Pr(e,t,o,r)):t.memoizedState!==null?(t.child=e.child,t.flags|=128,null):(i=r.fallback,l=t.mode,r=_l({mode:"visible",children:r.children},l,0,null),i=Nt(i,l,o,null),i.flags|=2,r.return=t,i.return=t,r.sibling=i,t.child=r,t.mode&1&&un(t,e.child,null,o),t.child.memoizedState=Hi(o),t.memoizedState=Wi,i);if(!(t.mode&1))return Pr(e,t,o,null);if(l.data==="$!"){if(r=l.nextSibling&&l.nextSibling.dataset,r)var u=r.dgst;return r=u,i=Error(S(419)),r=ri(i,r,void 0),Pr(e,t,o,r)}if(u=(o&e.childLanes)!==0,de||u){if(r=b,r!==null){switch(o&-o){case 4:l=2;break;case 16:l=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:l=32;break;case 536870912:l=268435456;break;default:l=0}l=l&(r.suspendedLanes|o)?0:l,l!==0&&l!==i.retryLane&&(i.retryLane=l,Ye(e,l),Oe(r,e,l,-1))}return Ko(),r=ri(Error(S(421))),Pr(e,t,o,r)}return l.data==="$?"?(t.flags|=128,t.child=e.child,t=cp.bind(null,e),l._reactRetry=t,null):(e=i.treeContext,ge=st(l.nextSibling),ye=t,$=!0,Te=null,e!==null&&(ke[Ce++]=We,ke[Ce++]=He,ke[Ce++]=Rt,We=e.id,He=e.overflow,Rt=t),t=Bo(t,r.children),t.flags|=4096,t)}function qu(e,t,n){e.lanes|=t;var r=e.alternate;r!==null&&(r.lanes|=t),Ii(e.return,t,n)}function li(e,t,n,r,l){var i=e.memoizedState;i===null?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:l}:(i.isBackwards=t,i.rendering=null,i.renderingStartTime=0,i.last=r,i.tail=n,i.tailMode=l)}function cc(e,t,n){var r=t.pendingProps,l=r.revealOrder,i=r.tail;if(ue(e,t,r.children,n),r=B.current,r&2)r=r&1|2,t.flags|=128;else{if(e!==null&&e.flags&128)e:for(e=t.child;e!==null;){if(e.tag===13)e.memoizedState!==null&&qu(e,n,t);else if(e.tag===19)qu(e,n,t);else if(e.child!==null){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;e.sibling===null;){if(e.return===null||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if(D(B,r),!(t.mode&1))t.memoizedState=null;else switch(l){case"forwards":for(n=t.child,l=null;n!==null;)e=n.alternate,e!==null&&il(e)===null&&(l=n),n=n.sibling;n=l,n===null?(l=t.child,t.child=null):(l=n.sibling,n.sibling=null),li(t,!1,l,n,i);break;case"backwards":for(n=null,l=t.child,t.child=null;l!==null;){if(e=l.alternate,e!==null&&il(e)===null){t.child=l;break}e=l.sibling,l.sibling=n,n=l,l=e}li(t,!0,n,null,i);break;case"together":li(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function $r(e,t){!(t.mode&1)&&e!==null&&(e.alternate=null,t.alternate=null,t.flags|=2)}function Xe(e,t,n){if(e!==null&&(t.dependencies=e.dependencies),Tt|=t.lanes,!(n&t.childLanes))return null;if(e!==null&&t.child!==e.child)throw Error(S(153));if(t.child!==null){for(e=t.child,n=dt(e,e.pendingProps),t.child=n,n.return=t;e.sibling!==null;)e=e.sibling,n=n.sibling=dt(e,e.pendingProps),n.return=t;n.sibling=null}return t.child}function qd(e,t,n){switch(t.tag){case 3:sc(t),on();break;case 5:Da(t);break;case 1:he(t.type)&&br(t);break;case 4:To(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,l=t.memoizedProps.value;D(nl,r._currentValue),r._currentValue=l;break;case 13:if(r=t.memoizedState,r!==null)return r.dehydrated!==null?(D(B,B.current&1),t.flags|=128,null):n&t.child.childLanes?ac(e,t,n):(D(B,B.current&1),e=Xe(e,t,n),e!==null?e.sibling:null);D(B,B.current&1);break;case 19:if(r=(n&t.childLanes)!==0,e.flags&128){if(r)return cc(e,t,n);t.flags|=128}if(l=t.memoizedState,l!==null&&(l.rendering=null,l.tail=null,l.lastEffect=null),D(B,B.current),r)break;return null;case 22:case 23:return t.lanes=0,oc(e,t,n)}return Xe(e,t,n)}var fc,Qi,dc,pc;fc=function(e,t){for(var n=t.child;n!==null;){if(n.tag===5||n.tag===6)e.appendChild(n.stateNode);else if(n.tag!==4&&n.child!==null){n.child.return=n,n=n.child;continue}if(n===t)break;for(;n.sibling===null;){if(n.return===null||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}};Qi=function(){};dc=function(e,t,n,r){var l=e.memoizedProps;if(l!==r){e=t.stateNode,_t(Be.current);var i=null;switch(n){case"input":l=pi(e,l),r=pi(e,r),i=[];break;case"select":l=V({},l,{value:void 0}),r=V({},r,{value:void 0}),i=[];break;case"textarea":l=vi(e,l),r=vi(e,r),i=[];break;default:typeof l.onClick!="function"&&typeof r.onClick=="function"&&(e.onclick=Zr)}yi(n,r);var o;n=null;for(a in l)if(!r.hasOwnProperty(a)&&l.hasOwnProperty(a)&&l[a]!=null)if(a==="style"){var u=l[a];for(o in u)u.hasOwnProperty(o)&&(n||(n={}),n[o]="")}else a!=="dangerouslySetInnerHTML"&&a!=="children"&&a!=="suppressContentEditableWarning"&&a!=="suppressHydrationWarning"&&a!=="autoFocus"&&(Bn.hasOwnProperty(a)?i||(i=[]):(i=i||[]).push(a,null));for(a in r){var s=r[a];if(u=l!=null?l[a]:void 0,r.hasOwnProperty(a)&&s!==u&&(s!=null||u!=null))if(a==="style")if(u){for(o in u)!u.hasOwnProperty(o)||s&&s.hasOwnProperty(o)||(n||(n={}),n[o]="");for(o in s)s.hasOwnProperty(o)&&u[o]!==s[o]&&(n||(n={}),n[o]=s[o])}else n||(i||(i=[]),i.push(a,n)),n=s;else a==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,u=u?u.__html:void 0,s!=null&&u!==s&&(i=i||[]).push(a,s)):a==="children"?typeof s!="string"&&typeof s!="number"||(i=i||[]).push(a,""+s):a!=="suppressContentEditableWarning"&&a!=="suppressHydrationWarning"&&(Bn.hasOwnProperty(a)?(s!=null&&a==="onScroll"&&I("scroll",e),i||u===s||(i=[])):(i=i||[]).push(a,s))}n&&(i=i||[]).push("style",n);var a=i;(t.updateQueue=a)&&(t.flags|=4)}};pc=function(e,t,n,r){n!==r&&(t.flags|=4)};function En(e,t){if(!$)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;t!==null;)t.alternate!==null&&(n=t),t=t.sibling;n===null?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;n!==null;)n.alternate!==null&&(r=n),n=n.sibling;r===null?t||e.tail===null?e.tail=null:e.tail.sibling=null:r.sibling=null}}function le(e){var t=e.alternate!==null&&e.alternate.child===e.child,n=0,r=0;if(t)for(var l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags&14680064,r|=l.flags&14680064,l.return=e,l=l.sibling;else for(l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags,r|=l.flags,l.return=e,l=l.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function bd(e,t,n){var r=t.pendingProps;switch(jo(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return le(t),null;case 1:return he(t.type)&&qr(),le(t),null;case 3:return r=t.stateNode,sn(),U(pe),U(oe),Oo(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),(e===null||e.child===null)&&(jr(t)?t.flags|=4:e===null||e.memoizedState.isDehydrated&&!(t.flags&256)||(t.flags|=1024,Te!==null&&(bi(Te),Te=null))),Qi(e,t),le(t),null;case 5:Mo(t);var l=_t(qn.current);if(n=t.type,e!==null&&t.stateNode!=null)dc(e,t,n,r,l),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(t.stateNode===null)throw Error(S(166));return le(t),null}if(e=_t(Be.current),jr(t)){r=t.stateNode,n=t.type;var i=t.memoizedProps;switch(r[Ue]=t,r[Jn]=i,e=(t.mode&1)!==0,n){case"dialog":I("cancel",r),I("close",r);break;case"iframe":case"object":case"embed":I("load",r);break;case"video":case"audio":for(l=0;l<Ln.length;l++)I(Ln[l],r);break;case"source":I("error",r);break;case"img":case"image":case"link":I("error",r),I("load",r);break;case"details":I("toggle",r);break;case"input":uu(r,i),I("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!i.multiple},I("invalid",r);break;case"textarea":au(r,i),I("invalid",r)}yi(n,i),l=null;for(var o in i)if(i.hasOwnProperty(o)){var u=i[o];o==="children"?typeof u=="string"?r.textContent!==u&&(i.suppressHydrationWarning!==!0&&Er(r.textContent,u,e),l=["children",u]):typeof u=="number"&&r.textContent!==""+u&&(i.suppressHydrationWarning!==!0&&Er(r.textContent,u,e),l=["children",""+u]):Bn.hasOwnProperty(o)&&u!=null&&o==="onScroll"&&I("scroll",r)}switch(n){case"input":vr(r),su(r,i,!0);break;case"textarea":vr(r),cu(r);break;case"select":case"option":break;default:typeof i.onClick=="function"&&(r.onclick=Zr)}r=l,t.updateQueue=r,r!==null&&(t.flags|=4)}else{o=l.nodeType===9?l:l.ownerDocument,e==="http://www.w3.org/1999/xhtml"&&(e=As(n)),e==="http://www.w3.org/1999/xhtml"?n==="script"?(e=o.createElement("div"),e.innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):typeof r.is=="string"?e=o.createElement(n,{is:r.is}):(e=o.createElement(n),n==="select"&&(o=e,r.multiple?o.multiple=!0:r.size&&(o.size=r.size))):e=o.createElementNS(e,n),e[Ue]=t,e[Jn]=r,fc(e,t,!1,!1),t.stateNode=e;e:{switch(o=wi(n,r),n){case"dialog":I("cancel",e),I("close",e),l=r;break;case"iframe":case"object":case"embed":I("load",e),l=r;break;case"video":case"audio":for(l=0;l<Ln.length;l++)I(Ln[l],e);l=r;break;case"source":I("error",e),l=r;break;case"img":case"image":case"link":I("error",e),I("load",e),l=r;break;case"details":I("toggle",e),l=r;break;case"input":uu(e,r),l=pi(e,r),I("invalid",e);break;case"option":l=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},l=V({},r,{value:void 0}),I("invalid",e);break;case"textarea":au(e,r),l=vi(e,r),I("invalid",e);break;default:l=r}yi(n,l),u=l;for(i in u)if(u.hasOwnProperty(i)){var s=u[i];i==="style"?Hs(e,s):i==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,s!=null&&Vs(e,s)):i==="children"?typeof s=="string"?(n!=="textarea"||s!=="")&&An(e,s):typeof s=="number"&&An(e,""+s):i!=="suppressContentEditableWarning"&&i!=="suppressHydrationWarning"&&i!=="autoFocus"&&(Bn.hasOwnProperty(i)?s!=null&&i==="onScroll"&&I("scroll",e):s!=null&&ao(e,i,s,o))}switch(n){case"input":vr(e),su(e,r,!1);break;case"textarea":vr(e),cu(e);break;case"option":r.value!=null&&e.setAttribute("value",""+ht(r.value));break;case"select":e.multiple=!!r.multiple,i=r.value,i!=null?Zt(e,!!r.multiple,i,!1):r.defaultValue!=null&&Zt(e,!!r.multiple,r.defaultValue,!0);break;default:typeof l.onClick=="function"&&(e.onclick=Zr)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}t.ref!==null&&(t.flags|=512,t.flags|=2097152)}return le(t),null;case 6:if(e&&t.stateNode!=null)pc(e,t,e.memoizedProps,r);else{if(typeof r!="string"&&t.stateNode===null)throw Error(S(166));if(n=_t(qn.current),_t(Be.current),jr(t)){if(r=t.stateNode,n=t.memoizedProps,r[Ue]=t,(i=r.nodeValue!==n)&&(e=ye,e!==null))switch(e.tag){case 3:Er(r.nodeValue,n,(e.mode&1)!==0);break;case 5:e.memoizedProps.suppressHydrationWarning!==!0&&Er(r.nodeValue,n,(e.mode&1)!==0)}i&&(t.flags|=4)}else r=(n.nodeType===9?n:n.ownerDocument).createTextNode(r),r[Ue]=t,t.stateNode=r}return le(t),null;case 13:if(U(B),r=t.memoizedState,e===null||e.memoizedState!==null&&e.memoizedState.dehydrated!==null){if($&&ge!==null&&t.mode&1&&!(t.flags&128))za(),on(),t.flags|=98560,i=!1;else if(i=jr(t),r!==null&&r.dehydrated!==null){if(e===null){if(!i)throw Error(S(318));if(i=t.memoizedState,i=i!==null?i.dehydrated:null,!i)throw Error(S(317));i[Ue]=t}else on(),!(t.flags&128)&&(t.memoizedState=null),t.flags|=4;le(t),i=!1}else Te!==null&&(bi(Te),Te=null),i=!0;if(!i)return t.flags&65536?t:null}return t.flags&128?(t.lanes=n,t):(r=r!==null,r!==(e!==null&&e.memoizedState!==null)&&r&&(t.child.flags|=8192,t.mode&1&&(e===null||B.current&1?J===0&&(J=3):Ko())),t.updateQueue!==null&&(t.flags|=4),le(t),null);case 4:return sn(),Qi(e,t),e===null&&Yn(t.stateNode.containerInfo),le(t),null;case 10:return Lo(t.type._context),le(t),null;case 17:return he(t.type)&&qr(),le(t),null;case 19:if(U(B),i=t.memoizedState,i===null)return le(t),null;if(r=(t.flags&128)!==0,o=i.rendering,o===null)if(r)En(i,!1);else{if(J!==0||e!==null&&e.flags&128)for(e=t.child;e!==null;){if(o=il(e),o!==null){for(t.flags|=128,En(i,!1),r=o.updateQueue,r!==null&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;n!==null;)i=n,e=r,i.flags&=14680066,o=i.alternate,o===null?(i.childLanes=0,i.lanes=e,i.child=null,i.subtreeFlags=0,i.memoizedProps=null,i.memoizedState=null,i.updateQueue=null,i.dependencies=null,i.stateNode=null):(i.childLanes=o.childLanes,i.lanes=o.lanes,i.child=o.child,i.subtreeFlags=0,i.deletions=null,i.memoizedProps=o.memoizedProps,i.memoizedState=o.memoizedState,i.updateQueue=o.updateQueue,i.type=o.type,e=o.dependencies,i.dependencies=e===null?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return D(B,B.current&1|2),t.child}e=e.sibling}i.tail!==null&&K()>cn&&(t.flags|=128,r=!0,En(i,!1),t.lanes=4194304)}else{if(!r)if(e=il(o),e!==null){if(t.flags|=128,r=!0,n=e.updateQueue,n!==null&&(t.updateQueue=n,t.flags|=4),En(i,!0),i.tail===null&&i.tailMode==="hidden"&&!o.alternate&&!$)return le(t),null}else 2*K()-i.renderingStartTime>cn&&n!==1073741824&&(t.flags|=128,r=!0,En(i,!1),t.lanes=4194304);i.isBackwards?(o.sibling=t.child,t.child=o):(n=i.last,n!==null?n.sibling=o:t.child=o,i.last=o)}return i.tail!==null?(t=i.tail,i.rendering=t,i.tail=t.sibling,i.renderingStartTime=K(),t.sibling=null,n=B.current,D(B,r?n&1|2:n&1),t):(le(t),null);case 22:case 23:return Qo(),r=t.memoizedState!==null,e!==null&&e.memoizedState!==null!==r&&(t.flags|=8192),r&&t.mode&1?ve&1073741824&&(le(t),t.subtreeFlags&6&&(t.flags|=8192)):le(t),null;case 24:return null;case 25:return null}throw Error(S(156,t.tag))}function ep(e,t){switch(jo(t),t.tag){case 1:return he(t.type)&&qr(),e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 3:return sn(),U(pe),U(oe),Oo(),e=t.flags,e&65536&&!(e&128)?(t.flags=e&-65537|128,t):null;case 5:return Mo(t),null;case 13:if(U(B),e=t.memoizedState,e!==null&&e.dehydrated!==null){if(t.alternate===null)throw Error(S(340));on()}return e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 19:return U(B),null;case 4:return sn(),null;case 10:return Lo(t.type._context),null;case 22:case 23:return Qo(),null;case 24:return null;default:return null}}var Nr=!1,ie=!1,tp=typeof WeakSet=="function"?WeakSet:Set,j=null;function Xt(e,t){var n=e.ref;if(n!==null)if(typeof n=="function")try{n(null)}catch(r){W(e,t,r)}else n.current=null}function Ki(e,t,n){try{n()}catch(r){W(e,t,r)}}var bu=!1;function np(e,t){if(Li=Yr,e=ya(),Co(e)){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{n=(n=e.ownerDocument)&&n.defaultView||window;var r=n.getSelection&&n.getSelection();if(r&&r.rangeCount!==0){n=r.anchorNode;var l=r.anchorOffset,i=r.focusNode;r=r.focusOffset;try{n.nodeType,i.nodeType}catch{n=null;break e}var o=0,u=-1,s=-1,a=0,h=0,p=e,m=null;t:for(;;){for(var w;p!==n||l!==0&&p.nodeType!==3||(u=o+l),p!==i||r!==0&&p.nodeType!==3||(s=o+r),p.nodeType===3&&(o+=p.nodeValue.length),(w=p.firstChild)!==null;)m=p,p=w;for(;;){if(p===e)break t;if(m===n&&++a===l&&(u=o),m===i&&++h===r&&(s=o),(w=p.nextSibling)!==null)break;p=m,m=p.parentNode}p=w}n=u===-1||s===-1?null:{start:u,end:s}}else n=null}n=n||{start:0,end:0}}else n=null;for(Ri={focusedElem:e,selectionRange:n},Yr=!1,j=t;j!==null;)if(t=j,e=t.child,(t.subtreeFlags&1028)!==0&&e!==null)e.return=t,j=e;else for(;j!==null;){t=j;try{var x=t.alternate;if(t.flags&1024)switch(t.tag){case 0:case 11:case 15:break;case 1:if(x!==null){var y=x.memoizedProps,C=x.memoizedState,f=t.stateNode,c=f.getSnapshotBeforeUpdate(t.elementType===t.type?y:Re(t.type,y),C);f.__reactInternalSnapshotBeforeUpdate=c}break;case 3:var d=t.stateNode.containerInfo;d.nodeType===1?d.textContent="":d.nodeType===9&&d.documentElement&&d.removeChild(d.documentElement);break;case 5:case 6:case 4:case 17:break;default:throw Error(S(163))}}catch(g){W(t,t.return,g)}if(e=t.sibling,e!==null){e.return=t.return,j=e;break}j=t.return}return x=bu,bu=!1,x}function In(e,t,n){var r=t.updateQueue;if(r=r!==null?r.lastEffect:null,r!==null){var l=r=r.next;do{if((l.tag&e)===e){var i=l.destroy;l.destroy=void 0,i!==void 0&&Ki(t,n,i)}l=l.next}while(l!==r)}}function El(e,t){if(t=t.updateQueue,t=t!==null?t.lastEffect:null,t!==null){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function Gi(e){var t=e.ref;if(t!==null){var n=e.stateNode;switch(e.tag){case 5:e=n;break;default:e=n}typeof t=="function"?t(e):t.current=e}}function hc(e){var t=e.alternate;t!==null&&(e.alternate=null,hc(t)),e.child=null,e.deletions=null,e.sibling=null,e.tag===5&&(t=e.stateNode,t!==null&&(delete t[Ue],delete t[Jn],delete t[Mi],delete t[Ud],delete t[$d])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function mc(e){return e.tag===5||e.tag===3||e.tag===4}function es(e){e:for(;;){for(;e.sibling===null;){if(e.return===null||mc(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;e.tag!==5&&e.tag!==6&&e.tag!==18;){if(e.flags&2||e.child===null||e.tag===4)continue e;e.child.return=e,e=e.child}if(!(e.flags&2))return e.stateNode}}function Yi(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.nodeType===8?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(n.nodeType===8?(t=n.parentNode,t.insertBefore(e,n)):(t=n,t.appendChild(e)),n=n._reactRootContainer,n!=null||t.onclick!==null||(t.onclick=Zr));else if(r!==4&&(e=e.child,e!==null))for(Yi(e,t,n),e=e.sibling;e!==null;)Yi(e,t,n),e=e.sibling}function Xi(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(r!==4&&(e=e.child,e!==null))for(Xi(e,t,n),e=e.sibling;e!==null;)Xi(e,t,n),e=e.sibling}var ee=null,ze=!1;function Ze(e,t,n){for(n=n.child;n!==null;)vc(e,t,n),n=n.sibling}function vc(e,t,n){if($e&&typeof $e.onCommitFiberUnmount=="function")try{$e.onCommitFiberUnmount(vl,n)}catch{}switch(n.tag){case 5:ie||Xt(n,t);case 6:var r=ee,l=ze;ee=null,Ze(e,t,n),ee=r,ze=l,ee!==null&&(ze?(e=ee,n=n.stateNode,e.nodeType===8?e.parentNode.removeChild(n):e.removeChild(n)):ee.removeChild(n.stateNode));break;case 18:ee!==null&&(ze?(e=ee,n=n.stateNode,e.nodeType===8?Zl(e.parentNode,n):e.nodeType===1&&Zl(e,n),Qn(e)):Zl(ee,n.stateNode));break;case 4:r=ee,l=ze,ee=n.stateNode.containerInfo,ze=!0,Ze(e,t,n),ee=r,ze=l;break;case 0:case 11:case 14:case 15:if(!ie&&(r=n.updateQueue,r!==null&&(r=r.lastEffect,r!==null))){l=r=r.next;do{var i=l,o=i.destroy;i=i.tag,o!==void 0&&(i&2||i&4)&&Ki(n,t,o),l=l.next}while(l!==r)}Ze(e,t,n);break;case 1:if(!ie&&(Xt(n,t),r=n.stateNode,typeof r.componentWillUnmount=="function"))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(u){W(n,t,u)}Ze(e,t,n);break;case 21:Ze(e,t,n);break;case 22:n.mode&1?(ie=(r=ie)||n.memoizedState!==null,Ze(e,t,n),ie=r):Ze(e,t,n);break;default:Ze(e,t,n)}}function ts(e){var t=e.updateQueue;if(t!==null){e.updateQueue=null;var n=e.stateNode;n===null&&(n=e.stateNode=new tp),t.forEach(function(r){var l=fp.bind(null,e,r);n.has(r)||(n.add(r),r.then(l,l))})}}function Le(e,t){var n=t.deletions;if(n!==null)for(var r=0;r<n.length;r++){var l=n[r];try{var i=e,o=t,u=o;e:for(;u!==null;){switch(u.tag){case 5:ee=u.stateNode,ze=!1;break e;case 3:ee=u.stateNode.containerInfo,ze=!0;break e;case 4:ee=u.stateNode.containerInfo,ze=!0;break e}u=u.return}if(ee===null)throw Error(S(160));vc(i,o,l),ee=null,ze=!1;var s=l.alternate;s!==null&&(s.return=null),l.return=null}catch(a){W(l,t,a)}}if(t.subtreeFlags&12854)for(t=t.child;t!==null;)gc(t,e),t=t.sibling}function gc(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(Le(t,e),De(e),r&4){try{In(3,e,e.return),El(3,e)}catch(y){W(e,e.return,y)}try{In(5,e,e.return)}catch(y){W(e,e.return,y)}}break;case 1:Le(t,e),De(e),r&512&&n!==null&&Xt(n,n.return);break;case 5:if(Le(t,e),De(e),r&512&&n!==null&&Xt(n,n.return),e.flags&32){var l=e.stateNode;try{An(l,"")}catch(y){W(e,e.return,y)}}if(r&4&&(l=e.stateNode,l!=null)){var i=e.memoizedProps,o=n!==null?n.memoizedProps:i,u=e.type,s=e.updateQueue;if(e.updateQueue=null,s!==null)try{u==="input"&&i.type==="radio"&&i.name!=null&&$s(l,i),wi(u,o);var a=wi(u,i);for(o=0;o<s.length;o+=2){var h=s[o],p=s[o+1];h==="style"?Hs(l,p):h==="dangerouslySetInnerHTML"?Vs(l,p):h==="children"?An(l,p):ao(l,h,p,a)}switch(u){case"input":hi(l,i);break;case"textarea":Bs(l,i);break;case"select":var m=l._wrapperState.wasMultiple;l._wrapperState.wasMultiple=!!i.multiple;var w=i.value;w!=null?Zt(l,!!i.multiple,w,!1):m!==!!i.multiple&&(i.defaultValue!=null?Zt(l,!!i.multiple,i.defaultValue,!0):Zt(l,!!i.multiple,i.multiple?[]:"",!1))}l[Jn]=i}catch(y){W(e,e.return,y)}}break;case 6:if(Le(t,e),De(e),r&4){if(e.stateNode===null)throw Error(S(162));l=e.stateNode,i=e.memoizedProps;try{l.nodeValue=i}catch(y){W(e,e.return,y)}}break;case 3:if(Le(t,e),De(e),r&4&&n!==null&&n.memoizedState.isDehydrated)try{Qn(t.containerInfo)}catch(y){W(e,e.return,y)}break;case 4:Le(t,e),De(e);break;case 13:Le(t,e),De(e),l=e.child,l.flags&8192&&(i=l.memoizedState!==null,l.stateNode.isHidden=i,!i||l.alternate!==null&&l.alternate.memoizedState!==null||(Wo=K())),r&4&&ts(e);break;case 22:if(h=n!==null&&n.memoizedState!==null,e.mode&1?(ie=(a=ie)||h,Le(t,e),ie=a):Le(t,e),De(e),r&8192){if(a=e.memoizedState!==null,(e.stateNode.isHidden=a)&&!h&&e.mode&1)for(j=e,h=e.child;h!==null;){for(p=j=h;j!==null;){switch(m=j,w=m.child,m.tag){case 0:case 11:case 14:case 15:In(4,m,m.return);break;case 1:Xt(m,m.return);var x=m.stateNode;if(typeof x.componentWillUnmount=="function"){r=m,n=m.return;try{t=r,x.props=t.memoizedProps,x.state=t.memoizedState,x.componentWillUnmount()}catch(y){W(r,n,y)}}break;case 5:Xt(m,m.return);break;case 22:if(m.memoizedState!==null){rs(p);continue}}w!==null?(w.return=m,j=w):rs(p)}h=h.sibling}e:for(h=null,p=e;;){if(p.tag===5){if(h===null){h=p;try{l=p.stateNode,a?(i=l.style,typeof i.setProperty=="function"?i.setProperty("display","none","important"):i.display="none"):(u=p.stateNode,s=p.memoizedProps.style,o=s!=null&&s.hasOwnProperty("display")?s.display:null,u.style.display=Ws("display",o))}catch(y){W(e,e.return,y)}}}else if(p.tag===6){if(h===null)try{p.stateNode.nodeValue=a?"":p.memoizedProps}catch(y){W(e,e.return,y)}}else if((p.tag!==22&&p.tag!==23||p.memoizedState===null||p===e)&&p.child!==null){p.child.return=p,p=p.child;continue}if(p===e)break e;for(;p.sibling===null;){if(p.return===null||p.return===e)break e;h===p&&(h=null),p=p.return}h===p&&(h=null),p.sibling.return=p.return,p=p.sibling}}break;case 19:Le(t,e),De(e),r&4&&ts(e);break;case 21:break;default:Le(t,e),De(e)}}function De(e){var t=e.flags;if(t&2){try{e:{for(var n=e.return;n!==null;){if(mc(n)){var r=n;break e}n=n.return}throw Error(S(160))}switch(r.tag){case 5:var l=r.stateNode;r.flags&32&&(An(l,""),r.flags&=-33);var i=es(e);Xi(e,i,l);break;case 3:case 4:var o=r.stateNode.containerInfo,u=es(e);Yi(e,u,o);break;default:throw Error(S(161))}}catch(s){W(e,e.return,s)}e.flags&=-3}t&4096&&(e.flags&=-4097)}function rp(e,t,n){j=e,yc(e)}function yc(e,t,n){for(var r=(e.mode&1)!==0;j!==null;){var l=j,i=l.child;if(l.tag===22&&r){var o=l.memoizedState!==null||Nr;if(!o){var u=l.alternate,s=u!==null&&u.memoizedState!==null||ie;u=Nr;var a=ie;if(Nr=o,(ie=s)&&!a)for(j=l;j!==null;)o=j,s=o.child,o.tag===22&&o.memoizedState!==null?ls(l):s!==null?(s.return=o,j=s):ls(l);for(;i!==null;)j=i,yc(i),i=i.sibling;j=l,Nr=u,ie=a}ns(e)}else l.subtreeFlags&8772&&i!==null?(i.return=l,j=i):ns(e)}}function ns(e){for(;j!==null;){var t=j;if(t.flags&8772){var n=t.alternate;try{if(t.flags&8772)switch(t.tag){case 0:case 11:case 15:ie||El(5,t);break;case 1:var r=t.stateNode;if(t.flags&4&&!ie)if(n===null)r.componentDidMount();else{var l=t.elementType===t.type?n.memoizedProps:Re(t.type,n.memoizedProps);r.componentDidUpdate(l,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var i=t.updateQueue;i!==null&&Bu(t,i,r);break;case 3:var o=t.updateQueue;if(o!==null){if(n=null,t.child!==null)switch(t.child.tag){case 5:n=t.child.stateNode;break;case 1:n=t.child.stateNode}Bu(t,o,n)}break;case 5:var u=t.stateNode;if(n===null&&t.flags&4){n=u;var s=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&n.focus();break;case"img":s.src&&(n.src=s.src)}}break;case 6:break;case 4:break;case 12:break;case 13:if(t.memoizedState===null){var a=t.alternate;if(a!==null){var h=a.memoizedState;if(h!==null){var p=h.dehydrated;p!==null&&Qn(p)}}}break;case 19:case 17:case 21:case 22:case 23:case 25:break;default:throw Error(S(163))}ie||t.flags&512&&Gi(t)}catch(m){W(t,t.return,m)}}if(t===e){j=null;break}if(n=t.sibling,n!==null){n.return=t.return,j=n;break}j=t.return}}function rs(e){for(;j!==null;){var t=j;if(t===e){j=null;break}var n=t.sibling;if(n!==null){n.return=t.return,j=n;break}j=t.return}}function ls(e){for(;j!==null;){var t=j;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{El(4,t)}catch(s){W(t,n,s)}break;case 1:var r=t.stateNode;if(typeof r.componentDidMount=="function"){var l=t.return;try{r.componentDidMount()}catch(s){W(t,l,s)}}var i=t.return;try{Gi(t)}catch(s){W(t,i,s)}break;case 5:var o=t.return;try{Gi(t)}catch(s){W(t,o,s)}}}catch(s){W(t,t.return,s)}if(t===e){j=null;break}var u=t.sibling;if(u!==null){u.return=t.return,j=u;break}j=t.return}}var lp=Math.ceil,sl=Je.ReactCurrentDispatcher,Ao=Je.ReactCurrentOwner,je=Je.ReactCurrentBatchConfig,O=0,b=null,G=null,te=0,ve=0,Jt=gt(0),J=0,nr=null,Tt=0,jl=0,Vo=0,Un=null,fe=null,Wo=0,cn=1/0,Ae=null,al=!1,Ji=null,ct=null,Lr=!1,rt=null,cl=0,$n=0,Zi=null,Br=-1,Ar=0;function se(){return O&6?K():Br!==-1?Br:Br=K()}function ft(e){return e.mode&1?O&2&&te!==0?te&-te:Ad.transition!==null?(Ar===0&&(Ar=na()),Ar):(e=F,e!==0||(e=window.event,e=e===void 0?16:aa(e.type)),e):1}function Oe(e,t,n,r){if(50<$n)throw $n=0,Zi=null,Error(S(185));or(e,n,r),(!(O&2)||e!==b)&&(e===b&&(!(O&2)&&(jl|=n),J===4&&tt(e,te)),me(e,r),n===1&&O===0&&!(t.mode&1)&&(cn=K()+500,Sl&&yt()))}function me(e,t){var n=e.callbackNode;Af(e,t);var r=Gr(e,e===b?te:0);if(r===0)n!==null&&pu(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(n!=null&&pu(n),t===1)e.tag===0?Bd(is.bind(null,e)):Na(is.bind(null,e)),Dd(function(){!(O&6)&&yt()}),n=null;else{switch(ra(r)){case 1:n=mo;break;case 4:n=ea;break;case 16:n=Kr;break;case 536870912:n=ta;break;default:n=Kr}n=_c(n,wc.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function wc(e,t){if(Br=-1,Ar=0,O&6)throw Error(S(327));var n=e.callbackNode;if(nn()&&e.callbackNode!==n)return null;var r=Gr(e,e===b?te:0);if(r===0)return null;if(r&30||r&e.expiredLanes||t)t=fl(e,r);else{t=r;var l=O;O|=2;var i=Sc();(b!==e||te!==t)&&(Ae=null,cn=K()+500,Pt(e,t));do try{up();break}catch(u){xc(e,u)}while(!0);No(),sl.current=i,O=l,G!==null?t=0:(b=null,te=0,t=J)}if(t!==0){if(t===2&&(l=Ei(e),l!==0&&(r=l,t=qi(e,l))),t===1)throw n=nr,Pt(e,0),tt(e,r),me(e,K()),n;if(t===6)tt(e,r);else{if(l=e.current.alternate,!(r&30)&&!ip(l)&&(t=fl(e,r),t===2&&(i=Ei(e),i!==0&&(r=i,t=qi(e,i))),t===1))throw n=nr,Pt(e,0),tt(e,r),me(e,K()),n;switch(e.finishedWork=l,e.finishedLanes=r,t){case 0:case 1:throw Error(S(345));case 2:Ct(e,fe,Ae);break;case 3:if(tt(e,r),(r&130023424)===r&&(t=Wo+500-K(),10<t)){if(Gr(e,0)!==0)break;if(l=e.suspendedLanes,(l&r)!==r){se(),e.pingedLanes|=e.suspendedLanes&l;break}e.timeoutHandle=Ti(Ct.bind(null,e,fe,Ae),t);break}Ct(e,fe,Ae);break;case 4:if(tt(e,r),(r&4194240)===r)break;for(t=e.eventTimes,l=-1;0<r;){var o=31-Me(r);i=1<<o,o=t[o],o>l&&(l=o),r&=~i}if(r=l,r=K()-r,r=(120>r?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*lp(r/1960))-r,10<r){e.timeoutHandle=Ti(Ct.bind(null,e,fe,Ae),r);break}Ct(e,fe,Ae);break;case 5:Ct(e,fe,Ae);break;default:throw Error(S(329))}}}return me(e,K()),e.callbackNode===n?wc.bind(null,e):null}function qi(e,t){var n=Un;return e.current.memoizedState.isDehydrated&&(Pt(e,t).flags|=256),e=fl(e,t),e!==2&&(t=fe,fe=n,t!==null&&bi(t)),e}function bi(e){fe===null?fe=e:fe.push.apply(fe,e)}function ip(e){for(var t=e;;){if(t.flags&16384){var n=t.updateQueue;if(n!==null&&(n=n.stores,n!==null))for(var r=0;r<n.length;r++){var l=n[r],i=l.getSnapshot;l=l.value;try{if(!Fe(i(),l))return!1}catch{return!1}}}if(n=t.child,t.subtreeFlags&16384&&n!==null)n.return=t,t=n;else{if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function tt(e,t){for(t&=~Vo,t&=~jl,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-Me(t),r=1<<n;e[n]=-1,t&=~r}}function is(e){if(O&6)throw Error(S(327));nn();var t=Gr(e,0);if(!(t&1))return me(e,K()),null;var n=fl(e,t);if(e.tag!==0&&n===2){var r=Ei(e);r!==0&&(t=r,n=qi(e,r))}if(n===1)throw n=nr,Pt(e,0),tt(e,t),me(e,K()),n;if(n===6)throw Error(S(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,Ct(e,fe,Ae),me(e,K()),null}function Ho(e,t){var n=O;O|=1;try{return e(t)}finally{O=n,O===0&&(cn=K()+500,Sl&&yt())}}function Mt(e){rt!==null&&rt.tag===0&&!(O&6)&&nn();var t=O;O|=1;var n=je.transition,r=F;try{if(je.transition=null,F=1,e)return e()}finally{F=r,je.transition=n,O=t,!(O&6)&&yt()}}function Qo(){ve=Jt.current,U(Jt)}function Pt(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(n!==-1&&(e.timeoutHandle=-1,Fd(n)),G!==null)for(n=G.return;n!==null;){var r=n;switch(jo(r),r.tag){case 1:r=r.type.childContextTypes,r!=null&&qr();break;case 3:sn(),U(pe),U(oe),Oo();break;case 5:Mo(r);break;case 4:sn();break;case 13:U(B);break;case 19:U(B);break;case 10:Lo(r.type._context);break;case 22:case 23:Qo()}n=n.return}if(b=e,G=e=dt(e.current,null),te=ve=t,J=0,nr=null,Vo=jl=Tt=0,fe=Un=null,jt!==null){for(t=0;t<jt.length;t++)if(n=jt[t],r=n.interleaved,r!==null){n.interleaved=null;var l=r.next,i=n.pending;if(i!==null){var o=i.next;i.next=l,r.next=o}n.pending=r}jt=null}return e}function xc(e,t){do{var n=G;try{if(No(),Ir.current=ul,ol){for(var r=A.memoizedState;r!==null;){var l=r.queue;l!==null&&(l.pending=null),r=r.next}ol=!1}if(zt=0,q=X=A=null,Dn=!1,bn=0,Ao.current=null,n===null||n.return===null){J=1,nr=t,G=null;break}e:{var i=e,o=n.return,u=n,s=t;if(t=te,u.flags|=32768,s!==null&&typeof s=="object"&&typeof s.then=="function"){var a=s,h=u,p=h.tag;if(!(h.mode&1)&&(p===0||p===11||p===15)){var m=h.alternate;m?(h.updateQueue=m.updateQueue,h.memoizedState=m.memoizedState,h.lanes=m.lanes):(h.updateQueue=null,h.memoizedState=null)}var w=Ku(o);if(w!==null){w.flags&=-257,Gu(w,o,u,i,t),w.mode&1&&Qu(i,a,t),t=w,s=a;var x=t.updateQueue;if(x===null){var y=new Set;y.add(s),t.updateQueue=y}else x.add(s);break e}else{if(!(t&1)){Qu(i,a,t),Ko();break e}s=Error(S(426))}}else if($&&u.mode&1){var C=Ku(o);if(C!==null){!(C.flags&65536)&&(C.flags|=256),Gu(C,o,u,i,t),_o(an(s,u));break e}}i=s=an(s,u),J!==4&&(J=2),Un===null?Un=[i]:Un.push(i),i=o;do{switch(i.tag){case 3:i.flags|=65536,t&=-t,i.lanes|=t;var f=rc(i,s,t);$u(i,f);break e;case 1:u=s;var c=i.type,d=i.stateNode;if(!(i.flags&128)&&(typeof c.getDerivedStateFromError=="function"||d!==null&&typeof d.componentDidCatch=="function"&&(ct===null||!ct.has(d)))){i.flags|=65536,t&=-t,i.lanes|=t;var g=lc(i,u,t);$u(i,g);break e}}i=i.return}while(i!==null)}Cc(n)}catch(k){t=k,G===n&&n!==null&&(G=n=n.return);continue}break}while(!0)}function Sc(){var e=sl.current;return sl.current=ul,e===null?ul:e}function Ko(){(J===0||J===3||J===2)&&(J=4),b===null||!(Tt&268435455)&&!(jl&268435455)||tt(b,te)}function fl(e,t){var n=O;O|=2;var r=Sc();(b!==e||te!==t)&&(Ae=null,Pt(e,t));do try{op();break}catch(l){xc(e,l)}while(!0);if(No(),O=n,sl.current=r,G!==null)throw Error(S(261));return b=null,te=0,J}function op(){for(;G!==null;)kc(G)}function up(){for(;G!==null&&!Tf();)kc(G)}function kc(e){var t=jc(e.alternate,e,ve);e.memoizedProps=e.pendingProps,t===null?Cc(e):G=t,Ao.current=null}function Cc(e){var t=e;do{var n=t.alternate;if(e=t.return,t.flags&32768){if(n=ep(n,t),n!==null){n.flags&=32767,G=n;return}if(e!==null)e.flags|=32768,e.subtreeFlags=0,e.deletions=null;else{J=6,G=null;return}}else if(n=bd(n,t,ve),n!==null){G=n;return}if(t=t.sibling,t!==null){G=t;return}G=t=e}while(t!==null);J===0&&(J=5)}function Ct(e,t,n){var r=F,l=je.transition;try{je.transition=null,F=1,sp(e,t,n,r)}finally{je.transition=l,F=r}return null}function sp(e,t,n,r){do nn();while(rt!==null);if(O&6)throw Error(S(327));n=e.finishedWork;var l=e.finishedLanes;if(n===null)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(S(177));e.callbackNode=null,e.callbackPriority=0;var i=n.lanes|n.childLanes;if(Vf(e,i),e===b&&(G=b=null,te=0),!(n.subtreeFlags&2064)&&!(n.flags&2064)||Lr||(Lr=!0,_c(Kr,function(){return nn(),null})),i=(n.flags&15990)!==0,n.subtreeFlags&15990||i){i=je.transition,je.transition=null;var o=F;F=1;var u=O;O|=4,Ao.current=null,np(e,n),gc(n,e),Nd(Ri),Yr=!!Li,Ri=Li=null,e.current=n,rp(n),Mf(),O=u,F=o,je.transition=i}else e.current=n;if(Lr&&(Lr=!1,rt=e,cl=l),i=e.pendingLanes,i===0&&(ct=null),Df(n.stateNode),me(e,K()),t!==null)for(r=e.onRecoverableError,n=0;n<t.length;n++)l=t[n],r(l.value,{componentStack:l.stack,digest:l.digest});if(al)throw al=!1,e=Ji,Ji=null,e;return cl&1&&e.tag!==0&&nn(),i=e.pendingLanes,i&1?e===Zi?$n++:($n=0,Zi=e):$n=0,yt(),null}function nn(){if(rt!==null){var e=ra(cl),t=je.transition,n=F;try{if(je.transition=null,F=16>e?16:e,rt===null)var r=!1;else{if(e=rt,rt=null,cl=0,O&6)throw Error(S(331));var l=O;for(O|=4,j=e.current;j!==null;){var i=j,o=i.child;if(j.flags&16){var u=i.deletions;if(u!==null){for(var s=0;s<u.length;s++){var a=u[s];for(j=a;j!==null;){var h=j;switch(h.tag){case 0:case 11:case 15:In(8,h,i)}var p=h.child;if(p!==null)p.return=h,j=p;else for(;j!==null;){h=j;var m=h.sibling,w=h.return;if(hc(h),h===a){j=null;break}if(m!==null){m.return=w,j=m;break}j=w}}}var x=i.alternate;if(x!==null){var y=x.child;if(y!==null){x.child=null;do{var C=y.sibling;y.sibling=null,y=C}while(y!==null)}}j=i}}if(i.subtreeFlags&2064&&o!==null)o.return=i,j=o;else e:for(;j!==null;){if(i=j,i.flags&2048)switch(i.tag){case 0:case 11:case 15:In(9,i,i.return)}var f=i.sibling;if(f!==null){f.return=i.return,j=f;break e}j=i.return}}var c=e.current;for(j=c;j!==null;){o=j;var d=o.child;if(o.subtreeFlags&2064&&d!==null)d.return=o,j=d;else e:for(o=c;j!==null;){if(u=j,u.flags&2048)try{switch(u.tag){case 0:case 11:case 15:El(9,u)}}catch(k){W(u,u.return,k)}if(u===o){j=null;break e}var g=u.sibling;if(g!==null){g.return=u.return,j=g;break e}j=u.return}}if(O=l,yt(),$e&&typeof $e.onPostCommitFiberRoot=="function")try{$e.onPostCommitFiberRoot(vl,e)}catch{}r=!0}return r}finally{F=n,je.transition=t}}return!1}function os(e,t,n){t=an(n,t),t=rc(e,t,1),e=at(e,t,1),t=se(),e!==null&&(or(e,1,t),me(e,t))}function W(e,t,n){if(e.tag===3)os(e,e,n);else for(;t!==null;){if(t.tag===3){os(t,e,n);break}else if(t.tag===1){var r=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof r.componentDidCatch=="function"&&(ct===null||!ct.has(r))){e=an(n,e),e=lc(t,e,1),t=at(t,e,1),e=se(),t!==null&&(or(t,1,e),me(t,e));break}}t=t.return}}function ap(e,t,n){var r=e.pingCache;r!==null&&r.delete(t),t=se(),e.pingedLanes|=e.suspendedLanes&n,b===e&&(te&n)===n&&(J===4||J===3&&(te&130023424)===te&&500>K()-Wo?Pt(e,0):Vo|=n),me(e,t)}function Ec(e,t){t===0&&(e.mode&1?(t=wr,wr<<=1,!(wr&130023424)&&(wr=4194304)):t=1);var n=se();e=Ye(e,t),e!==null&&(or(e,t,n),me(e,n))}function cp(e){var t=e.memoizedState,n=0;t!==null&&(n=t.retryLane),Ec(e,n)}function fp(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,l=e.memoizedState;l!==null&&(n=l.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(S(314))}r!==null&&r.delete(t),Ec(e,n)}var jc;jc=function(e,t,n){if(e!==null)if(e.memoizedProps!==t.pendingProps||pe.current)de=!0;else{if(!(e.lanes&n)&&!(t.flags&128))return de=!1,qd(e,t,n);de=!!(e.flags&131072)}else de=!1,$&&t.flags&1048576&&La(t,tl,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;$r(e,t),e=t.pendingProps;var l=ln(t,oe.current);tn(t,n),l=Do(null,t,r,e,l,n);var i=Io();return t.flags|=1,typeof l=="object"&&l!==null&&typeof l.render=="function"&&l.$$typeof===void 0?(t.tag=1,t.memoizedState=null,t.updateQueue=null,he(r)?(i=!0,br(t)):i=!1,t.memoizedState=l.state!==null&&l.state!==void 0?l.state:null,zo(t),l.updater=Cl,t.stateNode=l,l._reactInternals=t,$i(t,r,e,n),t=Vi(null,t,r,!0,i,n)):(t.tag=0,$&&i&&Eo(t),ue(null,t,l,n),t=t.child),t;case 16:r=t.elementType;e:{switch($r(e,t),e=t.pendingProps,l=r._init,r=l(r._payload),t.type=r,l=t.tag=pp(r),e=Re(r,e),l){case 0:t=Ai(null,t,r,e,n);break e;case 1:t=Ju(null,t,r,e,n);break e;case 11:t=Yu(null,t,r,e,n);break e;case 14:t=Xu(null,t,r,Re(r.type,e),n);break e}throw Error(S(306,r,""))}return t;case 0:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Re(r,l),Ai(e,t,r,l,n);case 1:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Re(r,l),Ju(e,t,r,l,n);case 3:e:{if(sc(t),e===null)throw Error(S(387));r=t.pendingProps,i=t.memoizedState,l=i.element,Fa(e,t),ll(t,r,null,n);var o=t.memoizedState;if(r=o.element,i.isDehydrated)if(i={element:r,isDehydrated:!1,cache:o.cache,pendingSuspenseBoundaries:o.pendingSuspenseBoundaries,transitions:o.transitions},t.updateQueue.baseState=i,t.memoizedState=i,t.flags&256){l=an(Error(S(423)),t),t=Zu(e,t,r,n,l);break e}else if(r!==l){l=an(Error(S(424)),t),t=Zu(e,t,r,n,l);break e}else for(ge=st(t.stateNode.containerInfo.firstChild),ye=t,$=!0,Te=null,n=Ma(t,null,r,n),t.child=n;n;)n.flags=n.flags&-3|4096,n=n.sibling;else{if(on(),r===l){t=Xe(e,t,n);break e}ue(e,t,r,n)}t=t.child}return t;case 5:return Da(t),e===null&&Di(t),r=t.type,l=t.pendingProps,i=e!==null?e.memoizedProps:null,o=l.children,zi(r,l)?o=null:i!==null&&zi(r,i)&&(t.flags|=32),uc(e,t),ue(e,t,o,n),t.child;case 6:return e===null&&Di(t),null;case 13:return ac(e,t,n);case 4:return To(t,t.stateNode.containerInfo),r=t.pendingProps,e===null?t.child=un(t,null,r,n):ue(e,t,r,n),t.child;case 11:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Re(r,l),Yu(e,t,r,l,n);case 7:return ue(e,t,t.pendingProps,n),t.child;case 8:return ue(e,t,t.pendingProps.children,n),t.child;case 12:return ue(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,l=t.pendingProps,i=t.memoizedProps,o=l.value,D(nl,r._currentValue),r._currentValue=o,i!==null)if(Fe(i.value,o)){if(i.children===l.children&&!pe.current){t=Xe(e,t,n);break e}}else for(i=t.child,i!==null&&(i.return=t);i!==null;){var u=i.dependencies;if(u!==null){o=i.child;for(var s=u.firstContext;s!==null;){if(s.context===r){if(i.tag===1){s=Qe(-1,n&-n),s.tag=2;var a=i.updateQueue;if(a!==null){a=a.shared;var h=a.pending;h===null?s.next=s:(s.next=h.next,h.next=s),a.pending=s}}i.lanes|=n,s=i.alternate,s!==null&&(s.lanes|=n),Ii(i.return,n,t),u.lanes|=n;break}s=s.next}}else if(i.tag===10)o=i.type===t.type?null:i.child;else if(i.tag===18){if(o=i.return,o===null)throw Error(S(341));o.lanes|=n,u=o.alternate,u!==null&&(u.lanes|=n),Ii(o,n,t),o=i.sibling}else o=i.child;if(o!==null)o.return=i;else for(o=i;o!==null;){if(o===t){o=null;break}if(i=o.sibling,i!==null){i.return=o.return,o=i;break}o=o.return}i=o}ue(e,t,l.children,n),t=t.child}return t;case 9:return l=t.type,r=t.pendingProps.children,tn(t,n),l=_e(l),r=r(l),t.flags|=1,ue(e,t,r,n),t.child;case 14:return r=t.type,l=Re(r,t.pendingProps),l=Re(r.type,l),Xu(e,t,r,l,n);case 15:return ic(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Re(r,l),$r(e,t),t.tag=1,he(r)?(e=!0,br(t)):e=!1,tn(t,n),nc(t,r,l),$i(t,r,l,n),Vi(null,t,r,!0,e,n);case 19:return cc(e,t,n);case 22:return oc(e,t,n)}throw Error(S(156,t.tag))};function _c(e,t){return bs(e,t)}function dp(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Ee(e,t,n,r){return new dp(e,t,n,r)}function Go(e){return e=e.prototype,!(!e||!e.isReactComponent)}function pp(e){if(typeof e=="function")return Go(e)?1:0;if(e!=null){if(e=e.$$typeof,e===fo)return 11;if(e===po)return 14}return 2}function dt(e,t){var n=e.alternate;return n===null?(n=Ee(e.tag,t,e.key,e.mode),n.elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=e.flags&14680064,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Vr(e,t,n,r,l,i){var o=2;if(r=e,typeof e=="function")Go(e)&&(o=1);else if(typeof e=="string")o=5;else e:switch(e){case Bt:return Nt(n.children,l,i,t);case co:o=8,l|=8;break;case ai:return e=Ee(12,n,t,l|2),e.elementType=ai,e.lanes=i,e;case ci:return e=Ee(13,n,t,l),e.elementType=ci,e.lanes=i,e;case fi:return e=Ee(19,n,t,l),e.elementType=fi,e.lanes=i,e;case Ds:return _l(n,l,i,t);default:if(typeof e=="object"&&e!==null)switch(e.$$typeof){case Os:o=10;break e;case Fs:o=9;break e;case fo:o=11;break e;case po:o=14;break e;case qe:o=16,r=null;break e}throw Error(S(130,e==null?e:typeof e,""))}return t=Ee(o,n,t,l),t.elementType=e,t.type=r,t.lanes=i,t}function Nt(e,t,n,r){return e=Ee(7,e,r,t),e.lanes=n,e}function _l(e,t,n,r){return e=Ee(22,e,r,t),e.elementType=Ds,e.lanes=n,e.stateNode={isHidden:!1},e}function ii(e,t,n){return e=Ee(6,e,null,t),e.lanes=n,e}function oi(e,t,n){return t=Ee(4,e.children!==null?e.children:[],e.key,t),t.lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function hp(e,t,n,r,l){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=Bl(0),this.expirationTimes=Bl(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=Bl(0),this.identifierPrefix=r,this.onRecoverableError=l,this.mutableSourceEagerHydrationData=null}function Yo(e,t,n,r,l,i,o,u,s){return e=new hp(e,t,n,u,s),t===1?(t=1,i===!0&&(t|=8)):t=0,i=Ee(3,null,null,t),e.current=i,i.stateNode=e,i.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},zo(i),e}function mp(e,t,n){var r=3<arguments.length&&arguments[3]!==void 0?arguments[3]:null;return{$$typeof:$t,key:r==null?null:""+r,children:e,containerInfo:t,implementation:n}}function Pc(e){if(!e)return mt;e=e._reactInternals;e:{if(Ft(e)!==e||e.tag!==1)throw Error(S(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(he(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(t!==null);throw Error(S(171))}if(e.tag===1){var n=e.type;if(he(n))return Pa(e,n,t)}return t}function Nc(e,t,n,r,l,i,o,u,s){return e=Yo(n,r,!0,e,l,i,o,u,s),e.context=Pc(null),n=e.current,r=se(),l=ft(n),i=Qe(r,l),i.callback=t??null,at(n,i,l),e.current.lanes=l,or(e,l,r),me(e,r),e}function Pl(e,t,n,r){var l=t.current,i=se(),o=ft(l);return n=Pc(n),t.context===null?t.context=n:t.pendingContext=n,t=Qe(i,o),t.payload={element:e},r=r===void 0?null:r,r!==null&&(t.callback=r),e=at(l,t,o),e!==null&&(Oe(e,l,o,i),Dr(e,l,o)),o}function dl(e){if(e=e.current,!e.child)return null;switch(e.child.tag){case 5:return e.child.stateNode;default:return e.child.stateNode}}function us(e,t){if(e=e.memoizedState,e!==null&&e.dehydrated!==null){var n=e.retryLane;e.retryLane=n!==0&&n<t?n:t}}function Xo(e,t){us(e,t),(e=e.alternate)&&us(e,t)}function vp(){return null}var Lc=typeof reportError=="function"?reportError:function(e){console.error(e)};function Jo(e){this._internalRoot=e}Nl.prototype.render=Jo.prototype.render=function(e){var t=this._internalRoot;if(t===null)throw Error(S(409));Pl(e,t,null,null)};Nl.prototype.unmount=Jo.prototype.unmount=function(){var e=this._internalRoot;if(e!==null){this._internalRoot=null;var t=e.containerInfo;Mt(function(){Pl(null,e,null,null)}),t[Ge]=null}};function Nl(e){this._internalRoot=e}Nl.prototype.unstable_scheduleHydration=function(e){if(e){var t=oa();e={blockedOn:null,target:e,priority:t};for(var n=0;n<et.length&&t!==0&&t<et[n].priority;n++);et.splice(n,0,e),n===0&&sa(e)}};function Zo(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11)}function Ll(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11&&(e.nodeType!==8||e.nodeValue!==" react-mount-point-unstable "))}function ss(){}function gp(e,t,n,r,l){if(l){if(typeof r=="function"){var i=r;r=function(){var a=dl(o);i.call(a)}}var o=Nc(t,r,e,0,null,!1,!1,"",ss);return e._reactRootContainer=o,e[Ge]=o.current,Yn(e.nodeType===8?e.parentNode:e),Mt(),o}for(;l=e.lastChild;)e.removeChild(l);if(typeof r=="function"){var u=r;r=function(){var a=dl(s);u.call(a)}}var s=Yo(e,0,!1,null,null,!1,!1,"",ss);return e._reactRootContainer=s,e[Ge]=s.current,Yn(e.nodeType===8?e.parentNode:e),Mt(function(){Pl(t,s,n,r)}),s}function Rl(e,t,n,r,l){var i=n._reactRootContainer;if(i){var o=i;if(typeof l=="function"){var u=l;l=function(){var s=dl(o);u.call(s)}}Pl(t,o,e,l)}else o=gp(n,t,e,l,r);return dl(o)}la=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=Nn(t.pendingLanes);n!==0&&(vo(t,n|1),me(t,K()),!(O&6)&&(cn=K()+500,yt()))}break;case 13:Mt(function(){var r=Ye(e,1);if(r!==null){var l=se();Oe(r,e,1,l)}}),Xo(e,1)}};go=function(e){if(e.tag===13){var t=Ye(e,134217728);if(t!==null){var n=se();Oe(t,e,134217728,n)}Xo(e,134217728)}};ia=function(e){if(e.tag===13){var t=ft(e),n=Ye(e,t);if(n!==null){var r=se();Oe(n,e,t,r)}Xo(e,t)}};oa=function(){return F};ua=function(e,t){var n=F;try{return F=e,t()}finally{F=n}};Si=function(e,t,n){switch(t){case"input":if(hi(e,n),t=n.name,n.type==="radio"&&t!=null){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var l=xl(r);if(!l)throw Error(S(90));Us(r),hi(r,l)}}}break;case"textarea":Bs(e,n);break;case"select":t=n.value,t!=null&&Zt(e,!!n.multiple,t,!1)}};Gs=Ho;Ys=Mt;var yp={usingClientEntryPoint:!1,Events:[sr,Ht,xl,Qs,Ks,Ho]},jn={findFiberByHostInstance:Et,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},wp={bundleType:jn.bundleType,version:jn.version,rendererPackageName:jn.rendererPackageName,rendererConfig:jn.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:Je.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return e=Zs(e),e===null?null:e.stateNode},findFiberByHostInstance:jn.findFiberByHostInstance||vp,findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var Rr=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!Rr.isDisabled&&Rr.supportsFiber)try{vl=Rr.inject(wp),$e=Rr}catch{}}xe.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=yp;xe.createPortal=function(e,t){var n=2<arguments.length&&arguments[2]!==void 0?arguments[2]:null;if(!Zo(t))throw Error(S(200));return mp(e,t,null,n)};xe.createRoot=function(e,t){if(!Zo(e))throw Error(S(299));var n=!1,r="",l=Lc;return t!=null&&(t.unstable_strictMode===!0&&(n=!0),t.identifierPrefix!==void 0&&(r=t.identifierPrefix),t.onRecoverableError!==void 0&&(l=t.onRecoverableError)),t=Yo(e,1,!1,null,null,n,!1,r,l),e[Ge]=t.current,Yn(e.nodeType===8?e.parentNode:e),new Jo(t)};xe.findDOMNode=function(e){if(e==null)return null;if(e.nodeType===1)return e;var t=e._reactInternals;if(t===void 0)throw typeof e.render=="function"?Error(S(188)):(e=Object.keys(e).join(","),Error(S(268,e)));return e=Zs(t),e=e===null?null:e.stateNode,e};xe.flushSync=function(e){return Mt(e)};xe.hydrate=function(e,t,n){if(!Ll(t))throw Error(S(200));return Rl(null,e,t,!0,n)};xe.hydrateRoot=function(e,t,n){if(!Zo(e))throw Error(S(405));var r=n!=null&&n.hydratedSources||null,l=!1,i="",o=Lc;if(n!=null&&(n.unstable_strictMode===!0&&(l=!0),n.identifierPrefix!==void 0&&(i=n.identifierPrefix),n.onRecoverableError!==void 0&&(o=n.onRecoverableError)),t=Nc(t,null,e,1,n??null,l,!1,i,o),e[Ge]=t.current,Yn(e),r)for(e=0;e<r.length;e++)n=r[e],l=n._getVersion,l=l(n._source),t.mutableSourceEagerHydrationData==null?t.mutableSourceEagerHydrationData=[n,l]:t.mutableSourceEagerHydrationData.push(n,l);return new Nl(t)};xe.render=function(e,t,n){if(!Ll(t))throw Error(S(200));return Rl(null,e,t,!1,n)};xe.unmountComponentAtNode=function(e){if(!Ll(e))throw Error(S(40));return e._reactRootContainer?(Mt(function(){Rl(null,null,e,!1,function(){e._reactRootContainer=null,e[Ge]=null})}),!0):!1};xe.unstable_batchedUpdates=Ho;xe.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!Ll(n))throw Error(S(200));if(e==null||e._reactInternals===void 0)throw Error(S(38));return Rl(e,t,n,!1,r)};xe.version="18.3.1-next-f1338f8080-20240426";function Rc(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(Rc)}catch(e){console.error(e)}}Rc(),Rs.exports=xe;var xp=Rs.exports,zc,as=xp;zc=as.createRoot,as.hydrateRoot;const hn="w-[25%] h-fit rounded-md grayscale transition-all duration-300 hover:grayscale-0 sm:w-[25%] lg:w-[15%]",Sp=v.jsx("a",{href:"https://developer.mozilla.org/en-US/docs/Web/JavaScript",target:"_blank",rel:"noreferrer",alt:"Javascript Documentation",className:hn+" flex items-center justify-center self-center",children:v.jsxs("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"100",height:"100",viewBox:"0 0 48 48",children:[v.jsx("title",{children:"Javascript"}),v.jsx("path",{fill:"#ffd600",d:"M6,42V6h36v36H6z"}),v.jsx("path",{fill:"#000001",d:"M29.538 32.947c.692 1.124 1.444 2.201 3.037 2.201 1.338 0 2.04-.665 2.04-1.585 0-1.101-.726-1.492-2.198-2.133l-.807-.344c-2.329-.988-3.878-2.226-3.878-4.841 0-2.41 1.845-4.244 4.728-4.244 2.053 0 3.528.711 4.592 2.573l-2.514 1.607c-.553-.988-1.151-1.377-2.078-1.377-.946 0-1.545.597-1.545 1.377 0 .964.6 1.354 1.985 1.951l.807.344C36.452 29.645 38 30.839 38 33.523 38 36.415 35.716 38 32.65 38c-2.999 0-4.702-1.505-5.65-3.368L29.538 32.947zM17.952 33.029c.506.906 1.275 1.603 2.381 1.603 1.058 0 1.667-.418 1.667-2.043V22h3.333v11.101c0 3.367-1.953 4.899-4.805 4.899-2.577 0-4.437-1.746-5.195-3.368L17.952 33.029z"})]})}),kp=v.jsx("a",{href:"https://supabase.com/",target:"_blank",rel:"noreferrer",alt:"Supabase Documentation",className:hn+" flex items-center justify-center self-center",children:v.jsxs("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"100",height:"100",viewBox:"0 0 48 48",children:[v.jsx("title",{children:"Supabase"}),v.jsx("g",{id:"Ð¡Ð»Ð¾Ð¹_1",children:v.jsx("path",{fill:"#71c674",d:"M43.9,22.5L27.5,43.3c-1.2,1.4-3.5,0.6-3.5-1.2V30H6.6c-2.1,0-3.3-2.5-2-4.1L20.7,5.1c1.1-1.5,3.4-0.7,3.4,1.2 v12.1h17.7C44.1,18.4,45.3,20.8,43.9,22.5z"})})]})}),Cp=v.jsx("a",{href:"https://www.typescriptlang.org/",target:"_blank",rel:"noreferrer",alt:"Typescript Documentation",className:hn+" flex items-center justify-center self-center",children:v.jsxs("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"100",height:"100",viewBox:"0 0 48 48",children:[v.jsx("title",{children:"Typescript"}),v.jsx("rect",{width:"36",height:"36",x:"6",y:"6",fill:"#1976d2"}),v.jsx("polygon",{fill:"#fff",points:"27.49,22 14.227,22 14.227,25.264 18.984,25.264 18.984,40 22.753,40 22.753,25.264 27.49,25.264"}),v.jsx("path",{fill:"#fff",d:"M39.194,26.084c0,0-1.787-1.192-3.807-1.192s-2.747,0.96-2.747,1.986 c0,2.648,7.381,2.383,7.381,7.712c0,8.209-11.254,4.568-11.254,4.568V35.22c0,0,2.152,1.622,4.733,1.622s2.483-1.688,2.483-1.92 c0-2.449-7.315-2.449-7.315-7.878c0-7.381,10.658-4.469,10.658-4.469L39.194,26.084z"})]})}),Ep=v.jsx("a",{href:"https://flutter.dev/",target:"_blank",rel:"noreferrer",alt:"Flutter Documentation",className:hn+" flex items-center justify-center self-center",children:v.jsxs("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"100",height:"100",viewBox:"0 0 48 48",children:[v.jsx("title",{children:"Flutter"}),v.jsx("polygon",{fill:"#40c4ff",points:"26,4 6,24 12,30 38,4"}),v.jsx("polygon",{fill:"#40c4ff",points:"38,22 27,33 21,27 26,22"}),v.jsx("rect",{width:"8.485",height:"8.485",x:"16.757",y:"28.757",fill:"#03a9f4",transform:"rotate(-45.001 21 33)"}),v.jsx("polygon",{fill:"#01579b",points:"38,44 26,44 21,39 27,33"}),v.jsx("polygon",{fill:"#084994",points:"21,39 30,36 27,33"})]})}),jp=v.jsx("a",{href:"https://tailwindcss.com/",target:"_blank",rel:"noreferrer",alt:"Tailwind Documentation",className:hn+" flex items-center justify-center self-center",children:v.jsxs("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"100",height:"100",viewBox:"0 0 48 48",children:[v.jsx("title",{children:"Tailwind"}),v.jsx("path",{fill:"#00acc1",d:"M24,9.604c-6.4,0-10.4,3.199-12,9.597c2.4-3.199,5.2-4.398,8.4-3.599 c1.826,0.456,3.131,1.781,4.576,3.247C27.328,21.236,30.051,24,36,24c6.4,0,10.4-3.199,12-9.598c-2.4,3.199-5.2,4.399-8.4,3.6 c-1.825-0.456-3.13-1.781-4.575-3.247C32.672,12.367,29.948,9.604,24,9.604L24,9.604z M12,24c-6.4,0-10.4,3.199-12,9.598 c2.4-3.199,5.2-4.399,8.4-3.599c1.825,0.457,3.13,1.781,4.575,3.246c2.353,2.388,5.077,5.152,11.025,5.152 c6.4,0,10.4-3.199,12-9.598c-2.4,3.199-5.2,4.399-8.4,3.599c-1.826-0.456-3.131-1.781-4.576-3.246C20.672,26.764,17.949,24,12,24 L12,24z"})]})}),_p=v.jsx("a",{href:"https://redux.js.org/",target:"_blank",rel:"noreferrer",alt:"Redux Documentation",className:hn+" flex items-center justify-center self-center",children:v.jsxs("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"100",height:"100",viewBox:"0 0 48 48",children:[v.jsx("title",{children:"Redux"}),v.jsx("path",{fill:"#7e57c2",d:"M23,4c-6.617,0-12,7.27-12,16.205c0,4.834,1.582,9.169,4.078,12.136C15.03,32.554,15,32.773,15,33	c0,1.657,1.343,3,3,3s3-1.343,3-3s-1.343-3-3-3c-0.315,0-0.612,0.062-0.897,0.152C15.206,27.731,14,24.175,14,20.205	C14,12.924,18.037,7,23,7c3.837,0,7.111,3.547,8.404,8.518c1.122,0.346,2.237,0.782,3.33,1.308C33.579,9.508,28.759,4,23,4z"}),v.jsx("path",{fill:"#7e57c2",d:"M35.507,20.084c-3.947-2.392-8.374-3.442-12.182-2.959C22.775,16.444,21.943,16,21,16	c-1.657,0-3,1.343-3,3s1.343,3,3,3c1.272,0,2.353-0.795,2.789-1.912c3.118-0.379,6.812,0.531,10.163,2.563	c6.403,3.881,9.67,10.569,7.282,14.911c-0.827,1.504-2.286,2.572-4.218,3.09c-2.286,0.611-5.007,0.394-7.727-0.528	c-0.839,0.772-1.749,1.498-2.725,2.168c2.552,1.117,5.196,1.704,7.669,1.704c1.24,0,2.438-0.147,3.559-0.447	c2.741-0.733,4.841-2.304,6.071-4.542C47.016,33.276,43.267,24.787,35.507,20.084z"}),v.jsx("path",{fill:"#7e57c2",d:"M35,28.992C35,27.34,33.657,26,32,26s-3,1.34-3,2.992c0,0.669,0.228,1.281,0.6,1.779	c-1.279,2.802-3.744,5.567-7.062,7.578c-3.865,2.344-8.185,3.202-11.555,2.302c-1.932-0.518-3.391-1.586-4.218-3.09	c-1.702-3.094-0.521-7.376,2.61-10.988c-0.323-1.144-0.562-2.34-0.706-3.575c-5.07,4.797-7.109,11.323-4.532,16.009	c1.23,2.238,3.33,3.809,6.071,4.542c1.121,0.3,2.318,0.447,3.559,0.447c3.346,0,7.007-1.068,10.326-3.08	c3.836-2.325,6.683-5.577,8.209-8.962C33.815,31.801,35,30.541,35,28.992z"})]})}),Pp=[Sp,Cp,_p,jp,Ep,kp];function Np({className:e=""}){return v.jsx("div",{className:e,children:Pp.map((t,n)=>v.jsx(E.Fragment,{children:t},n))})}function Lp({className:e,w:t,h:n}){return v.jsxs(v.Fragment,{children:[v.jsxs("p",{className:"col-span-2 row-span-2 h-fit w-full text-wrap px-4 py-1.5 text-sm sm:text-xl lg:text-2xl",children:["Ali Aljaffer",v.jsx("br",{}),"Software Engineer, Web Developer"]}),v.jsx("p",{className:"col-span-2 row-span-2 w-full text-wrap px-4 py-1.5 indent-4 text-sm sm:text-xl lg:text-2xl",children:"Fresh Computer Science graduate with a minor in Mathematics from Wright State University, Dayton, Ohio. Interested in Web Development, Mathematics, and Data Analysis. I started Web Development in 2023 during my 3rd year of university, and have been working on improving my skillset in this topic ever since. I also started learning mobile development early 2024, so I'm excited to start working on real-world apps on both iOS and Android platforms!"}),v.jsx("p",{className:"col-span-2 row-span-2 w-full text-wrap px-4 py-1.5 indent-4 text-sm sm:text-xl lg:text-2xl",children:"In my free time, I build mechanical keyboards, do photography and astrophotography, both digital and analog, and I love to cook! "}),v.jsx(Np,{imgs_size:"w-[10%]",className:"flex w-[100%] flex-row flex-wrap items-start justify-around gap-1 px-4 lg:justify-between lg:gap-1",length:8})]})}function Rp({children:e,className:t}){return v.jsx("div",{className:`rounded-lg bg-c2 ${t}`,children:e})}/**
- * @remix-run/router v1.19.1
- *
- * Copyright (c) Remix Software Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.md file in the root directory of this source tree.
- *
- * @license MIT
- */function rr(){return rr=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},rr.apply(this,arguments)}var lt;(function(e){e.Pop="POP",e.Push="PUSH",e.Replace="REPLACE"})(lt||(lt={}));const cs="popstate";function zp(e){e===void 0&&(e={});function t(r,l){let{pathname:i,search:o,hash:u}=r.location;return eo("",{pathname:i,search:o,hash:u},l.state&&l.state.usr||null,l.state&&l.state.key||"default")}function n(r,l){return typeof l=="string"?l:pl(l)}return Mp(t,n,null,e)}function Y(e,t){if(e===!1||e===null||typeof e>"u")throw new Error(t)}function Tc(e,t){if(!e){typeof console<"u"&&console.warn(t);try{throw new Error(t)}catch{}}}function Tp(){return Math.random().toString(36).substr(2,8)}function fs(e,t){return{usr:e.state,key:e.key,idx:t}}function eo(e,t,n,r){return n===void 0&&(n=null),rr({pathname:typeof e=="string"?e:e.pathname,search:"",hash:""},typeof t=="string"?mn(t):t,{state:n,key:t&&t.key||r||Tp()})}function pl(e){let{pathname:t="/",search:n="",hash:r=""}=e;return n&&n!=="?"&&(t+=n.charAt(0)==="?"?n:"?"+n),r&&r!=="#"&&(t+=r.charAt(0)==="#"?r:"#"+r),t}function mn(e){let t={};if(e){let n=e.indexOf("#");n>=0&&(t.hash=e.substr(n),e=e.substr(0,n));let r=e.indexOf("?");r>=0&&(t.search=e.substr(r),e=e.substr(0,r)),e&&(t.pathname=e)}return t}function Mp(e,t,n,r){r===void 0&&(r={});let{window:l=document.defaultView,v5Compat:i=!1}=r,o=l.history,u=lt.Pop,s=null,a=h();a==null&&(a=0,o.replaceState(rr({},o.state,{idx:a}),""));function h(){return(o.state||{idx:null}).idx}function p(){u=lt.Pop;let C=h(),f=C==null?null:C-a;a=C,s&&s({action:u,location:y.location,delta:f})}function m(C,f){u=lt.Push;let c=eo(y.location,C,f);a=h()+1;let d=fs(c,a),g=y.createHref(c);try{o.pushState(d,"",g)}catch(k){if(k instanceof DOMException&&k.name==="DataCloneError")throw k;l.location.assign(g)}i&&s&&s({action:u,location:y.location,delta:1})}function w(C,f){u=lt.Replace;let c=eo(y.location,C,f);a=h();let d=fs(c,a),g=y.createHref(c);o.replaceState(d,"",g),i&&s&&s({action:u,location:y.location,delta:0})}function x(C){let f=l.location.origin!=="null"?l.location.origin:l.location.href,c=typeof C=="string"?C:pl(C);return c=c.replace(/ $/,"%20"),Y(f,"No window.location.(origin|href) available to create URL for href: "+c),new URL(c,f)}let y={get action(){return u},get location(){return e(l,o)},listen(C){if(s)throw new Error("A history only accepts one active listener");return l.addEventListener(cs,p),s=C,()=>{l.removeEventListener(cs,p),s=null}},createHref(C){return t(l,C)},createURL:x,encodeLocation(C){let f=x(C);return{pathname:f.pathname,search:f.search,hash:f.hash}},push:m,replace:w,go(C){return o.go(C)}};return y}var ds;(function(e){e.data="data",e.deferred="deferred",e.redirect="redirect",e.error="error"})(ds||(ds={}));function Op(e,t,n){return n===void 0&&(n="/"),Fp(e,t,n,!1)}function Fp(e,t,n,r){let l=typeof t=="string"?mn(t):t,i=qo(l.pathname||"/",n);if(i==null)return null;let o=Mc(e);Dp(o);let u=null;for(let s=0;u==null&&s<o.length;++s){let a=Gp(i);u=Qp(o[s],a,r)}return u}function Mc(e,t,n,r){t===void 0&&(t=[]),n===void 0&&(n=[]),r===void 0&&(r="");let l=(i,o,u)=>{let s={relativePath:u===void 0?i.path||"":u,caseSensitive:i.caseSensitive===!0,childrenIndex:o,route:i};s.relativePath.startsWith("/")&&(Y(s.relativePath.startsWith(r),'Absolute route path "'+s.relativePath+'" nested under path '+('"'+r+'" is not valid. An absolute child route path ')+"must start with the combined path of all its parent routes."),s.relativePath=s.relativePath.slice(r.length));let a=pt([r,s.relativePath]),h=n.concat(s);i.children&&i.children.length>0&&(Y(i.index!==!0,"Index routes must not have child routes. Please remove "+('all child routes from route path "'+a+'".')),Mc(i.children,t,h,a)),!(i.path==null&&!i.index)&&t.push({path:a,score:Wp(a,i.index),routesMeta:h})};return e.forEach((i,o)=>{var u;if(i.path===""||!((u=i.path)!=null&&u.includes("?")))l(i,o);else for(let s of Oc(i.path))l(i,o,s)}),t}function Oc(e){let t=e.split("/");if(t.length===0)return[];let[n,...r]=t,l=n.endsWith("?"),i=n.replace(/\?$/,"");if(r.length===0)return l?[i,""]:[i];let o=Oc(r.join("/")),u=[];return u.push(...o.map(s=>s===""?i:[i,s].join("/"))),l&&u.push(...o),u.map(s=>e.startsWith("/")&&s===""?"/":s)}function Dp(e){e.sort((t,n)=>t.score!==n.score?n.score-t.score:Hp(t.routesMeta.map(r=>r.childrenIndex),n.routesMeta.map(r=>r.childrenIndex)))}const Ip=/^:[\w-]+$/,Up=3,$p=2,Bp=1,Ap=10,Vp=-2,ps=e=>e==="*";function Wp(e,t){let n=e.split("/"),r=n.length;return n.some(ps)&&(r+=Vp),t&&(r+=$p),n.filter(l=>!ps(l)).reduce((l,i)=>l+(Ip.test(i)?Up:i===""?Bp:Ap),r)}function Hp(e,t){return e.length===t.length&&e.slice(0,-1).every((r,l)=>r===t[l])?e[e.length-1]-t[t.length-1]:0}function Qp(e,t,n){let{routesMeta:r}=e,l={},i="/",o=[];for(let u=0;u<r.length;++u){let s=r[u],a=u===r.length-1,h=i==="/"?t:t.slice(i.length)||"/",p=hs({path:s.relativePath,caseSensitive:s.caseSensitive,end:a},h),m=s.route;if(!p&&a&&n&&!r[r.length-1].route.index&&(p=hs({path:s.relativePath,caseSensitive:s.caseSensitive,end:!1},h)),!p)return null;Object.assign(l,p.params),o.push({params:l,pathname:pt([i,p.pathname]),pathnameBase:Zp(pt([i,p.pathnameBase])),route:m}),p.pathnameBase!=="/"&&(i=pt([i,p.pathnameBase]))}return o}function hs(e,t){typeof e=="string"&&(e={path:e,caseSensitive:!1,end:!0});let[n,r]=Kp(e.path,e.caseSensitive,e.end),l=t.match(n);if(!l)return null;let i=l[0],o=i.replace(/(.)\/+$/,"$1"),u=l.slice(1);return{params:r.reduce((a,h,p)=>{let{paramName:m,isOptional:w}=h;if(m==="*"){let y=u[p]||"";o=i.slice(0,i.length-y.length).replace(/(.)\/+$/,"$1")}const x=u[p];return w&&!x?a[m]=void 0:a[m]=(x||"").replace(/%2F/g,"/"),a},{}),pathname:i,pathnameBase:o,pattern:e}}function Kp(e,t,n){t===void 0&&(t=!1),n===void 0&&(n=!0),Tc(e==="*"||!e.endsWith("*")||e.endsWith("/*"),'Route path "'+e+'" will be treated as if it were '+('"'+e.replace(/\*$/,"/*")+'" because the `*` character must ')+"always follow a `/` in the pattern. To get rid of this warning, "+('please change the route path to "'+e.replace(/\*$/,"/*")+'".'));let r=[],l="^"+e.replace(/\/*\*?$/,"").replace(/^\/*/,"/").replace(/[\\.*+^${}|()[\]]/g,"\\$&").replace(/\/:([\w-]+)(\?)?/g,(o,u,s)=>(r.push({paramName:u,isOptional:s!=null}),s?"/?([^\\/]+)?":"/([^\\/]+)"));return e.endsWith("*")?(r.push({paramName:"*"}),l+=e==="*"||e==="/*"?"(.*)$":"(?:\\/(.+)|\\/*)$"):n?l+="\\/*$":e!==""&&e!=="/"&&(l+="(?:(?=\\/|$))"),[new RegExp(l,t?void 0:"i"),r]}function Gp(e){try{return e.split("/").map(t=>decodeURIComponent(t).replace(/\//g,"%2F")).join("/")}catch(t){return Tc(!1,'The URL path "'+e+'" could not be decoded because it is is a malformed URL segment. This is probably due to a bad percent '+("encoding ("+t+").")),e}}function qo(e,t){if(t==="/")return e;if(!e.toLowerCase().startsWith(t.toLowerCase()))return null;let n=t.endsWith("/")?t.length-1:t.length,r=e.charAt(n);return r&&r!=="/"?null:e.slice(n)||"/"}function Yp(e,t){t===void 0&&(t="/");let{pathname:n,search:r="",hash:l=""}=typeof e=="string"?mn(e):e;return{pathname:n?n.startsWith("/")?n:Xp(n,t):t,search:qp(r),hash:bp(l)}}function Xp(e,t){let n=t.replace(/\/+$/,"").split("/");return e.split("/").forEach(l=>{l===".."?n.length>1&&n.pop():l!=="."&&n.push(l)}),n.length>1?n.join("/"):"/"}function ui(e,t,n,r){return"Cannot include a '"+e+"' character in a manually specified "+("`to."+t+"` field ["+JSON.stringify(r)+"].  Please separate it out to the ")+("`to."+n+"` field. Alternatively you may provide the full path as ")+'a string in <Link to="..."> and the router will parse it for you.'}function Jp(e){return e.filter((t,n)=>n===0||t.route.path&&t.route.path.length>0)}function Fc(e,t){let n=Jp(e);return t?n.map((r,l)=>l===n.length-1?r.pathname:r.pathnameBase):n.map(r=>r.pathnameBase)}function Dc(e,t,n,r){r===void 0&&(r=!1);let l;typeof e=="string"?l=mn(e):(l=rr({},e),Y(!l.pathname||!l.pathname.includes("?"),ui("?","pathname","search",l)),Y(!l.pathname||!l.pathname.includes("#"),ui("#","pathname","hash",l)),Y(!l.search||!l.search.includes("#"),ui("#","search","hash",l)));let i=e===""||l.pathname==="",o=i?"/":l.pathname,u;if(o==null)u=n;else{let p=t.length-1;if(!r&&o.startsWith("..")){let m=o.split("/");for(;m[0]==="..";)m.shift(),p-=1;l.pathname=m.join("/")}u=p>=0?t[p]:"/"}let s=Yp(l,u),a=o&&o!=="/"&&o.endsWith("/"),h=(i||o===".")&&n.endsWith("/");return!s.pathname.endsWith("/")&&(a||h)&&(s.pathname+="/"),s}const pt=e=>e.join("/").replace(/\/\/+/g,"/"),Zp=e=>e.replace(/\/+$/,"").replace(/^\/*/,"/"),qp=e=>!e||e==="?"?"":e.startsWith("?")?e:"?"+e,bp=e=>!e||e==="#"?"":e.startsWith("#")?e:"#"+e;function e0(e){return e!=null&&typeof e.status=="number"&&typeof e.statusText=="string"&&typeof e.internal=="boolean"&&"data"in e}const Ic=["post","put","patch","delete"];new Set(Ic);const t0=["get",...Ic];new Set(t0);/**
- * React Router v6.26.1
- *
- * Copyright (c) Remix Software Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.md file in the root directory of this source tree.
- *
- * @license MIT
- */function lr(){return lr=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},lr.apply(this,arguments)}const bo=E.createContext(null),n0=E.createContext(null),Dt=E.createContext(null),zl=E.createContext(null),It=E.createContext({outlet:null,matches:[],isDataRoute:!1}),Uc=E.createContext(null);function r0(e,t){let{relative:n}=t===void 0?{}:t;cr()||Y(!1);let{basename:r,navigator:l}=E.useContext(Dt),{hash:i,pathname:o,search:u}=Bc(e,{relative:n}),s=o;return r!=="/"&&(s=o==="/"?r:pt([r,o])),l.createHref({pathname:s,search:u,hash:i})}function cr(){return E.useContext(zl)!=null}function fr(){return cr()||Y(!1),E.useContext(zl).location}function $c(e){E.useContext(Dt).static||E.useLayoutEffect(e)}function l0(){let{isDataRoute:e}=E.useContext(It);return e?g0():i0()}function i0(){cr()||Y(!1);let e=E.useContext(bo),{basename:t,future:n,navigator:r}=E.useContext(Dt),{matches:l}=E.useContext(It),{pathname:i}=fr(),o=JSON.stringify(Fc(l,n.v7_relativeSplatPath)),u=E.useRef(!1);return $c(()=>{u.current=!0}),E.useCallback(function(a,h){if(h===void 0&&(h={}),!u.current)return;if(typeof a=="number"){r.go(a);return}let p=Dc(a,JSON.parse(o),i,h.relative==="path");e==null&&t!=="/"&&(p.pathname=p.pathname==="/"?t:pt([t,p.pathname])),(h.replace?r.replace:r.push)(p,h.state,h)},[t,r,o,i,e])}function Bc(e,t){let{relative:n}=t===void 0?{}:t,{future:r}=E.useContext(Dt),{matches:l}=E.useContext(It),{pathname:i}=fr(),o=JSON.stringify(Fc(l,r.v7_relativeSplatPath));return E.useMemo(()=>Dc(e,JSON.parse(o),i,n==="path"),[e,o,i,n])}function o0(e,t){return u0(e,t)}function u0(e,t,n,r){cr()||Y(!1);let{navigator:l}=E.useContext(Dt),{matches:i}=E.useContext(It),o=i[i.length-1],u=o?o.params:{};o&&o.pathname;let s=o?o.pathnameBase:"/";o&&o.route;let a=fr(),h;if(t){var p;let C=typeof t=="string"?mn(t):t;s==="/"||(p=C.pathname)!=null&&p.startsWith(s)||Y(!1),h=C}else h=a;let m=h.pathname||"/",w=m;if(s!=="/"){let C=s.replace(/^\//,"").split("/");w="/"+m.replace(/^\//,"").split("/").slice(C.length).join("/")}let x=Op(e,{pathname:w}),y=d0(x&&x.map(C=>Object.assign({},C,{params:Object.assign({},u,C.params),pathname:pt([s,l.encodeLocation?l.encodeLocation(C.pathname).pathname:C.pathname]),pathnameBase:C.pathnameBase==="/"?s:pt([s,l.encodeLocation?l.encodeLocation(C.pathnameBase).pathname:C.pathnameBase])})),i,n,r);return t&&y?E.createElement(zl.Provider,{value:{location:lr({pathname:"/",search:"",hash:"",state:null,key:"default"},h),navigationType:lt.Pop}},y):y}function s0(){let e=v0(),t=e0(e)?e.status+" "+e.statusText:e instanceof Error?e.message:JSON.stringify(e),n=e instanceof Error?e.stack:null,l={padding:"0.5rem",backgroundColor:"rgba(200,200,200, 0.5)"};return E.createElement(E.Fragment,null,E.createElement("h2",null,"Unexpected Application Error!"),E.createElement("h3",{style:{fontStyle:"italic"}},t),n?E.createElement("pre",{style:l},n):null,null)}const a0=E.createElement(s0,null);class c0 extends E.Component{constructor(t){super(t),this.state={location:t.location,revalidation:t.revalidation,error:t.error}}static getDerivedStateFromError(t){return{error:t}}static getDerivedStateFromProps(t,n){return n.location!==t.location||n.revalidation!=="idle"&&t.revalidation==="idle"?{error:t.error,location:t.location,revalidation:t.revalidation}:{error:t.error!==void 0?t.error:n.error,location:n.location,revalidation:t.revalidation||n.revalidation}}componentDidCatch(t,n){console.error("React Router caught the following error during render",t,n)}render(){return this.state.error!==void 0?E.createElement(It.Provider,{value:this.props.routeContext},E.createElement(Uc.Provider,{value:this.state.error,children:this.props.component})):this.props.children}}function f0(e){let{routeContext:t,match:n,children:r}=e,l=E.useContext(bo);return l&&l.static&&l.staticContext&&(n.route.errorElement||n.route.ErrorBoundary)&&(l.staticContext._deepestRenderedBoundaryId=n.route.id),E.createElement(It.Provider,{value:t},r)}function d0(e,t,n,r){var l;if(t===void 0&&(t=[]),n===void 0&&(n=null),r===void 0&&(r=null),e==null){var i;if(!n)return null;if(n.errors)e=n.matches;else if((i=r)!=null&&i.v7_partialHydration&&t.length===0&&!n.initialized&&n.matches.length>0)e=n.matches;else return null}let o=e,u=(l=n)==null?void 0:l.errors;if(u!=null){let h=o.findIndex(p=>p.route.id&&(u==null?void 0:u[p.route.id])!==void 0);h>=0||Y(!1),o=o.slice(0,Math.min(o.length,h+1))}let s=!1,a=-1;if(n&&r&&r.v7_partialHydration)for(let h=0;h<o.length;h++){let p=o[h];if((p.route.HydrateFallback||p.route.hydrateFallbackElement)&&(a=h),p.route.id){let{loaderData:m,errors:w}=n,x=p.route.loader&&m[p.route.id]===void 0&&(!w||w[p.route.id]===void 0);if(p.route.lazy||x){s=!0,a>=0?o=o.slice(0,a+1):o=[o[0]];break}}}return o.reduceRight((h,p,m)=>{let w,x=!1,y=null,C=null;n&&(w=u&&p.route.id?u[p.route.id]:void 0,y=p.route.errorElement||a0,s&&(a<0&&m===0?(x=!0,C=null):a===m&&(x=!0,C=p.route.hydrateFallbackElement||null)));let f=t.concat(o.slice(0,m+1)),c=()=>{let d;return w?d=y:x?d=C:p.route.Component?d=E.createElement(p.route.Component,null):p.route.element?d=p.route.element:d=h,E.createElement(f0,{match:p,routeContext:{outlet:h,matches:f,isDataRoute:n!=null},children:d})};return n&&(p.route.ErrorBoundary||p.route.errorElement||m===0)?E.createElement(c0,{location:n.location,revalidation:n.revalidation,component:y,error:w,children:c(),routeContext:{outlet:null,matches:f,isDataRoute:!0}}):c()},null)}var Ac=function(e){return e.UseBlocker="useBlocker",e.UseRevalidator="useRevalidator",e.UseNavigateStable="useNavigate",e}(Ac||{}),hl=function(e){return e.UseBlocker="useBlocker",e.UseLoaderData="useLoaderData",e.UseActionData="useActionData",e.UseRouteError="useRouteError",e.UseNavigation="useNavigation",e.UseRouteLoaderData="useRouteLoaderData",e.UseMatches="useMatches",e.UseRevalidator="useRevalidator",e.UseNavigateStable="useNavigate",e.UseRouteId="useRouteId",e}(hl||{});function p0(e){let t=E.useContext(bo);return t||Y(!1),t}function h0(e){let t=E.useContext(n0);return t||Y(!1),t}function m0(e){let t=E.useContext(It);return t||Y(!1),t}function Vc(e){let t=m0(),n=t.matches[t.matches.length-1];return n.route.id||Y(!1),n.route.id}function v0(){var e;let t=E.useContext(Uc),n=h0(hl.UseRouteError),r=Vc(hl.UseRouteError);return t!==void 0?t:(e=n.errors)==null?void 0:e[r]}function g0(){let{router:e}=p0(Ac.UseNavigateStable),t=Vc(hl.UseNavigateStable),n=E.useRef(!1);return $c(()=>{n.current=!0}),E.useCallback(function(l,i){i===void 0&&(i={}),n.current&&(typeof l=="number"?e.navigate(l):e.navigate(l,lr({fromRouteId:t},i)))},[e,t])}function Rn(e){Y(!1)}function y0(e){let{basename:t="/",children:n=null,location:r,navigationType:l=lt.Pop,navigator:i,static:o=!1,future:u}=e;cr()&&Y(!1);let s=t.replace(/^\/*/,"/"),a=E.useMemo(()=>({basename:s,navigator:i,static:o,future:lr({v7_relativeSplatPath:!1},u)}),[s,u,i,o]);typeof r=="string"&&(r=mn(r));let{pathname:h="/",search:p="",hash:m="",state:w=null,key:x="default"}=r,y=E.useMemo(()=>{let C=qo(h,s);return C==null?null:{location:{pathname:C,search:p,hash:m,state:w,key:x},navigationType:l}},[s,h,p,m,w,x,l]);return y==null?null:E.createElement(Dt.Provider,{value:a},E.createElement(zl.Provider,{children:n,value:y}))}function w0(e){let{children:t,location:n}=e;return o0(to(t),n)}new Promise(()=>{});function to(e,t){t===void 0&&(t=[]);let n=[];return E.Children.forEach(e,(r,l)=>{if(!E.isValidElement(r))return;let i=[...t,l];if(r.type===E.Fragment){n.push.apply(n,to(r.props.children,i));return}r.type!==Rn&&Y(!1),!r.props.index||!r.props.children||Y(!1);let o={id:r.props.id||i.join("-"),caseSensitive:r.props.caseSensitive,element:r.props.element,Component:r.props.Component,index:r.props.index,path:r.props.path,loader:r.props.loader,action:r.props.action,errorElement:r.props.errorElement,ErrorBoundary:r.props.ErrorBoundary,hasErrorBoundary:r.props.ErrorBoundary!=null||r.props.errorElement!=null,shouldRevalidate:r.props.shouldRevalidate,handle:r.props.handle,lazy:r.props.lazy};r.props.children&&(o.children=to(r.props.children,i)),n.push(o)}),n}/**
- * React Router DOM v6.26.1
- *
- * Copyright (c) Remix Software Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.md file in the root directory of this source tree.
- *
- * @license MIT
- */function no(){return no=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},no.apply(this,arguments)}function x0(e,t){if(e==null)return{};var n={},r=Object.keys(e),l,i;for(i=0;i<r.length;i++)l=r[i],!(t.indexOf(l)>=0)&&(n[l]=e[l]);return n}function S0(e){return!!(e.metaKey||e.altKey||e.ctrlKey||e.shiftKey)}function k0(e,t){return e.button===0&&(!t||t==="_self")&&!S0(e)}const C0=["onClick","relative","reloadDocument","replace","state","target","to","preventScrollReset","unstable_viewTransition"],E0="6";try{window.__reactRouterVersion=E0}catch{}const j0="startTransition",ms=af[j0];function _0(e){let{basename:t,children:n,future:r,window:l}=e,i=E.useRef();i.current==null&&(i.current=zp({window:l,v5Compat:!0}));let o=i.current,[u,s]=E.useState({action:o.action,location:o.location}),{v7_startTransition:a}=r||{},h=E.useCallback(p=>{a&&ms?ms(()=>s(p)):s(p)},[s,a]);return E.useLayoutEffect(()=>o.listen(h),[o,h]),E.createElement(y0,{basename:t,children:n,location:u.location,navigationType:u.action,navigator:o,future:r})}const P0=typeof window<"u"&&typeof window.document<"u"&&typeof window.document.createElement<"u",N0=/^(?:[a-z][a-z0-9+.-]*:|\/\/)/i,L0=E.forwardRef(function(t,n){let{onClick:r,relative:l,reloadDocument:i,replace:o,state:u,target:s,to:a,preventScrollReset:h,unstable_viewTransition:p}=t,m=x0(t,C0),{basename:w}=E.useContext(Dt),x,y=!1;if(typeof a=="string"&&N0.test(a)&&(x=a,P0))try{let d=new URL(window.location.href),g=a.startsWith("//")?new URL(d.protocol+a):new URL(a),k=qo(g.pathname,w);g.origin===d.origin&&k!=null?a=k+g.search+g.hash:y=!0}catch{}let C=r0(a,{relative:l}),f=R0(a,{replace:o,state:u,target:s,preventScrollReset:h,relative:l,unstable_viewTransition:p});function c(d){r&&r(d),d.defaultPrevented||f(d)}return E.createElement("a",no({},m,{href:x||C,onClick:y||i?r:c,ref:n,target:s}))});var vs;(function(e){e.UseScrollRestoration="useScrollRestoration",e.UseSubmit="useSubmit",e.UseSubmitFetcher="useSubmitFetcher",e.UseFetcher="useFetcher",e.useViewTransitionState="useViewTransitionState"})(vs||(vs={}));var gs;(function(e){e.UseFetcher="useFetcher",e.UseFetchers="useFetchers",e.UseScrollRestoration="useScrollRestoration"})(gs||(gs={}));function R0(e,t){let{target:n,replace:r,state:l,preventScrollReset:i,relative:o,unstable_viewTransition:u}=t===void 0?{}:t,s=l0(),a=fr(),h=Bc(e,{relative:o});return E.useCallback(p=>{if(k0(p,n)){p.preventDefault();let m=r!==void 0?r:pl(a)===pl(h);s(e,{replace:m,state:l,preventScrollReset:i,relative:o,unstable_viewTransition:u})}},[a,s,h,r,l,n,e,i,o,u])}function St({onClick:e,children:t,activeColor:n="bg-stone-900",idleColor:r="bg-stone-500",variation:l="",className:i="",to:o=void 0}){const u="opacity-80 rounded-sm p-3 uppercase sm:p-6",s={active:`${u} ${n} text-lg text-c1`,idle:`${u} ${r} text-lg text-c5`,small:`${u} py-0.5 lg:py-1 lg:h-full text-md text-stone-600 font-semibold hover:font-bold tracking-widest  hover:text-stone-800 hover:border-b-2 hover:shadow-cool hover:border-stone-300 hover:rounded-lg border-b-2 border-r-2 border-transparent duration-300 transition-all sm:text-xl md:py-1 md:px-4 md:text-md md:shadow-lg lg:text-2xl lg:rounded-xl lg:px-8 md:shadow-stone-300 md:rounded-lg`};return o!==void 0?v.jsx(L0,{className:i+" "+l==""?s.active:s[l],to:o,children:t}):v.jsx("button",{onClick:e,className:i+" "+l==""?s.active:s[l],children:t})}function z0(){return fr().pathname}function T0({width:e}){const t=z0(),r=!["/","/projects","/contact"].includes(t);return v.jsx("div",{className:`${e} flex items-center ${r?"justify-center":"justify-between"} gap-4 px-1`,children:t==="/"?v.jsxs(v.Fragment,{children:[v.jsx(St,{variation:"small",to:"/contact",children:"Contact"}),v.jsx(St,{variation:"small",to:"/projects",children:"Projects"})]}):t==="/contact"?v.jsxs(v.Fragment,{children:[v.jsx(St,{variation:"small",to:"/",children:"Home"}),v.jsx(St,{variation:"small",to:"/projects",children:"Projects"})]}):t==="/projects"?v.jsxs(v.Fragment,{children:[v.jsx(St,{variation:"small",to:"/contact",children:"Contact"}),v.jsx(St,{variation:"small",to:"/",children:"Home"})]}):v.jsx(St,{variation:"small",to:"/",children:"Home"})})}function M0(){return v.jsxs("div",{className:"flex flex-col items-center justify-center gap-4 p-4 text-4xl",children:[v.jsx("h1",{className:"text-center",children:"404"}),v.jsx("p",{className:"text-center",children:"Page not found!"})]})}const ys=[{projectId:1,projectName:"🎵 spotiVersus",projectType:"Full-Stack development,Entertainment, Music, Data Model Design, Database Schema Design, REST APIs, Data Retrieval",tech:["react","vite","tailwind","react router","typescript","supabase","vercel","postgreSQL"],githubRepoURL:"https://github.com/aliAljaffer/spotiVersus",hostedSiteURL:"https://spoti-versus.vercel.app/",isFinished:!0,reasonNotFinished:"Fully working state. I have more features planned, so can't really say it's finished.",description:"spotiVersus is a mini-game of guessing which of the two artists displayed has a higher number of monthly listeners on Spotify. At the start, a number of artists is pulled from a database."},{projectId:3,projectName:"🍕 fast react pizza",projectType:"Full-Stack development, Responsive Design, Database Management, UI/UX",tech:["react","vite","redux toolkit","tailwind","react router","vercel"],githubRepoURL:"https://github.com/aliAljaffer/fr-pizza",hostedSiteURL:"https://fast-r-pizza.vercel.app",isFinished:!0,reasonNotFinished:"",description:"A mock-website for a small Pizza shop. The business rules did not require user authentication, so login/registration was not implemented. Each order is registered and assigned a persistant unique ID."},{projectId:2,projectName:"🗂️ MOSS for Pilot",projectType:"Process Automation, File Organization, Integration with Other Tools, Open-Closed Design, File Parsing",tech:["Python","ZipFile","OS","shutil"],githubRepoURL:"https://github.com/aliAljaffer/MOSS-for-Pilot-LMS",hostedSiteURL:"https://github.com/aliAljaffer/CSE-Discord-Bot",isFinished:!0,reasonNotFinished:"",description:"Mass-checking of student submitted code downloaded through Pilot Learning Management System (LMS) for purposes of cheating deterrence. A python script that expects a zip file downloaded from an LMS, which then unzips it and creates a folder named after the student's last name with their most-recently submitted files. This script was later adapted by the CS Dev team at Wright State University and refactored to be used as a /command in the CS Discord server."},{projectId:4,projectName:"🧮 WithinLimits",projectType:"Entertainment, Games, Bounded-Limits Ruleset, Input Handling, Game Logic Design",tech:["Java","JavaSwing GUI","Model-View-Controller Design Pattern"],githubRepoURL:"https://github.com/aliAljaffer/projects/tree/main/WithinLimits",hostedSiteURL:"",isFinished:!0,reasonNotFinished:"Will refactor into Flutter code to deploy for iOS and Android.",description:"Math game where the player must keep the running total inbetween two bounds. A random number is shown, and the player has to choose between three operations to perform on the random number and the running total number, ensuring that the operation would not result in a total outside of the set bounds."},{projectId:5,projectName:"⚽️ Soccer Predictor",projectType:"Educational, Machine Learning, Data Analysis, Feature Extraction, Data Preprocessing, Random Forest, Data Resampling, Feature Derivation",tech:["Python, Matplotlib, keras, numpy, pandas, Jupyter Notebooks"],githubRepoURL:"https://github.com/aliAljaffer/soccer-match-predictor",hostedSiteURL:"",isFinished:!0,reasonNotFinished:"",description:"Soccer match predictions using machine learning techniques, including data preprocessing and cleanup, feature extraction and derivation, and model optimization."},{projectId:6,projectName:"🧩 Data Structures",projectType:"Educational, Memory Management and Memory Leak Analysis, Hashing (jsHash), Object-Oriented Programming, Recursion, Complexity Analysis, Unit Testing",tech:["C++, Catch2"],githubRepoURL:"https://github.com/aliAljaffer/DSA",hostedSiteURL:"https://alialjaffer.github.io/DSA/",isFinished:!0,reasonNotFinished:"",description:"Data structures implementations in C++ with focus on performance and low memory usage. Implementing: Doubly-Linked List (Sequence), Balanced Binary Tree (AVL Tree), Tries (Auto-Completion Trie), HashDatabase."},{projectId:7,projectName:"🏡 Qiddiyah Cabins",projectType:"Full-Stack Development, CRUD Operations, Data Visualization, Asset Management System, Authentication",tech:["react","react query","supabase","postgreSQL","styled-components","auth0"],githubRepoURL:"",hostedSiteURL:"",isFinished:!1,reasonNotFinished:"Work in progress.",description:"Qiddiyah Cabins - Find Your Desert Sanctuary. An internal dashboard designed for the management of a cabin-renting firm. The dashboard allows for authenticated staff members to perform CRUD operations on a postgreSQL database for the firm's cabins. More features planned such as statistics for income, guest retention, average number of stays, etc."}];function O0({project:e,tech:t,idx:n,numProjects:r}){const[l,i]=E.useState(!1),{projectName:o,projectType:u,githubRepoURL:s,hostedSiteURL:a,isFinished:h,reasonNotFinished:p,description:m}=e,w=()=>{i(y=>!y)},x=v.jsx("p",{className:"row-span-3 inline h-full overflow-y-auto text-wrap text-xs transition-all duration-300",children:m});return v.jsxs("div",{className:"relative grid h-[65%] w-full shrink-0 grid-cols-1 grid-rows-[fit_fit_fit_fit] gap-4 overflow-visible rounded-2xl px-4 py-3 transition-all duration-300 hover:bg-stone-600 sm:px-3 md:w-[40%] md:gap-8 lg:w-[33%] xl:h-[85%] xl:w-[25%]",children:[v.jsxs("span",{className:"absolute -top-0.5 right-0 text-xs",children:["(",n+1," / ",r,") ",n+1<r?v.jsx(v.Fragment,{children:"→"}):""]}),v.jsx("p",{className:"w-full text-xl font-bold capitalize tracking-wide sm:tracking-wide",children:o}),l?x:v.jsxs(v.Fragment,{children:[v.jsx("p",{className:"inline h-full text-wrap text-sm text-stone-500",children:u}),v.jsx("p",{className:"text-wrap text-sm capitalize text-stone-400",children:t.map((y,C)=>y+(C===t.length-1?"":", "))}),v.jsxs("div",{className:"flex h-fit items-start justify-between",children:[s?v.jsxs("a",{href:s,target:"_blank",rel:"noreferrer",alt:`${o} github repository`,className:"flex w-fit cursor-pointer items-center gap-2 place-self-start rounded-lg bg-stone-300 px-4 py-1.5 text-sm text-stone-600 shadow-md shadow-stone-800 transition-all duration-300 hover:-translate-y-0.5 hover:bg-stone-50 hover:font-extrabold",children:[v.jsx("span",{children:v.jsx("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"18",height:"18",viewBox:"0 0 30 30",className:"fill-stone-600",children:v.jsx("path",{d:"M15,3C8.373,3,3,8.373,3,15c0,5.623,3.872,10.328,9.092,11.63C12.036,26.468,12,26.28,12,26.047v-2.051 c-0.487,0-1.303,0-1.508,0c-0.821,0-1.551-0.353-1.905-1.009c-0.393-0.729-0.461-1.844-1.435-2.526 c-0.289-0.227-0.069-0.486,0.264-0.451c0.615,0.174,1.125,0.596,1.605,1.222c0.478,0.627,0.703,0.769,1.596,0.769 c0.433,0,1.081-0.025,1.691-0.121c0.328-0.833,0.895-1.6,1.588-1.962c-3.996-0.411-5.903-2.399-5.903-5.098 c0-1.162,0.495-2.286,1.336-3.233C9.053,10.647,8.706,8.73,9.435,8c1.798,0,2.885,1.166,3.146,1.481C13.477,9.174,14.461,9,15.495,9 c1.036,0,2.024,0.174,2.922,0.483C18.675,9.17,19.763,8,21.565,8c0.732,0.731,0.381,2.656,0.102,3.594 c0.836,0.945,1.328,2.066,1.328,3.226c0,2.697-1.904,4.684-5.894,5.097C18.199,20.49,19,22.1,19,23.313v2.734 c0,0.104-0.023,0.179-0.035,0.268C23.641,24.676,27,20.236,27,15C27,8.373,21.627,3,15,3z"})})})," ","Repo"]}):v.jsx("span",{className:"cursor-pointer text-sm line-through",children:"Repo not ready"}),a&&v.jsxs("a",{href:a,target:"_blank",rel:"noreferrer",alt:`${o} hosted website`,className:"flex w-fit cursor-pointer items-center justify-center gap-2 rounded-lg bg-stone-300 px-4 py-1.5 text-sm text-stone-600 shadow-md shadow-stone-800 transition-all duration-300 hover:-translate-y-0.5 hover:bg-stone-50 hover:font-extrabold",children:[v.jsx("span",{children:v.jsxs("svg",{width:"18",height:"18",viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg",children:[v.jsx("g",{id:"SVGRepo_bgCarrier",strokeWidth:"0"}),v.jsx("g",{id:"SVGRepo_tracerCarrier",strokeLinecap:"round",strokeLinejoin:"round"}),v.jsxs("g",{id:"SVGRepo_iconCarrier",children:[" ",v.jsx("path",{d:"M14 12C14 14.7614 11.7614 17 9 17H7C4.23858 17 2 14.7614 2 12C2 9.23858 4.23858 7 7 7H7.5M10 12C10 9.23858 12.2386 7 15 7H17C19.7614 7 22 9.23858 22 12C22 14.7614 19.7614 17 17 17H16.5",className:"stroke-stone-600",strokeWidth:"2",strokeLinecap:"round"})," "]})]})}),"Website"]})]})]}),v.jsx("button",{className:"inline w-full cursor-pointer border-none text-sm tracking-wide text-stone-100 underline transition-all duration-300 hover:translate-y-0.5",onClick:w,children:`${l?"Hide":"Show"} project summary`})]})}function F0(){const[e,t]=E.useState(!0),n=E.useRef(null),r=ys.filter(s=>!s.isFinished),l=ys.filter(s=>s.isFinished),i=e?l:r,o=()=>{t(s=>!s)},u=s=>{};return v.jsxs("div",{className:"text-md col-span-2 row-span-2 flex h-[90dvh] w-full flex-col items-start justify-start gap-4 overflow-auto text-wrap p-4 sm:text-xl lg:text-2xl xl:h-[50dvh]",children:[v.jsxs("div",{className:"flex w-full flex-row items-center justify-center",children:[v.jsx("p",{className:"text-md tracking-tight",children:e?"Past projects:":"Currently working on:"}),v.jsxs("button",{onClick:o,className:"ml-auto inline w-fit cursor-pointer place-self-start rounded-lg bg-stone-300 px-4 py-1.5 text-xs text-stone-600 shadow-md shadow-stone-800 transition-all duration-300 hover:-translate-y-0.5 hover:bg-stone-50 hover:font-extrabold sm:text-sm",children:[`${e?"In-progress":"Finished"} projects`," →"]})]}),v.jsx("div",{ref:n,onWheel:u,className:"mt-auto flex h-[95%] w-full items-center gap-12 overflow-x-scroll scroll-smooth lg:flex-row lg:scrollbar-thin lg:scrollbar-track-stone-600 lg:scrollbar-thumb-stone-400",children:i.map((s,a)=>v.jsx(E.Fragment,{children:v.jsx(O0,{idx:a,numProjects:i.length,project:s,tech:s.tech})},s.projectId))})]})}function D0(){const e="text-sm sm:text-lg";return v.jsxs("div",{className:"grid grid-cols-1 grid-rows-3 place-content-center items-center justify-center gap-5 sm:grid-cols-2 sm:grid-rows-2",children:[v.jsxs("p",{className:`flex items-center justify-start gap-2 sm:justify-self-center ${e}`,children:[v.jsx("span",{children:v.jsxs("svg",{fill:"#e7e5e4",height:"24",width:"24",version:"1.1",id:"Capa_1",xmlns:"http://www.w3.org/2000/svg",xmlnsXlink:"http://www.w3.org/1999/xlink",viewBox:"0 0 75.294 75.294",xmlSpace:"preserve",stroke:"#e7e5e4",children:[v.jsx("g",{id:"SVGRepo_bgCarrier",strokeWidth:"0"}),v.jsx("g",{id:"SVGRepo_tracerCarrier",strokeLinecap:"round",strokeLinejoin:"round"}),v.jsxs("g",{id:"SVGRepo_iconCarrier",children:[v.jsxs("g",{children:[v.jsx("path",{d:"M66.097,12.089h-56.9C4.126,12.089,0,16.215,0,21.286v32.722c0,5.071,4.126,9.197,9.197,9.197h56.9 c5.071,0,9.197-4.126,9.197-9.197V21.287C75.295,16.215,71.169,12.089,66.097,12.089z M61.603,18.089L37.647,33.523L13.691,18.089 H61.603z M66.097,57.206h-56.9C7.434,57.206,6,55.771,6,54.009V21.457l29.796,19.16c0.04,0.025,0.083,0.042,0.124,0.065 c0.043,0.024,0.087,0.047,0.131,0.069c0.231,0.119,0.469,0.215,0.712,0.278c0.025,0.007,0.05,0.01,0.075,0.016 c0.267,0.063,0.537,0.102,0.807,0.102c0.001,0,0.002,0,0.002,0c0.002,0,0.003,0,0.004,0c0.27,0,0.54-0.038,0.807-0.102 c0.025-0.006,0.05-0.009,0.075-0.016c0.243-0.063,0.48-0.159,0.712-0.278c0.044-0.022,0.088-0.045,0.131-0.069 c0.041-0.023,0.084-0.04,0.124-0.065l29.796-19.16v32.551C69.295,55.771,67.86,57.206,66.097,57.206z"})," "]})," "]})]})}),v.jsx("a",{href:"mailto:ahjfr@icloud.com",className:"underline",children:"ahjfr@icloud.com"})]}),v.jsxs("p",{className:`flex items-center justify-start gap-2 sm:justify-self-center ${e}`,children:[v.jsx("span",{children:v.jsx("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"30",height:"30",viewBox:"0 0 30 30",className:"fill-stone-200",children:v.jsx("path",{d:"M15,3C8.373,3,3,8.373,3,15c0,5.623,3.872,10.328,9.092,11.63C12.036,26.468,12,26.28,12,26.047v-2.051 c-0.487,0-1.303,0-1.508,0c-0.821,0-1.551-0.353-1.905-1.009c-0.393-0.729-0.461-1.844-1.435-2.526 c-0.289-0.227-0.069-0.486,0.264-0.451c0.615,0.174,1.125,0.596,1.605,1.222c0.478,0.627,0.703,0.769,1.596,0.769 c0.433,0,1.081-0.025,1.691-0.121c0.328-0.833,0.895-1.6,1.588-1.962c-3.996-0.411-5.903-2.399-5.903-5.098 c0-1.162,0.495-2.286,1.336-3.233C9.053,10.647,8.706,8.73,9.435,8c1.798,0,2.885,1.166,3.146,1.481C13.477,9.174,14.461,9,15.495,9 c1.036,0,2.024,0.174,2.922,0.483C18.675,9.17,19.763,8,21.565,8c0.732,0.731,0.381,2.656,0.102,3.594 c0.836,0.945,1.328,2.066,1.328,3.226c0,2.697-1.904,4.684-5.894,5.097C18.199,20.49,19,22.1,19,23.313v2.734 c0,0.104-0.023,0.179-0.035,0.268C23.641,24.676,27,20.236,27,15C27,8.373,21.627,3,15,3z"})})}),v.jsx("a",{target:"_blank",rel:"noreferrer",href:"https://github.com/aliAljaffer",className:"underline",children:"aliAljaffer"})]}),v.jsxs("p",{className:`flex items-center gap-2 justify-self-start sm:col-span-2 sm:justify-self-center ${e}`,children:[v.jsx("span",{children:v.jsx("svg",{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"24",height:"24",viewBox:"0 0 24 24",className:"fill-stone-200",children:v.jsx("path",{d:"M19,3H5C3.895,3,3,3.895,3,5v14c0,1.105,0.895,2,2,2h14c1.105,0,2-0.895,2-2V5C21,3.895,20.105,3,19,3z M7.738,17L7.738,17 c-0.697,0-1.262-0.565-1.262-1.262v-4.477C6.477,10.565,7.042,10,7.738,10h0C8.435,10,9,10.565,9,11.262v4.477 C9,16.435,8.435,17,7.738,17z M7.694,8.717c-0.771,0-1.286-0.514-1.286-1.2s0.514-1.2,1.371-1.2c0.771,0,1.286,0.514,1.286,1.2 S8.551,8.717,7.694,8.717z M16.779,17L16.779,17c-0.674,0-1.221-0.547-1.221-1.221v-2.605c0-1.058-0.651-1.174-0.895-1.174 s-1.058,0.035-1.058,1.174v2.605c0,0.674-0.547,1.221-1.221,1.221h-0.081c-0.674,0-1.221-0.547-1.221-1.221v-4.517 c0-0.697,0.565-1.262,1.262-1.262h0c0.697,0,1.262,0.565,1.262,1.262c0,0,0.282-1.262,2.198-1.262C17.023,10,18,10.977,18,13.174 v2.605C18,16.453,17.453,17,16.779,17z"})})}),v.jsx("a",{href:"https://linkedin.com/in/aliAljaffer",target:"_blank",rel:"noreferrer",className:"underline",children:"aliAljaffer"})]})]})}function I0(){const e=" w-[85dvw] sm:w-[85dvw] lg:w-[80dvw] lg:max-w-[1024px]",t=" h-[75dvh] max-h-[75dvh]  lg:h-[69dvh] lg:max-h-[90dvh] xl:max-h-[80dvh] ";return v.jsxs(_0,{children:[" ",v.jsxs("main",{className:"relative flex min-h-screen w-full flex-col items-center justify-center gap-2 bg-stone-200 text-sm text-c5 lg:gap-3",children:[v.jsx(Rp,{className:t+" "+e+" flex flex-col items-center justify-around overflow-auto bg-stone-700 py-2.5 sm:gap-1 sm:p-4 lg:justify-around lg:gap-4",children:v.jsxs(w0,{children:[v.jsx(Rn,{index:!0,path:"/",element:v.jsx(Lp,{w:e,h:t})}),v.jsx(Rn,{path:"/contact",element:v.jsx(D0,{})}),v.jsx(Rn,{path:"/projects",element:v.jsx(F0,{})}),v.jsx(Rn,{path:"*",element:v.jsx(M0,{})})]})}),v.jsx(T0,{width:e})]})]})}zc(document.getElementById("root")).render(v.jsx(E.StrictMode,{children:v.jsx(I0,{})}));
