+++ docs/index.html
@@ -17,8 +17,8 @@
       href="https://fonts.googleapis.com/css2?family=Ubuntu+Mono&family=PT+Mono&display=swap"
       rel="stylesheet"
     />
-    <script type="module" crossorigin src="/assets/index-BA9C0b2H.js"></script>
-    <link rel="stylesheet" crossorigin href="/assets/index-BDSMXxDQ.css">
+    <script type="module" crossorigin src="/assets/index-C6u0fpwo.js"></script>
+    <link rel="stylesheet" crossorigin href="/assets/index-CQDtUP8O.css">
   </head>
   <body>
     <div id="root"></div>
